Distribution.Backpack.ComponentsGraph

Description

See https://github.com/ezyang/ghc-proposals/blob/backpack/proposals/0000-backpack.rst

Documentation

type ComponentsGraph = Graph (Node ComponentName Component)

A graph of source-level components by their source-level dependencies

type ComponentsWithDeps = [(Component, [ComponentName])]

A list of components associated with the source level dependencies between them.

mkComponentsGraph :: ComponentRequestedSpec -> PackageDescription -> Either [ComponentName] ComponentsGraph

Create a Graph of Component, or report a cycle if there is a problem.

componentsGraphToList :: ComponentsGraph -> ComponentsWithDeps

Given the package description and a PackageDescription (used to determine if a package name is internal or not), sort the components in dependency order (fewest dependencies first). This is NOT necessarily the build order (although it is in the absence of Backpack.)

dispComponentsWithDeps :: ComponentsWithDeps -> Doc

Pretty-print ComponentsWithDeps.

componentCycleMsg :: PackageIdentifier -> [ComponentName] -> Doc

Error message when there is a cycle; takes the SCC of components.