Distribution.CabalSpecVersion

Documentation

data CabalSpecVersion

Different Cabal-the-spec versions.

We branch based on this at least in the parser.

Constructors

CabalSpecV1_0
this is older than CabalSpecV1_2

CabalSpecV1_2
new syntax (sections)

CabalSpecV1_4

CabalSpecV1_6

CabalSpecV1_8

CabalSpecV1_10

CabalSpecV1_12

CabalSpecV1_18

CabalSpecV1_20

CabalSpecV1_22

CabalSpecV1_24

CabalSpecV2_0

CabalSpecV2_2

CabalSpecV2_4

CabalSpecV3_0

showCabalSpecVersion :: CabalSpecVersion -> String

Show cabal spec version, but not the way in the .cabal files

Since: 3.0.0.0

cabalSpecLatest :: CabalSpecVersion

cabalSpecFromVersionDigits :: [Int] -> CabalSpecVersion

specHasCommonStanzas :: CabalSpecVersion -> HasCommonStanzas

specHasElif :: CabalSpecVersion -> HasElif

data HasElif

Constructors

HasElif

NoElif

data HasCommonStanzas

Constructors

HasCommonStanzas

NoCommonStanzas

data HasGlobstar

Constructors

HasGlobstar

NoGlobstar