Distribution.Compat.Binary

Documentation

decodeOrFailIO :: Binary a => ByteString -> IO (Either String a)

decodeFileOrFail' :: Binary a => FilePath -> IO (Either String a)

Lazily reconstruct a value previously written to a file.

class Binary t where

The Binary class provides put and get, methods to encode and decode a Haskell value to a lazy ByteString. It mirrors the Read and Show classes for textual representation of Haskell types, and is suitable for serialising Haskell values to disk, over the network.

For decoding and generating simple external binary formats (e.g. C structures), Binary may be used, but in general is not suitable for complex protocols. Instead use the Put and Get primitives directly.

Instances of Binary should satisfy the following property:

 decode . encode == id

That is, the get and put methods should be the inverse of each other. A range of instances are provided for basic Haskell types.

Methods

put :: t -> Put

Encode a value in the Put monad.

get :: Get t

Decode a value in the Get monad

decode :: Binary a => ByteString -> a

Decode a value from a lazy ByteString, reconstructing the original structure.

encode :: Binary a => a -> ByteString

Encode a value using binary serialisation to a lazy ByteString.

encodeFile :: Binary a => FilePath -> a -> IO ()

Lazily serialise a value to a file