Distribution.Compat.Prelude.Internal

Description

Warning: This modules' API is not stable. Use at your own risk, or better yet, use base-compat!

This module re-exports the non-exposed Distribution.Compat.Prelude module for reuse by cabal-install's Distribution.Client.Compat.Prelude module.

It is highly discouraged to rely on this module for Setup.hs scripts since its API is not stable.

Documentation

module Distribution.Compat.Prelude