Distribution.FieldGrammar

Description

This module provides a way to specify a grammar of .cabal -like files.

Field grammar type

class FieldGrammar g where

FieldGrammar is parametrised by

s which is a structure we are parsing. We need this to provide prettyprinter functionality

a type of the field.

Note: We'd like to have forall s. Applicative (f s) context.

Minimal complete definition

blurFieldGrammar, uniqueFieldAla, booleanFieldDef, optionalFieldAla, optionalFieldDefAla, monoidalFieldAla, prefixedFields, knownField, hiddenField, deprecatedSince, availableSince

Methods

blurFieldGrammar :: ALens' a b -> g b c -> g a c

Unfocus, zoom out, blur FieldGrammar.

uniqueFieldAla

Arguments

:: (Parsec b, Pretty b, Newtype b a)

=> FieldName
field name

-> (a -> b)
Newtype pack

-> ALens' s a
lens into the field

-> g s a

Field which should be defined, exactly once.

booleanFieldDef

Arguments

:: FieldName
field name

-> ALens' s Bool
lens into the field

-> Bool
default

-> g s Bool

Boolean field with a default value.

optionalFieldAla

Arguments

:: (Parsec b, Pretty b, Newtype b a)

=> FieldName
field name

-> (a -> b)
pack

-> ALens' s (Maybe a)
lens into the field

-> g s (Maybe a)

Optional field.

optionalFieldDefAla

Arguments

:: (Parsec b, Pretty b, Newtype b a, Eq a)

=> FieldName
field name

-> (a -> b)
Newtype pack

-> ALens' s a
Lens' s a: lens into the field

-> a
default value

-> g s a

Optional field with default value.

monoidalFieldAla

Arguments

:: (Parsec b, Pretty b, Monoid a, Newtype b a)

=> FieldName
field name

-> (a -> b)
pack

-> ALens' s a
lens into the field

-> g s a

Monoidal field.

Values are combined with mappend.

Note: optionalFieldAla is a monoidalField with Last monoid.

prefixedFields

Arguments

:: FieldName
field name prefix

-> ALens' s [(String, String)]
lens into the field

-> g s [(String, String)]

Parser matching all fields with a name starting with a prefix.

knownField :: FieldName -> g s ()

Known field, which we don't parse, neither pretty print.

hiddenField :: g s a -> g s a

Field which is parsed but not pretty printed.

deprecatedSince

Arguments

:: [Int]
version

-> String
deprecation message

-> g s a

-> g s a

Deprecated since

availableSince

Arguments

:: [Int]
spec version

-> a
default value

-> g s a

-> g s a

Annotate field with since spec-version.

uniqueField

Arguments

:: (FieldGrammar g, Parsec a, Pretty a)

=> FieldName
field name

-> ALens' s a
lens into the field

-> g s a

Field which can be defined at most once.

optionalField

Arguments

:: (FieldGrammar g, Parsec a, Pretty a)

=> FieldName
field name

-> ALens' s (Maybe a)
lens into the field

-> g s (Maybe a)

Field which can be defined at most once.

optionalFieldDef

Arguments

:: (FieldGrammar g, Functor (g s), Parsec a, Pretty a, Eq a)

=> FieldName
field name

-> ALens' s a
Lens' s a: lens into the field

-> a
default value

-> g s a

Optional field with default value.

monoidalField

Arguments

:: (FieldGrammar g, Parsec a, Pretty a, Monoid a)

=> FieldName
field name

-> ALens' s a
lens into the field

-> g s a

Field which can be define multiple times, and the results are mappended.

deprecatedField'

Arguments

:: FieldGrammar g

=> String
deprecation message

-> g s a

-> g s a

Deprecated field. If found, warning is issued.

Note: also it's not pretty printed!

Concrete grammar implementations

data ParsecFieldGrammar s a

type ParsecFieldGrammar' a = ParsecFieldGrammar a a

parseFieldGrammar :: CabalSpecVersion -> Fields Position -> ParsecFieldGrammar s a -> ParseResult a

fieldGrammarKnownFieldList :: ParsecFieldGrammar s a -> [FieldName]

data PrettyFieldGrammar s a

type PrettyFieldGrammar' a = PrettyFieldGrammar a a

prettyFieldGrammar :: PrettyFieldGrammar s a -> s -> Doc

We can use PrettyFieldGrammar to pp print the s.

Note: there is not trailing ($+$ text "").

Auxlilary

(^^^) :: a -> (a -> b) -> b infixl 5

Reverse function application which binds tighter than <$> and <*>. Useful for refining grammar specification.

<*> monoidalFieldAla "extensions"           (alaList' FSep MQuoted)       oldExtensions
    ^^^ deprecatedSince [1,12] "Please use 'default-extensions' or 'other-extensions' fields."

data Section ann

The Section constructor of Field.

Constructors

MkSection !(Name ann) [SectionArg ann] [Field ann]

type Fields ann = Map FieldName [NamelessField ann]

partitionFields :: [Field ann] -> (Fields ann, [[Section ann]])

Partition field list into field map and groups of sections.

takeFields :: [Field ann] -> (Fields ann, [Field ann])

Take all fields from the front.

runFieldParser :: Position -> ParsecParser a -> CabalSpecVersion -> [FieldLine Position] -> ParseResult a

runFieldParser' :: Position -> ParsecParser a -> CabalSpecVersion -> FieldLineStream -> ParseResult a