Distribution.Fields.Field

Description

Cabal-like file AST types: Field, Section etc

These types are parametrized by an annotation.

Cabal file

data Field ann

A Cabal-like file consists of a series of fields (foo: bar) and sections (library ...).

Constructors

Field !(Name ann) [FieldLine ann]

Section !(Name ann) [SectionArg ann] [Field ann]

fieldName :: Field ann -> Name ann

Section of field name

fieldAnn :: Field ann -> ann

fieldUniverse :: Field ann -> [Field ann]

All transitive descendands of Field, including itself.

Note: the resulting list is never empty.

data FieldLine ann

A line of text representing the value of a field from a Cabal file. A field may contain multiple lines.

Invariant: ByteString has no newlines.

Constructors

FieldLine !ann !ByteString

fieldLineAnn :: FieldLine ann -> ann

Since: 3.0.0.0

fieldLineBS :: FieldLine ann -> ByteString

Since: 3.0.0.0

data SectionArg ann

Section arguments, e.g. name of the library

Constructors

SecArgName !ann !ByteString
identifier, or omething which loos like number. Also many dot numbers, i.e. "7.6.3"

SecArgStr !ann !ByteString
quoted string

SecArgOther !ann !ByteString
everything else, mm. operators (e.g. in if-section conditionals)

sectionArgAnn :: SectionArg ann -> ann

Extract annotation from SectionArg.

Name

type FieldName = ByteString

data Name ann

A field name.

Invariant: ByteString is lower-case ASCII.

Constructors

Name !ann !FieldName

mkName :: ann -> FieldName -> Name ann

getName :: Name ann -> FieldName

nameAnn :: Name ann -> ann