Distribution.InstalledPackageInfo

Description

This is the information about an installed package that is communicated to the ghc-pkg program in order to register a package. ghc-pkg now consumes this package format (as of version 6.4). This is specific to GHC at the moment.

The .cabal file format is for describing a package that is not yet installed. It has a lot of flexibility, like conditionals and dependency ranges. As such, that format is not at all suitable for describing a package that has already been built and installed. By the time we get to that stage, we have resolved all conditionals and resolved dependency version constraints to exact versions of dependent packages. So, this module defines the InstalledPackageInfo data structure that contains all the info we keep about an installed package. There is a parser and pretty printer. The textual format is rather simpler than the .cabal format: there are no sections, for example.

Documentation

data InstalledPackageInfo

Constructors

InstalledPackageInfo

Fields

sourcePackageId :: PackageId

installedUnitId :: UnitId

installedComponentId_ :: ComponentId

instantiatedWith :: [(ModuleName, OpenModule)]

compatPackageKey :: String

license :: License

copyright :: String

maintainer :: String

author :: String

stability :: String

homepage :: String

pkgUrl :: String

synopsis :: String

description :: String

category :: String

abiHash :: AbiHash

indefinite :: Bool

exposed :: Bool

exposedModules :: [ExposedModule]

hiddenModules :: [ModuleName]

trusted :: Bool

importDirs :: [FilePath]

libraryDirs :: [FilePath]

libraryDynDirs :: [FilePath]

overrides libraryDirs

dataDir :: FilePath

hsLibraries :: [String]

extraLibraries :: [String]

extraGHCiLibraries :: [String]

includeDirs :: [FilePath]

includes :: [String]

depends :: [UnitId]

abiDepends :: [AbiDependency]

ccOptions :: [String]

ldOptions :: [String]

frameworkDirs :: [FilePath]

frameworks :: [String]

haddockInterfaces :: [FilePath]

haddockHTMLs :: [FilePath]

pkgRoot :: Maybe FilePath

installedPackageId :: InstalledPackageInfo -> UnitId

Deprecated: Use installedUnitId instead

Backwards compatibility with Cabal pre-1.24. This type synonym is slightly awful because in cabal-install we define an InstalledPackageId but it's a ComponentId, not a UnitId!

installedComponentId :: InstalledPackageInfo -> ComponentId

requiredSignatures :: InstalledPackageInfo -> Set ModuleName

Returns the set of module names which need to be filled for an indefinite package, or the empty set if the package is definite.

installedOpenUnitId :: InstalledPackageInfo -> OpenUnitId

Get the indefinite unit identity representing this package. This IS NOT guaranteed to give you a substitution; for instantiated packages you will get DefiniteUnitId (installedUnitId ipi). For indefinite libraries, however, you will correctly get an OpenUnitId with the appropriate OpenModuleSubst.

data ExposedModule

Constructors

ExposedModule

Fields

exposedName :: ModuleName

exposedReexport :: Maybe OpenModule

data AbiDependency

An ABI dependency is a dependency on a library which also records the ABI hash (abiHash) of the library it depends on.

The primary utility of this is to enable an extra sanity when GHC loads libraries: it can check if the dependency has a matching ABI and if not, refuse to load this library. This information is critical if we are shadowing libraries; differences in the ABI hash let us know what packages get shadowed by the new version of a package.

Constructors

AbiDependency

Fields

depUnitId :: UnitId

depAbiHash :: AbiHash

data ParseResult a

Constructors

ParseFailed PError

ParseOk [PWarning] a

data PError

Constructors

AmbiguousParse String LineNo

NoParse String LineNo

TabsError LineNo

FromString String (Maybe LineNo)

data PWarning

emptyInstalledPackageInfo :: InstalledPackageInfo

parseInstalledPackageInfo :: String -> ParseResult InstalledPackageInfo

showInstalledPackageInfo :: InstalledPackageInfo -> String

showInstalledPackageInfoField :: String -> Maybe (InstalledPackageInfo -> String)

showSimpleInstalledPackageInfoField :: String -> Maybe (InstalledPackageInfo -> String)

fieldsInstalledPackageInfo :: [FieldDescr InstalledPackageInfo]