Distribution.Make

Description

This is an alternative build system that delegates everything to the make program. All the commands just end up calling make with appropriate arguments. The intention was to allow preexisting packages that used makefiles to be wrapped into Cabal packages. In practice essentially all such packages were converted over to the "Simple" build system instead. Consequently this module is not used much and it certainly only sees cursory maintenance and no testing. Perhaps at some point we should stop pretending that it works.

Uses the parsed command-line from Distribution.Simple.Setup in order to build Haskell tools using a back-end build system based on make. Obviously we assume that there is a configure script, and that after the ConfigCmd has been run, there is a Makefile. Further assumptions:

ConfigCmd

We assume the configure script accepts --with-hc, --with-hc-pkg, --prefix, --bindir, --libdir, --libexecdir, --datadir.

BuildCmd

We assume that the default Makefile target will build everything.

InstallCmd

We assume there is an install target. Note that we assume that this does *not* register the package!

CopyCmd

We assume there is a copy target, and a variable $(destdir). The copy target should probably just invoke make install recursively (e.g. $(MAKE) install prefix=$(destdir)/$(prefix) bindir=$(destdir)/$(bindir). The reason we can't invoke make install directly here is that we don't know the value of $(prefix).

SDistCmd

We assume there is a dist target.

RegisterCmd

We assume there is a register target and a variable $(user).

UnregisterCmd

We assume there is an unregister target.

HaddockCmd

We assume there is a docs or doc target.

Documentation

module Distribution.Package

data License

Indicates the license under which a package's source code is released. Versions of the licenses not listed here will be rejected by Hackage and cause cabal check to issue a warning.

Constructors

GPL (Maybe Version)
GNU General Public License, version 2 or version 3.

AGPL (Maybe Version)
GNU Affero General Public License, version 3.

LGPL (Maybe Version)
GNU Lesser General Public License, version 2.1 or version 3.

BSD2
2-clause BSD license.

BSD3
3-clause BSD license.

BSD4
4-clause BSD license. This license has not been approved by the OSI and is incompatible with the GNU GPL. It is provided for historical reasons and should be avoided.

MIT
MIT license.

ISC
ISC license

MPL Version
Mozilla Public License, version 2.0.

Apache (Maybe Version)
Apache License, version 2.0.

PublicDomain
The author of a package disclaims any copyright to its source code and dedicates it to the public domain. This is not a software license. Please note that it is not possible to dedicate works to the public domain in every jurisdiction, nor is a work that is in the public domain in one jurisdiction necessarily in the public domain elsewhere.

AllRightsReserved
Explicitly 'All Rights Reserved', eg for proprietary software. The package may not be legally modified or redistributed by anyone but the rightsholder.

UnspecifiedLicense
No license specified which legally defaults to 'All Rights Reserved'. The package may not be legally modified or redistributed by anyone but the rightsholder.

OtherLicense
Any other software license.

UnknownLicense String
Indicates an erroneous license name.

data Version

A Version represents the version of a software entity.

Instances of Eq and Ord are provided, which gives exact equality and lexicographic ordering of the version number components (i.e. 2.1 > 2.0, 1.2.3 > 1.2.2, etc.).

This type is opaque and distinct from the Version type in Data.Version since Cabal-2.0. The difference extends to the Binary instance using a different (and more compact) encoding.

Since: Cabal-syntax-2.0.0.2

defaultMain :: IO ()

defaultMainArgs :: [String] -> IO ()