Distribution.PackageDescription.Parsec

Description

This defined parsers and partial pretty printers for the .cabal format.

Package descriptions

readGenericPackageDescription :: Verbosity -> FilePath -> IO GenericPackageDescription

Parse the given package file.

parseGenericPackageDescription :: ByteString -> ParseResult GenericPackageDescription

Parses the given file into a GenericPackageDescription.

In Cabal 1.2 the syntax for package descriptions was changed to a format with sections and possibly indented property descriptions.

TODO: add lex warnings

parseGenericPackageDescriptionMaybe :: ByteString -> Maybe GenericPackageDescription

Maybe variant of parseGenericPackageDescription

Parsing

data ParseResult a

A monad with failure and accumulating errors and warnings.

runParseResult :: ParseResult a -> ([PWarning], [PError], Maybe a)

Destruct a ParseResult into the emitted warnings and errors, and possibly the final result if there were no errors.

Supplementary build information