Distribution.PackageDescription.Parsec.FieldDescr

Description

GenericPackageDescription Field descriptions

Package description

pkgDescrFieldDescrs :: [FieldDescr PackageDescription]

storeXFieldsPD :: UnknownFieldParser PackageDescription

Store any fields beginning with "x-" in the customFields field of a PackageDescription. All other fields will generate a warning.

Library

libFieldDescrs :: [FieldDescr Library]

storeXFieldsLib :: UnknownFieldParser Library

Foreign library

foreignLibFieldDescrs :: [FieldDescr ForeignLib]

storeXFieldsForeignLib :: UnknownFieldParser ForeignLib

Executable

executableFieldDescrs :: [FieldDescr Executable]

storeXFieldsExe :: UnknownFieldParser Executable

Test suite

data TestSuiteStanza

An intermediate type just used for parsing the test-suite stanza. After validation it is converted into the proper TestSuite type.

Constructors

TestSuiteStanza

Fields

testStanzaTestType :: Maybe TestType

testStanzaMainIs :: Maybe FilePath

testStanzaTestModule :: Maybe ModuleName

testStanzaBuildInfo :: BuildInfo

emptyTestStanza :: TestSuiteStanza

testSuiteFieldDescrs :: [FieldDescr TestSuiteStanza]

storeXFieldsTest :: UnknownFieldParser TestSuiteStanza

validateTestSuite :: Position -> TestSuiteStanza -> ParseResult TestSuite

Benchmark

data BenchmarkStanza

An intermediate type just used for parsing the benchmark stanza. After validation it is converted into the proper Benchmark type.

Constructors

BenchmarkStanza

Fields

benchmarkStanzaBenchmarkType :: Maybe BenchmarkType

benchmarkStanzaMainIs :: Maybe FilePath

benchmarkStanzaBenchmarkModule :: Maybe ModuleName

benchmarkStanzaBuildInfo :: BuildInfo

emptyBenchmarkStanza :: BenchmarkStanza

benchmarkFieldDescrs :: [FieldDescr BenchmarkStanza]

storeXFieldsBenchmark :: UnknownFieldParser BenchmarkStanza

validateBenchmark :: Position -> BenchmarkStanza -> ParseResult Benchmark

Flag

flagFieldDescrs :: [FieldDescr Flag]

Source repository

sourceRepoFieldDescrs :: [FieldDescr SourceRepo]

Setup build info

setupBInfoFieldDescrs :: [FieldDescr SetupBuildInfo]