Distribution.PackageDescription.Quirks

Description

Since: 2.2.0.0

Documentation

patchQuirks :: ByteString -> (Bool, ByteString)

Patch legacy .cabal file contents to allow parsec parser to accept all of Hackage.

Bool part of the result tells whether the output is modified.

Since: 2.2.0.0