Distribution.Parsec.Common

Description

Module containing small types

Diagnostics

data PError

Parser error.

Constructors

PError Position String

showPError :: FilePath -> PError -> String

data PWarning

Parser warning.

Constructors

PWarning !PWarnType !Position String

data PWarnType

Type of parser warning. We do classify warnings.

Different application may decide not to show some, or have fatal behaviour on others

Constructors

PWTOther
Unclassified warning

PWTUTF
Invalid UTF encoding

PWTBoolCase
true or false, not True or False

PWTVersionTag
there are version with tags

PWTNewSyntax
New syntax used, but no cabal-version: >= 1.2 specified

PWTOldSyntax
Old syntax used, and cabal-version >= 1.2 specified

PWTDeprecatedField

PWTInvalidSubsection

PWTUnknownField

PWTUnknownSection

PWTTrailingFields

PWTExtraMainIs
extra main-is field

PWTExtraTestModule
extra test-module field

PWTExtraBenchmarkModule
extra benchmark-module field

PWTLexNBSP

PWTLexBOM

PWTLexTab

PWTQuirkyCabalFile
legacy cabal file that we know how to patch

PWTDoubleDash
Double dash token, most likely it's a mistake - it's not a comment

PWTMultipleSingularField
e.g. name or version should be specified only once.

PWTBuildTypeDefault
Workaround for derive-package having build-type: Default. See https://github.com/haskell/cabal/issues/5020.

PWTVersionLeadingZeros
See https://github.com/haskell-infra/hackage-trustees/issues/128

showPWarning :: FilePath -> PWarning -> String

Position

data Position

1-indexed row and column positions in a file.

Constructors

Position !Int !Int

incPos :: Int -> Position -> Position

Shift position by n columns to the right.

retPos :: Position -> Position

Shift position to beginning of next row.

showPos :: Position -> String

zeroPos :: Position