Distribution.Parsec.Newtypes

Description

This module provides newtype wrappers to be used with Distribution.FieldGrammar.

List

alaList :: sep -> [a] -> List sep (Identity a) a

alaList and alaList' are simply List, with additional phantom arguments to constraint the resulting type

>>> :t alaList VCat
alaList VCat :: [a] -> List VCat (Identity a) a

>>> :t alaList' FSep Token
alaList' FSep Token :: [String] -> List FSep Token String

alaList' :: sep -> (a -> b) -> [a] -> List sep b a

More general version of alaList.

Modifiers

data CommaVCat

Vertical list with commas. Displayed with vcat

Constructors

CommaVCat

data CommaFSep

Paragraph fill list with commas. Displayed with fsep

Constructors

CommaFSep

data VCat

Vertical list with optional commas. Displayed with vcat.

Constructors

VCat

data FSep

Paragraph fill list with optional commas. Displayed with fsep.

Constructors

FSep

data NoCommaFSep

Paragraph fill list without commas. Displayed with fsep.

Constructors

NoCommaFSep

Type

data List sep b a

List separated with optional commas. Displayed with sep, arguments of type a are parsed and pretty-printed as b.

Version & License

newtype SpecVersion

Version range or just version, i.e. cabal-version field.

There are few things to consider:

Starting with 2.2 the cabal-version field should be the first field in the file and only exact version is accepted. Therefore if we get e.g. >= 2.2, we fail. See https://github.com/haskell/cabal/issues/4899

Constructors

SpecVersion

Fields

getSpecVersion :: Either Version VersionRange

newtype TestedWith

Version range or just version

Constructors

TestedWith

Fields

getTestedWith :: (CompilerFlavor, VersionRange)

newtype SpecLicense

SPDX License expression or legacy license

Constructors

SpecLicense

Fields

getSpecLicense :: Either License License

Identifiers

newtype Token

Haskell string or [^ ,]+

Constructors

Token

Fields

getToken :: String

newtype Token'

Haskell string or [^ ]+

Constructors

Token'

Fields

getToken' :: String

newtype MQuoted a

Either "quoted" or un-quoted.

Constructors

MQuoted

Fields

getMQuoted :: a

newtype FreeText

This is almost many anyChar, but it

trims whitespace from ends of the lines,

converts lines with only single dot into empty line.

Constructors

FreeText

Fields

getFreeText :: String

newtype FilePathNT

Filepath are parsed as Token.

Constructors

FilePathNT

Fields

getFilePathNT :: String