Distribution.Parsec.Types.Field

Description

Cabal-like file AST types: Field, Section etc

These types are parametrized by an annotation.

Cabal file

data Field ann

A Cabal-like file consists of a series of fields (foo: bar) and sections (library ...).

Constructors

Field !(Name ann) [FieldLine ann]

Section !(Name ann) [SectionArg ann] [Field ann]

fieldAnn :: Field ann -> ann

data FieldLine ann

A line of text representing the value of a field from a Cabal file. A field may contain multiple lines.

Invariant: ByteString has no newlines.

Constructors

FieldLine !ann !ByteString

data SectionArg ann

Section arguments, e.g. name of the library

Constructors

SecArgName !ann !ByteString
identifier

SecArgStr !ann !String
quoted string

SecArgNum !ann !ByteString
Something which loos like number. Also many dot numbers, i.e. "7.6.3"

SecArgOther !ann !ByteString
everything else, mm. operators (e.g. in if-section conditionals)

sectionArgAnn :: SectionArg ann -> ann

Extract annotation from SectionArg.

Name

data Name ann

A field name.

Invariant: ByteString is lower-case ASCII.

Constructors

Name !ann !ByteString

mkName :: ann -> ByteString -> Name ann

getName :: Name ann -> ByteString

nameAnn :: Name ann -> ann