Distribution.Parsec.Types.FieldDescr

Description

Field name

type FieldName = ByteString

Field description

data FieldDescr a

Constructors

FieldDescr

Fields

fieldName :: FieldName

fieldPretty :: a -> Doc

fieldParser :: a -> FieldParser a

liftField :: (b -> a) -> (a -> b -> b) -> FieldDescr a -> FieldDescr b

simpleField

Arguments

:: FieldName
fieldname

-> (a -> Doc)
show

-> FieldParser a
parser

-> (b -> a)
getter

-> (a -> b -> b)
setter

-> FieldDescr b

boolField :: FieldName -> (b -> Bool) -> (Bool -> b -> b) -> FieldDescr b

optsField :: FieldName -> CompilerFlavor -> (b -> [(CompilerFlavor, [String])]) -> ([(CompilerFlavor, [String])] -> b -> b) -> FieldDescr b

listField :: FieldName -> (a -> Doc) -> FieldParser a -> (b -> [a]) -> ([a] -> b -> b) -> FieldDescr b

listFieldWithSep :: Separator -> FieldName -> (a -> Doc) -> FieldParser a -> (b -> [a]) -> ([a] -> b -> b) -> FieldDescr b

commaListField :: FieldName -> (a -> Doc) -> FieldParser a -> (b -> [a]) -> ([a] -> b -> b) -> FieldDescr b

commaListFieldWithSep :: Separator -> FieldName -> (a -> Doc) -> FieldParser a -> (b -> [a]) -> ([a] -> b -> b) -> FieldDescr b

spaceListField :: FieldName -> (a -> Doc) -> FieldParser a -> (b -> [a]) -> ([a] -> b -> b) -> FieldDescr b

Pretty printing

ppFields :: [FieldDescr a] -> a -> Doc

ppField :: FieldName -> Doc -> Doc

Unknown fields

type UnknownFieldParser a = FieldName -> String -> a -> Maybe a

How to handle unknown fields.

warnUnrec :: UnknownFieldParser a

A default unrecognized field parser which simply returns Nothing, i.e. ignores all unrecognized fields, so warnings will be generated.

ignoreUnrec :: UnknownFieldParser a

A default unrecognized field parser which silently (i.e. no warnings will be generated) ignores unrecognized fields, by returning the structure being built unmodified.