Distribution.Parsec.Types.ParseResult

Description

A parse result type for parsers from AST to Haskell types.

Documentation

data ParseResult a

A monad with failure and accumulating errors and warnings.

runParseResult :: ParseResult a -> ([PWarning], [PError], Maybe a)

Destruct a ParseResult into the emitted warnings and errors, and possibly the final result if there were no errors.

recoverWith :: ParseResult a -> a -> ParseResult a

Recover the parse result, so we can proceed parsing. runParseResult will still result in Nothing, if there are recorded errors.

parseWarning :: Position -> PWarnType -> String -> ParseResult ()

parseFailure :: Position -> String -> ParseResult ()

Add an error, but not fail the parser yet.

For fatal failure use parseFatalFailure

parseFatalFailure :: Position -> String -> ParseResult a

Add an fatal error.

parseFatalFailure' :: ParseResult a

parseWarnings' :: [PWarning] -> ParseResult ()