Distribution.Pretty

Documentation

class Pretty a where

Minimal complete definition

pretty

Methods

pretty :: a -> Doc

prettyVersioned :: CabalSpecVersion -> a -> Doc

prettyShow :: Pretty a => a -> String

defaultStyle :: Style

The default rendering style used in Cabal for console output. It has a fixed page width and adds line breaks automatically.

flatStyle :: Style

A style for rendering all on one line.

Utilities

showFilePath :: FilePath -> Doc

showToken :: String -> Doc

showFreeText :: String -> Doc

Pretty-print free-format text, ensuring that it is vertically aligned, and with blank lines replaced by dots for correct re-parsing.

showFreeTextV3 :: String -> Doc

Pretty-print free-format text. Since cabal-version: 3.0 we don't replace blank lines with dots.

Since: Cabal-3.0.0.0

Deprecated

type Separator = [Doc] -> Doc