Distribution.Simple.Build.Inputs

Inputs of actions for building components

data PreBuildComponentInputs

The information required for a build computation which is available right before building each component, i.e. the pre-build component inputs.

Constructors

PreBuildComponentInputs

Fields

buildingWhat :: BuildingWhat

What kind of build are we doing?

localBuildInfo :: LocalBuildInfo

Information about the package

targetInfo :: TargetInfo

Information about an individual component

Queries over the component being built

buildVerbosity :: PreBuildComponentInputs -> Verbosity

Get the Verbosity from the context the component being built is in.

buildComponent :: PreBuildComponentInputs -> Component

Get the Component being built.

buildIsLib :: PreBuildComponentInputs -> Bool

Is the Component being built a Library?

buildCLBI :: PreBuildComponentInputs -> ComponentLocalBuildInfo

Get the ComponentLocalBuildInfo for the component being built.

buildBI :: PreBuildComponentInputs -> BuildInfo

Get the BuildInfo of the component being built.

buildCompiler :: PreBuildComponentInputs -> Compiler

Get the Compiler being used to build the component.

Re-exports

data BuildingWhat

What kind of build phase are we doing/hooking into?

Is this a normal build, or is it perhaps for running an interactive session or Haddock?

Constructors

BuildNormal BuildFlags
A normal build.

BuildRepl ReplFlags
Build steps for an interactive session.

BuildHaddock HaddockFlags
Build steps for generating documentation.

BuildHscolour HscolourFlags
Build steps for Hscolour.

data LocalBuildInfo

Data cached after configuration step. See also ConfigFlags.

Constructors

NewLocalBuildInfo

Fields

localBuildDescr :: LocalBuildDescr

Information about a package determined by Cabal after the configuration step.

localBuildConfig :: LocalBuildConfig

Information about a package configuration that can be modified by the user at configuration time.

Bundled Patterns

pattern LocalBuildInfo :: ConfigFlags -> FlagAssignment -> ComponentRequestedSpec -> [String] -> InstallDirTemplates -> Compiler -> Platform -> Maybe (SymbolicPath Pkg 'File) -> Graph ComponentLocalBuildInfo -> Map ComponentName [ComponentLocalBuildInfo] -> Map (PackageName, ComponentName) PromisedComponent -> InstalledPackageIndex -> PackageDescription -> ProgramDb -> PackageDBStack -> Bool -> Bool -> Bool -> Bool -> Bool -> Bool -> Bool -> Bool -> ProfDetailLevel -> ProfDetailLevel -> OptimisationLevel -> DebugInfoLevel -> Bool -> Bool -> Bool -> Bool -> Bool -> Bool -> Bool -> [UnitId] -> Bool -> LocalBuildInfo
This pattern synonym is for backwards compatibility, to adapt to LocalBuildInfo being split into LocalBuildDescr and LocalBuildConfig.

data TargetInfo

The TargetInfo contains all the information necessary to build a specific target (e.g., componentmodulefile) in a package. In principle, one can get the Component from a ComponentLocalBuildInfo and LocalBuildInfo, but it is much more convenient to have the component in hand.

Constructors

TargetInfo

Fields

targetCLBI :: ComponentLocalBuildInfo

targetComponent :: Component

buildingWhatCommonFlags :: BuildingWhat -> CommonSetupFlags

buildingWhatVerbosity :: BuildingWhat -> Verbosity

buildingWhatWorkingDir :: BuildingWhat -> Maybe (SymbolicPath CWD ('Dir Pkg))

buildingWhatDistPref :: BuildingWhat -> SymbolicPath Pkg ('Dir Dist)