Distribution.Simple.Compiler

Description

This should be a much more sophisticated abstraction than it is. Currently it's just a bit of data about the compiler, like its flavour and name and version. The reason it's just data is because currently it has to be in Read and Show so it can be saved along with the LocalBuildInfo. The only interesting bit of info it contains is a mapping between language extensions and compiler command line flags. This module also defines a PackageDB type which is used to refer to package databases. Most compilers only know about a single global package collection but GHC has a global and per-user one and it lets you create arbitrary other package databases. We do not yet fully support this latter feature.

Haskell implementations

module Distribution.Compiler

data Compiler

Constructors

Compiler

Fields

compilerId :: CompilerId

Compiler flavour and version.

compilerAbiTag :: AbiTag

Tag for distinguishing incompatible ABI's on the same architecture/os.

compilerCompat :: [CompilerId]

Other implementations that this compiler claims to be compatible with.

compilerLanguages :: [(Language, CompilerFlag)]

Supported language standards.

compilerExtensions :: [(Extension, Maybe CompilerFlag)]

Supported extensions.

compilerProperties :: Map String String

A key-value map for properties not covered by the above fields.

showCompilerId :: Compiler -> String

showCompilerIdWithAbi :: Compiler -> String

compilerFlavor :: Compiler -> CompilerFlavor

compilerVersion :: Compiler -> Version

compilerCompatFlavor :: CompilerFlavor -> Compiler -> Bool

Is this compiler compatible with the compiler flavour we're interested in?

For example this checks if the compiler is actually GHC or is another compiler that claims to be compatible with some version of GHC, e.g. GHCJS.

if compilerCompatFlavor GHC compiler then ... else ...

compilerCompatVersion :: CompilerFlavor -> Compiler -> Maybe Version

Is this compiler compatible with the compiler flavour we're interested in, and if so what version does it claim to be compatible with.

For example this checks if the compiler is actually GHC-7.x or is another compiler that claims to be compatible with some GHC-7.x version.

case compilerCompatVersion GHC compiler of
  Just (Version (7:_)) -> ...
  _                    -> ...

compilerInfo :: Compiler -> CompilerInfo

Support for package databases

type PackageDB = PackageDBX (SymbolicPath Pkg ('Dir PkgDB))

type PackageDBStack = PackageDBStackX (SymbolicPath Pkg ('Dir PkgDB))

type PackageDBCWD = PackageDBX FilePath

type PackageDBStackCWD = PackageDBStackX FilePath

data PackageDBX fp

Some compilers have a notion of a database of available packages. For some there is just one global db of packages, other compilers support a per-user or an arbitrary db specified at some location in the file system. This can be used to build isolated environments of packages, for example to build a collection of related packages without installing them globally.

Abstracted over

Constructors

GlobalPackageDB

UserPackageDB

SpecificPackageDB fp
NB: the path might be relative or it might be absolute

type PackageDBStackX from = [PackageDBX from]

We typically get packages from several databases, and stack them together. This type lets us be explicit about that stacking. For example typical stacks include:

[GlobalPackageDB]
[GlobalPackageDB, UserPackageDB]
[GlobalPackageDB, SpecificPackageDB "package.conf.inplace"]

Note that the GlobalPackageDB is invariably at the bottom since it contains the rts, base and other special compiler-specific packages.

We are not restricted to using just the above combinations. In particular we can use several custom package dbs and the user package db together.

When it comes to writing, the top most (last) package is used.

type PackageDBS from = PackageDBX (SymbolicPath from ('Dir PkgDB))

type PackageDBStackS from = PackageDBStackX (SymbolicPath from ('Dir PkgDB))

registrationPackageDB :: PackageDBStackX from -> PackageDBX from

Return the package that we should register into. This is the package db at the top of the stack.

absolutePackageDBPaths :: Maybe (SymbolicPath CWD ('Dir Pkg)) -> PackageDBStack -> IO PackageDBStack

Make package paths absolute

absolutePackageDBPath :: Maybe (SymbolicPath CWD ('Dir Pkg)) -> PackageDB -> IO PackageDB

interpretPackageDB :: Maybe (SymbolicPath CWD ('Dir Pkg)) -> PackageDB -> PackageDBCWD

interpretPackageDBStack :: Maybe (SymbolicPath CWD ('Dir Pkg)) -> PackageDBStack -> PackageDBStackCWD

coercePackageDB :: PackageDBCWD -> PackageDBX (SymbolicPath CWD ('Dir PkgDB))

Transform a package db using a FilePath into one using symbolic paths.

coercePackageDBStack :: [PackageDBCWD] -> [PackageDBX (SymbolicPath CWD ('Dir PkgDB))]

Support for optimisation levels

data OptimisationLevel

Some compilers support optimising. Some have different levels. For compilers that do not the level is just capped to the level they do support.

Constructors

NoOptimisation

NormalOptimisation

MaximumOptimisation

flagToOptimisationLevel :: Maybe String -> OptimisationLevel

Support for debug info levels

data DebugInfoLevel

Some compilers support emitting debug info. Some have different levels. For compilers that do not the level is just capped to the level they do support.

Constructors

NoDebugInfo

MinimalDebugInfo

NormalDebugInfo

MaximalDebugInfo

flagToDebugInfoLevel :: Maybe String -> DebugInfoLevel

Support for language extensions

type CompilerFlag = String

languageToFlags :: Compiler -> Maybe Language -> [CompilerFlag]

unsupportedLanguages :: Compiler -> [Language] -> [Language]

extensionsToFlags :: Compiler -> [Extension] -> [CompilerFlag]

For the given compiler, return the flags for the supported extensions.

unsupportedExtensions :: Compiler -> [Extension] -> [Extension]

For the given compiler, return the extensions it does not support.

parmakeSupported :: Compiler -> Bool

Does this compiler support parallel --make mode?

reexportedModulesSupported :: Compiler -> Bool

Does this compiler support reexported-modules?

renamingPackageFlagsSupported :: Compiler -> Bool

Does this compiler support thinning/renaming on package flags?

unifiedIPIDRequired :: Compiler -> Bool

Does this compiler have unified IPIDs (so no package keys)

packageKeySupported :: Compiler -> Bool

Does this compiler support package keys?

unitIdSupported :: Compiler -> Bool

Does this compiler support unit IDs?

coverageSupported :: Compiler -> Bool

Does this compiler support Haskell program coverage?

profilingSupported :: Compiler -> Bool

Does this compiler support profiling?

profilingDynamicSupported :: Compiler -> Maybe Bool

Is the compiler distributed with profiling dynamic libraries

profilingDynamicSupportedOrUnknown :: Compiler -> Bool

Either profiling dynamic is definitely supported or we don't know (so assume it is)

profilingVanillaSupported :: Compiler -> Maybe Bool

Is the compiler distributed with profiling libraries

profilingVanillaSupportedOrUnknown :: Compiler -> Bool

Either profiling is definitely supported or we don't know (so assume it is)

dynamicSupported :: Compiler -> Maybe Bool

Is the compiler distributed with dynamic libraries

backpackSupported :: Compiler -> Bool

Does this compiler support Backpack?

arResponseFilesSupported :: Compiler -> Bool

Does this compiler's "ar" command supports response file arguments (i.e. @file-style arguments).

arDashLSupported :: Compiler -> Bool

Does this compiler's "ar" command support llvm-ar's -L flag, which compels the archiver to add an input archive's members rather than adding the archive itself.

libraryDynDirSupported :: Compiler -> Bool

Does this compiler support a package database entry with: "dynamic-library-dirs"?

libraryVisibilitySupported :: Compiler -> Bool

Does this compiler support a package database entry with: "visibility"?

jsemSupported :: Compiler -> Bool

Does this compiler support the -jsem option?

Support for profiling detail levels

data ProfDetailLevel

Some compilers (notably GHC) support profiling and can instrument programs so the system can account costs to different functions. There are different levels of detail that can be used for this accounting. For compilers that do not support this notion or the particular detail levels, this is either ignored or just capped to some similar level they do support.

Constructors

ProfDetailNone

ProfDetailDefault

ProfDetailExportedFunctions

ProfDetailToplevelFunctions

ProfDetailAllFunctions

ProfDetailTopLate

ProfDetailOther String

knownProfDetailLevels :: [(String, [String], ProfDetailLevel)]

flagToProfDetailLevel :: String -> ProfDetailLevel

showProfDetailLevel :: ProfDetailLevel -> String