Distribution.Simple.FileMonitor.Types

Description

Types for monitoring files and directories.

Globs with respect to a root

data RootedGlob

A file path specified by globbing, relative to some root directory.

Constructors

RootedGlob

Fields

FilePathRoot

what the glob is relative to

Glob

the glob

data FilePathRoot

Constructors

FilePathRelative

FilePathRoot FilePath
e.g. "/", "c:" or result of takeDrive

FilePathHomeDir

data Glob

A filepath specified by globbing.

File monitoring

data MonitorFilePath

A description of a file (or set of files) to monitor for changes.

Where file paths are relative they are relative to a common directory (e.g. project root), not necessarily the process current directory.

Constructors

MonitorFile

Fields

monitorKindFile :: !MonitorKindFile

monitorKindDir :: !MonitorKindDir

monitorPath :: !FilePath

MonitorFileGlob

Fields

monitorKindFile :: !MonitorKindFile

monitorKindDir :: !MonitorKindDir

monitorPathGlob :: !RootedGlob

data MonitorKindFile

Constructors

FileExists

FileModTime

FileHashed

FileNotExists

data MonitorKindDir

Constructors

DirExists

DirModTime

DirNotExists

Utility constructors of MonitorFilePath

monitorFile :: FilePath -> MonitorFilePath

Monitor a single file for changes, based on its modification time. The monitored file is considered to have changed if it no longer exists or if its modification time has changed.

monitorFileHashed :: FilePath -> MonitorFilePath

Monitor a single file for changes, based on its modification time and content hash. The monitored file is considered to have changed if it no longer exists or if its modification time and content hash have changed.

monitorNonExistentFile :: FilePath -> MonitorFilePath

Monitor a single non-existent file for changes. The monitored file is considered to have changed if it exists.

monitorFileExistence :: FilePath -> MonitorFilePath

Monitor a single file for existence only. The monitored file is considered to have changed if it no longer exists.

monitorDirectory :: FilePath -> MonitorFilePath

Monitor a single directory for changes, based on its modification time. The monitored directory is considered to have changed if it no longer exists or if its modification time has changed.

monitorNonExistentDirectory :: FilePath -> MonitorFilePath

Monitor a single non-existent directory for changes. The monitored directory is considered to have changed if it exists.

monitorDirectoryExistence :: FilePath -> MonitorFilePath

Monitor a single directory for existence. The monitored directory is considered to have changed only if it no longer exists.

monitorFileOrDirectory :: FilePath -> MonitorFilePath

Monitor a single file or directory for changes, based on its modification time. The monitored file is considered to have changed if it no longer exists or if its modification time has changed.

monitorFileGlob :: RootedGlob -> MonitorFilePath

Monitor a set of files (or directories) identified by a file glob. The monitored glob is considered to have changed if the set of files matching the glob changes (i.e. creations or deletions), or for files if the modification time and content hash of any matching file has changed.

monitorFileGlobExistence :: RootedGlob -> MonitorFilePath

Monitor a set of files (or directories) identified by a file glob for existence only. The monitored glob is considered to have changed if the set of files matching the glob changes (i.e. creations or deletions).

monitorFileSearchPath :: [FilePath] -> FilePath -> [MonitorFilePath]

Creates a list of files to monitor when you search for a file which unsuccessfully looked in notFoundAtPaths before finding it at foundAtPath.

monitorFileHashedSearchPath :: [FilePath] -> FilePath -> [MonitorFilePath]

Similar to monitorFileSearchPath, but also instructs us to monitor the hash of the found file.