Distribution.Simple.GHCJS

Documentation

configure :: Verbosity -> Maybe FilePath -> Maybe FilePath -> ProgramDb -> IO (Compiler, Maybe Platform, ProgramDb)

getInstalledPackages :: Verbosity -> PackageDBStack -> ProgramDb -> IO InstalledPackageIndex

Given a package DB stack, return all installed packages.

getPackageDBContents :: Verbosity -> PackageDB -> ProgramDb -> IO InstalledPackageIndex

Given a single package DB, return all installed packages.

buildLib :: Verbosity -> Flag (Maybe Int) -> PackageDescription -> LocalBuildInfo -> Library -> ComponentLocalBuildInfo -> IO ()

buildExe :: Verbosity -> Flag (Maybe Int) -> PackageDescription -> LocalBuildInfo -> Executable -> ComponentLocalBuildInfo -> IO ()

replLib :: Verbosity -> Flag (Maybe Int) -> PackageDescription -> LocalBuildInfo -> Library -> ComponentLocalBuildInfo -> IO ()

replExe :: Verbosity -> Flag (Maybe Int) -> PackageDescription -> LocalBuildInfo -> Executable -> ComponentLocalBuildInfo -> IO ()

startInterpreter :: Verbosity -> ProgramDb -> Compiler -> Platform -> PackageDBStack -> IO ()

Start a REPL without loading any source files.

installLib

Arguments

:: Verbosity

-> LocalBuildInfo

-> FilePath
install location

-> FilePath
install location for dynamic libraries

-> FilePath
Build location

-> PackageDescription

-> Library

-> ComponentLocalBuildInfo

-> IO ()

Install for ghc, .hi, .a and, if --with-ghci given, .o

installExe

Arguments

:: Verbosity

-> LocalBuildInfo

-> FilePath
Where to copy the files to

-> FilePath
Build location

-> (FilePath, FilePath)
Executable (prefix,suffix)

-> PackageDescription

-> Executable

-> IO ()

libAbiHash :: Verbosity -> PackageDescription -> LocalBuildInfo -> Library -> ComponentLocalBuildInfo -> IO String

hcPkgInfo :: ProgramDb -> HcPkgInfo

registerPackage :: Verbosity -> ProgramDb -> PackageDBStack -> InstalledPackageInfo -> RegisterOptions -> IO ()

componentGhcOptions :: Verbosity -> LocalBuildInfo -> BuildInfo -> ComponentLocalBuildInfo -> FilePath -> GhcOptions

getLibDir :: Verbosity -> LocalBuildInfo -> IO FilePath

isDynamic :: Compiler -> Bool

getGlobalPackageDB :: Verbosity -> ConfiguredProgram -> IO FilePath

Return the FilePath to the global GHC package database.

runCmd :: ProgramDb -> FilePath -> (FilePath, FilePath, [String])

Get the JavaScript file name and command and arguments to run a program compiled by GHCJS the exe should be the base program name without exe extension