Distribution.Simple.Glob.Internal

Description

Internal module for simple file globbing. Please import Distribution.Simple.Glob instead.

Documentation

data Glob

A filepath specified by globbing.

Constructors

GlobDir !GlobPieces !Glob
dirGlob/glob

GlobDirRecursive !GlobPieces
**/glob, where ** denotes recursively traversing all directories and matching filenames on glob.

GlobFile !GlobPieces
A file glob.

GlobDirTrailing
Trailing dir; a glob ending in /.

type GlobPieces = [GlobPiece]

A single directory or file component of a globbed path

data GlobPiece

A piece of a globbing pattern

Constructors

WildCard
A wildcard *

Literal String
A literal string dirABC

Union [GlobPieces]
A union of patterns, e.g. dir{a,*.txt,c}...

dispGlobPieces :: GlobPieces -> Doc

isGlobEscapedChar :: Char -> Bool