Distribution.Simple.Haddock

Description

This module deals with the haddock and hscolour commands. It uses information about installed packages (from ghc-pkg) to find the locations of documentation for dependent packages, so it can create links.

The hscolour support allows generating HTML versions of the original source, with coloured syntax highlighting.

Documentation

haddock :: PackageDescription -> LocalBuildInfo -> [PPSuffixHandler] -> HaddockFlags -> IO ()

createHaddockIndex :: Verbosity -> ProgramDb -> Compiler -> Platform -> HaddockProjectFlags -> IO ()

Execute Haddock configured with HaddocksFlags. It is used to build index and contents for documentation of multiple packages.

hscolour :: PackageDescription -> LocalBuildInfo -> [PPSuffixHandler] -> HscolourFlags -> IO ()

haddockPackagePaths :: [InstalledPackageInfo] -> Maybe (InstalledPackageInfo -> FilePath) -> IO ([(FilePath, Maybe FilePath, Maybe FilePath, Visibility)], Maybe String)

Given a list of InstalledPackageInfos, return a list of interfaces and HTML paths, and an optional warning for packages with missing documentation.

data Visibility

Governs whether modules from a given interface should be visible or hidden in the Haddock generated content page. We don't expose this functionality to the user, but simply use Visible for only local packages. Visibility of modules is available since haddock-2.26.1.

Constructors

Visible

Hidden