Distribution.Simple.Install

Description

This is the entry point into installing a built package. Performs the "./setup install" and "./setup copy" actions. It moves files into place based on the prefix argument. It does the generic bits and then calls compiler-specific functions to do the rest.

Documentation

install

Arguments

:: PackageDescription
information from the .cabal file

-> LocalBuildInfo
information from the configure step

-> CopyFlags
flags sent to copy or install

-> IO ()

Perform the "./setup install" and "./setup copy" actions. Move files into place based on the prefix argument.

This does NOT register libraries, you should call register to do that.