Distribution.Simple.LHC

Description

This is a fairly large module. It contains most of the GHC-specific code for configuring, building and installing packages. It also exports a function for finding out what packages are already installed. Configuring involves finding the ghc and ghc-pkg programs, finding what language extensions this version of ghc supports and returning a Compiler value.

getInstalledPackages involves calling the ghc-pkg program to find out what packages are installed.

Building is somewhat complex as there is quite a bit of information to take into account. We have to build libs and programs, possibly for profiling and shared libs. We have to support building libraries that will be usable by GHCi and also ghc's -split-objs feature. We have to compile any C files using ghc. Linking, especially for split-objs is remarkably complex, partly because there tend to be 1,000's of .o files and this can often be more than we can pass to the ld or ar programs in one go.

Installing for libs and exes involves finding the right files and copying them to the right places. One of the more tricky things about this module is remembering the layout of files in the build directory (which is not explicitly documented) and thus what search dirs are used for various kinds of files.

Documentation

configure :: Verbosity -> Maybe FilePath -> Maybe FilePath -> ProgramDb -> IO (Compiler, Maybe Platform, ProgramDb)

getInstalledPackages :: Verbosity -> PackageDBStack -> ProgramDb -> IO InstalledPackageIndex

buildLib :: Verbosity -> PackageDescription -> LocalBuildInfo -> Library -> ComponentLocalBuildInfo -> IO ()

Build a library with LHC.

buildExe :: Verbosity -> PackageDescription -> LocalBuildInfo -> Executable -> ComponentLocalBuildInfo -> IO ()

Build an executable with LHC.

installLib

Arguments

:: Verbosity

-> LocalBuildInfo

-> FilePath
install location

-> FilePath
install location for dynamic libraries

-> FilePath
Build location

-> PackageDescription

-> Library

-> ComponentLocalBuildInfo

-> IO ()

Install for ghc, .hi, .a and, if --with-ghci given, .o

installExe

Arguments

:: Verbosity

-> LocalBuildInfo

-> FilePath
Where to copy the files to

-> FilePath
Build location

-> (FilePath, FilePath)
Executable (prefix,suffix)

-> PackageDescription

-> Executable

-> IO ()

Install executables for GHC.

registerPackage :: Verbosity -> ProgramDb -> PackageDBStack -> InstalledPackageInfo -> RegisterOptions -> IO ()

hcPkgInfo :: ProgramDb -> HcPkgInfo

ghcOptions :: LocalBuildInfo -> BuildInfo -> ComponentLocalBuildInfo -> FilePath -> [String]

ghcVerbosityOptions :: Verbosity -> [String]