Distribution.Simple.Program.Hpc

Description

This module provides an library interface to the hpc program.

Documentation

markup

Arguments

:: ConfiguredProgram

-> Version

-> Verbosity

-> FilePath
Path to .tix file

-> [FilePath]
Paths to .mix file directories

-> FilePath
Path where html output should be located

-> [ModuleName]
List of modules to include in the report

-> IO ()

Invoke hpc with the given parameters.

Prior to HPC version 0.7 (packaged with GHC 7.8), hpc did not handle multiple .mix paths correctly, so we print a warning, and only pass it the first path in the list. This means that e.g. test suites that import their library as a dependency can still work, but those that include the library modules directly (in other-modules) don't.

union

Arguments

:: ConfiguredProgram

-> Verbosity

-> [FilePath]
Paths to .tix files

-> FilePath
Path to resultant .tix file

-> [ModuleName]
List of modules to exclude from union

-> IO ()