Distribution.Simple.Program.Ld

Description

This module provides an library interface to the ld linker program.

Documentation

combineObjectFiles :: Verbosity -> ConfiguredProgram -> FilePath -> [FilePath] -> IO ()

Call ld -r to link a bunch of object files together.