Distribution.Simple.SetupHooks.Errors

Documentation

data SetupHooksException

An error involving the SetupHooks module of a package with Hooks build-type.

Constructors

CannotApplyComponentDiff CannotApplyComponentDiffReason
Cannot apply a diff to a component in a per-component configure hook.

RulesException RulesException
An error with pre-build rules.

data CannotApplyComponentDiffReason

Constructors

MismatchedComponentTypes Component Component

IllegalComponentDiff Component (NonEmpty IllegalComponentDiffReason)

data IllegalComponentDiffReason

Constructors

CannotChangeName

CannotChangeComponentField String

CannotChangeBuildInfoField String

data RulesException

AN error involving the Rules in the SetupHooks module of a package with the Hooks build-type.

Constructors

CyclicRuleDependencies (NonEmpty (RuleBinary, [Tree RuleBinary]))
There are cycles in the dependency graph of fine-grained rules.

CantFindSourceForRuleDependencies
When executing fine-grained rules compiled into the external hooks executable, we failed to find dependencies of a rule.

Fields

RuleBinary

(NonEmpty Location)

missing dependencies

MissingRuleOutputs
When executing fine-grained rules compiled into the external hooks executable, a rule failed to generate the outputs it claimed it would.

Fields

RuleBinary

(NonEmpty Location)

missing outputs

InvalidRuleOutputIndex
An invalid reference to a rule output, e.g. an out-of-range index.

Fields

RuleId

rule

RuleId

dependency

(NonEmpty Location)

outputs of dependency

Word

the invalid index

DuplicateRuleId !RuleId !Rule !Rule
A duplicate RuleId in the construction of pre-build rules.

setupHooksExceptionCode :: SetupHooksException -> Int

setupHooksExceptionMessage :: SetupHooksException -> String