Distribution.Simple.SetupHooks.Internal

Description

Internal implementation module. Users of build-type: Hooks should import Distribution.Simple.SetupHooks instead.

The setup hooks datatype

data SetupHooks

Hooks into the cabal build phases.

Usage:

In your .cabal file, declare build-type: Hooks (with a cabal-version greater than or equal to 3.14),

In your .cabal file, include a custom-setup stanza which declares the dependencies of your SetupHooks module; this will usually contain a dependency on the Cabal-hooks package.

Provide a SetupHooks.hs module next to your .cabal file; it must export setupHooks :: SetupHooks.

Constructors

SetupHooks

Fields

configureHooks :: ConfigureHooks

Hooks into the configure phase.

buildHooks :: BuildHooks

Hooks into the build phase.

These hooks are relevant to any build-like phase, such as repl or haddock.

installHooks :: InstallHooks

Hooks into the copy/install phase.

noSetupHooks :: SetupHooks

Empty hooks.

Configure hooks

data ConfigureHooks

Configure-time hooks.

Order of execution:

preConfPackageHook,

configure the package,

postConfPackageHook,

preConfComponentHook,

configure the components.

Constructors

ConfigureHooks

Fields

preConfPackageHook :: Maybe PreConfPackageHook

Package-wide pre-configure hook. See PreConfPackageHook.

postConfPackageHook :: Maybe PostConfPackageHook

Package-wide post-configure hook. See PostConfPackageHook.

preConfComponentHook :: Maybe PreConfComponentHook

Per-component pre-configure hook. See PreConfComponentHook.

noConfigureHooks :: ConfigureHooks

Empty configure phase hooks.

Per-package configure hooks

data PreConfPackageInputs

Inputs to the package-wide pre-configure step.

Constructors

PreConfPackageInputs

Fields

configFlags :: ConfigFlags

localBuildConfig :: LocalBuildConfig

Warning: the ProgramDb in the withPrograms field will not contain any unconfigured programs.

compiler :: Compiler

platform :: Platform

data PreConfPackageOutputs

Outputs of the package-wide pre-configure step.

Prefer using noPreConfPackageOutputs and overriding the fields you care about, to avoid depending on implementation details of this datatype.

Constructors

PreConfPackageOutputs

Fields

buildOptions :: BuildOptions

extraConfiguredProgs :: ConfiguredProgs

noPreConfPackageOutputs :: PreConfPackageInputs -> PreConfPackageOutputs

Use this smart constructor to declare an empty set of changes by the package-wide pre-configure hook, and override the fields you care about.

Use this rather than PreConfPackageOutputs to avoid relying on internal implementation details of the latter.

type PreConfPackageHook = PreConfPackageInputs -> IO PreConfPackageOutputs

data PostConfPackageInputs

Inputs to the package-wide post-configure step.

Constructors

PostConfPackageInputs

Fields

localBuildConfig :: LocalBuildConfig

packageBuildDescr :: PackageBuildDescr

type PostConfPackageHook = PostConfPackageInputs -> IO ()

Package-wide post-configure step.

Perform side effects. Last opportunity for any package-wide logic; any subsequent hooks work per-component.

Per-component configure hooks

data PreConfComponentInputs

Inputs to the per-component pre-configure step.

Constructors

PreConfComponentInputs

Fields

localBuildConfig :: LocalBuildConfig

packageBuildDescr :: PackageBuildDescr

component :: Component

data PreConfComponentOutputs

Outputs of the per-component pre-configure step.

Prefer using noPreComponentOutputs and overriding the fields you care about, to avoid depending on implementation details of this datatype.

Constructors

PreConfComponentOutputs

Fields

componentDiff :: ComponentDiff

noPreConfComponentOutputs :: PreConfComponentInputs -> PreConfComponentOutputs

Use this smart constructor to declare an empty set of changes by a per-component pre-configure hook, and override the fields you care about.

Use this rather than PreConfComponentOutputs to avoid relying on internal implementation details of the latter.

type PreConfComponentHook = PreConfComponentInputs -> IO PreConfComponentOutputs

Per-component pre-configure step.

For each component of the package, this hook can perform side effects, and return a diff to the passed in component, e.g. to declare additional autogenerated modules.

newtype ComponentDiff

A diff to a Cabal Component, that gets combined monoidally into an existing Component.

Constructors

ComponentDiff

Fields

componentDiff :: Component

emptyComponentDiff :: ComponentName -> ComponentDiff

buildInfoComponentDiff :: ComponentName -> BuildInfo -> ComponentDiff

type LibraryDiff = Library

type ForeignLibDiff = ForeignLib

type ExecutableDiff = Executable

type TestSuiteDiff = TestSuite

type BenchmarkDiff = Benchmark

type BuildInfoDiff = BuildInfo

Build hooks

data BuildHooks

Build-time hooks.

Constructors

BuildHooks

Fields

preBuildComponentRules :: Maybe PreBuildComponentRules

Per-component fine-grained pre-build rules.

postBuildComponentHook :: Maybe PostBuildComponentHook

Per-component post-build hook.

noBuildHooks :: BuildHooks

Empty build hooks.

data BuildingWhat

What kind of build phase are we doing/hooking into?

Is this a normal build, or is it perhaps for running an interactive session or Haddock?

Constructors

BuildNormal BuildFlags
A normal build.

BuildRepl ReplFlags
Build steps for an interactive session.

BuildHaddock HaddockFlags
Build steps for generating documentation.

BuildHscolour HscolourFlags
Build steps for Hscolour.

buildingWhatVerbosity :: BuildingWhat -> Verbosity

buildingWhatWorkingDir :: BuildingWhat -> Maybe (SymbolicPath CWD ('Dir Pkg))

buildingWhatDistPref :: BuildingWhat -> SymbolicPath Pkg ('Dir Dist)

Pre-build rules

data PreBuildComponentInputs

Constructors

PreBuildComponentInputs

Fields

buildingWhat :: BuildingWhat

what kind of build phase are we hooking into?

localBuildInfo :: LocalBuildInfo

information about the package

targetInfo :: TargetInfo

information about an individual component

type PreBuildComponentRules = Rules PreBuildComponentInputs

Post-build hook

data PostBuildComponentInputs

Constructors

PostBuildComponentInputs

Fields

buildFlags :: BuildFlags

localBuildInfo :: LocalBuildInfo

targetInfo :: TargetInfo

type PostBuildComponentHook = PostBuildComponentInputs -> IO ()

Install hooks

data InstallHooks

Copy/install hooks.

Constructors

InstallHooks

Fields

installComponentHook :: Maybe InstallComponentHook

Per-component install hook.

noInstallHooks :: InstallHooks

Empty copy/install hooks.

data InstallComponentInputs

Constructors

InstallComponentInputs

Fields

copyFlags :: CopyFlags

localBuildInfo :: LocalBuildInfo

targetInfo :: TargetInfo

type InstallComponentHook = InstallComponentInputs -> IO ()

A per-component install hook, which can only perform side effects (e.g. copying files).

Internals

Per-component hook utilities

applyComponentDiffs :: Verbosity -> (Component -> IO (Maybe ComponentDiff)) -> PackageDescription -> IO PackageDescription

forComponents_ :: PackageDescription -> (Component -> IO ()) -> IO ()

Executing build rules

executeRules :: Verbosity -> LocalBuildInfo -> TargetInfo -> Map RuleId Rule -> IO ()

Run all pre-build rules.

This function should only be called internally within Cabal, as it is used to implement the (legacy) Setup.hs interface. The build tool (e.g. cabal-install or hls) should instead go through the separate hooks executable, which allows us to only rerun the out-of-date rules (instead of running all of these rules at once).

HookedBuildInfo compatibility code

hookedBuildInfoComponents :: HookedBuildInfo -> Set ComponentName

hookedBuildInfoComponentDiff_maybe :: HookedBuildInfo -> ComponentName -> Maybe (IO ComponentDiff)