Distribution.Simple.SetupHooks.Rule

Description

Internal module that defines fine-grained rules for setup hooks. Users should import SetupHooks instead.

Rules

Rule

type Rule = RuleData 'User

data RuleData (scope :: Scope)

A rule consists of:

an action to run to execute the rule,

a description of the rule inputs and outputs.

Use staticRule or dynamicRule to construct a rule, overriding specific fields, rather than directly using the Rule constructor.

Constructors

Rule
Please use the staticRule or dynamicRule smart constructors instead of this constructor, in order to avoid relying on internal implementation details.

Fields

ruleCommands :: !(RuleCmds scope)

To run this rule, which Commands should we execute?

staticDependencies :: ![Dependency]

Static dependencies of this rule.

results :: !(NonEmpty Location)

Results of this rule.

data RuleId

A unique identifier for a Rule.

Constructors

RuleId

Fields

ruleNameSpace :: !RulesNameSpace

ruleName :: !ShortText

staticRule :: Typeable arg => Command arg (IO ()) -> [Dependency] -> NonEmpty Location -> Rule

A rule with static dependencies.

Prefer using this smart constructor instead of Rule whenever possible.

dynamicRule :: (Typeable depsArg, Typeable depsRes, Typeable arg) => StaticPtr (Dict (Binary depsRes, Show depsRes, Eq depsRes)) -> Command depsArg (IO ([Dependency], depsRes)) -> Command arg (depsRes -> IO ()) -> [Dependency] -> NonEmpty Location -> Rule

A rule with dynamic dependencies.

Prefer using this smart constructor instead of Rule whenever possible.

Commands

data RuleCommands (scope :: Scope) (deps :: Scope -> Type -> Type -> Type) (ruleCmd :: Scope -> Type -> Type -> Type) where

Commands to execute a rule:

for a rule with static dependencies, a single command,

for a rule with dynamic dependencies, a command for computing dynamic dependencies, and a command for executing the rule.

Constructors

StaticRuleCommand
A rule with statically-known dependencies.

Fields

:: forall arg (deps :: Scope -> Type -> Type -> Type) (ruleCmd :: Scope -> Type -> Type -> Type) (scope :: Scope). If (scope == 'System) (arg ~ ByteString) ()

=> { staticRuleCommand :: !(ruleCmd scope arg (IO ()))

The command to execute the rule.

, staticRuleArgRep :: !(If (scope == 'System) SomeTypeRep (TypeRep arg))

A TypeRep for arg.

} -> RuleCommands scope deps ruleCmd

DynamicRuleCommands

Fields

:: forall depsArg depsRes arg (deps :: Scope -> Type -> Type -> Type) (ruleCmd :: Scope -> Type -> Type -> Type) (scope :: Scope). If (scope == 'System) (depsArg ~ ByteString, depsRes ~ ByteString, arg ~ ByteString) ()

=> { dynamicRuleInstances :: !(Static scope (Dict (Binary depsRes, Show depsRes, Eq depsRes)))

A rule with dynamic dependencies, which consists of two parts:

a dynamic dependency computation, that returns additional edges to be added to the build graph together with an additional piece of data,

the command to execute the rule itself, which receives the additional piece of data returned by the dependency computation.

, dynamicDeps :: !(deps scope depsArg depsRes)

A dynamic dependency computation. The resulting dependencies will be injected into the build graph, and the result of the computation will be passed on to the command that executes the rule.

, dynamicRuleCommand :: !(ruleCmd scope arg (depsRes -> IO ()))

The command to execute the rule. It will receive the result of the dynamic dependency computation.

, dynamicRuleTypeRep :: !(If (scope == 'System) SomeTypeRep (TypeRep (depsArg, depsRes, arg)))

A TypeRep for the triple (depsArg,depsRes,arg).

} -> RuleCommands scope deps ruleCmd

type Command = CommandData 'User

A command consists of a statically-known action together with a (possibly dynamic) argument to that action.

For example, the action can consist of running an executable (such as happy or c2hs), while the argument consists of the variable component of the command, e.g. the specific file to run happy on.

data CommandData (scope :: Scope) arg res

Internal datatype used for commands, both for the Hooks API (Command) and for the build system.

Constructors

Command

Fields

actionPtr :: !(Static scope (arg -> res))

The (statically-known) action to execute.

actionArg :: !(ScopedArgument scope arg)

The (possibly dynamic) argument to pass to the action.

cmdInstances :: !(Static scope (Dict (Binary arg, Show arg)))

Static evidence that the argument can be serialised and deserialised.

runCommand :: Command args res -> res

Run a Command.

mkCommand :: StaticPtr (Dict (Binary arg, Show arg)) -> StaticPtr (arg -> res) -> arg -> Command arg res

Construct a command.

Prefer using this smart constructor instead of Command whenever possible.

data Dict c where

A wrapper used to pass evidence of a constraint as an explicit value.

Constructors

Dict :: forall c. c => Dict c

Helpers for executing commands

type RuleCmds (scope :: Scope) = RuleCommands scope DynDepsCmd CommandData

Both the rule command and the (optional) dynamic dependency command.

type RuleDynDepsCmd (scope :: Scope) = RuleCommands scope DynDepsCmd (NoCmd :: Scope -> Type -> Type -> Type)

Only the (optional) dynamic dependency command.

type RuleExecCmd (scope :: Scope) = RuleCommands scope (DepsRes :: Scope -> Type -> Type -> Type) CommandData

The rule command together with the result of the (optional) dynamic dependency computation.

newtype DynDepsCmd (scope :: Scope) depsArg depsRes

A dynamic dependency command.

Constructors

DynDepsCmd

Fields

dynDepsCmd :: CommandData scope depsArg (IO ([Dependency], depsRes))

newtype DepsRes (scope :: Scope) (depsArg :: k) depsRes

The result of a dynamic dependency computation.

Constructors

DepsRes

Fields

depsRes :: ScopedArgument scope depsRes

ruleDepsCmd :: forall (scope :: Scope). RuleCmds scope -> RuleDynDepsCmd scope

Project out the (optional) dependency computation command, so that it can be serialised without serialising anything else.

runRuleDynDepsCmd :: RuleDynDepsCmd 'User -> Maybe (IO ([Dependency], ByteString))

Obtain the (optional) IO action that computes dynamic dependencies.

ruleExecCmd :: forall (scope :: Scope). SScope scope -> RuleCmds scope -> Maybe ByteString -> RuleExecCmd scope

Project out the command for running the rule, passing in the result of the dependency computation if there was one.

runRuleExecCmd :: RuleExecCmd 'User -> IO ()

Obtain the IO action that executes a rule.

Collections of rules

newtype Rules env

A collection of Rules.

Use the rules smart constructor instead of directly using the Rules constructor.

Rules are registered using registerRule,

Monitored files or directories are declared using addRuleMonitors; a change in these will trigger the recomputation of all rules.

The env type parameter represents an extra argument, which usually consists of information known to Cabal such as LocalBuildInfo and ComponentLocalBuildInfo.

Constructors

Rules

Fields

runRules :: env -> RulesM ()

data Dependency

A dependency of a rule.

Constructors

RuleDependency !RuleOutput
A dependency on an output of another rule.

FileDependency !Location
A direct dependency on a file at a particular location on disk.

This should not be used for files that are generated by other rules; use RuleDependency instead.

data RuleOutput

A reference to an output of another rule.

Constructors

RuleOutput

Fields

outputOfRule :: !RuleId

which rule's outputs are we referring to?

outputIndex :: !Word

which particular output of that rule?

rules

Arguments

:: StaticPtr label
unique label for this collection of rules

-> (env -> RulesM ())
the computation of rules

-> Rules env

Construct a collection of rules with a given label.

A label for the rules can be constructed using the static keyword, using the StaticPointers extension. NB: separate calls to rules should have different labels.

Example usage:

myRules :: Rules env
myRules = rules (static ()) $ \ env -> do { .. } -- use the monadic API here

noRules :: RulesM ()

An empty collection of rules.

Rule inputs/outputs

data Location where

A (fully resolved) location of a dependency or result of a rule, consisting of a base directory and of a file path relative to that base directory path.

In practice, this will be something like Location dir (moduleNameSymbolicPath mod . "hs"), where:

for a file dependency, dir is one of the Cabal search directories,

for an output, dir is a directory such as autogenComponentModulesDir or componentBuildDir.

Constructors

Location

Fields

:: forall baseDir. { locationBaseDir :: !(SymbolicPath Pkg ('Dir baseDir))

Base directory.

, locationRelPath :: !(RelativePath baseDir 'File)

File path relative to base directory (including file extension).

} -> Location

location :: Location -> SymbolicPath Pkg 'File

Get a (relative or absolute) un-interpreted path to a Location.

File/directory monitoring

data MonitorFilePath

A description of a file (or set of files) to monitor for changes.

Where file paths are relative they are relative to a common directory (e.g. project root), not necessarily the process current directory.

Constructors

MonitorFile

Fields

monitorKindFile :: !MonitorKindFile

monitorKindDir :: !MonitorKindDir

monitorPath :: !FilePath

MonitorFileGlob

Fields

monitorKindFile :: !MonitorKindFile

monitorKindDir :: !MonitorKindDir

monitorPathGlob :: !RootedGlob

data MonitorKindFile

Constructors

FileExists

FileModTime

FileHashed

FileNotExists

data MonitorKindDir

Constructors

DirExists

DirModTime

DirNotExists

Monadic API for generation of ActionId

type RulesM a = RulesT IO a

Monad for constructing rules.

newtype RulesT (m :: Type -> Type) a

Monad transformer for defining rules. Usually wraps the IO monad, allowing IO actions to be performed using liftIO.

Constructors

RulesT

Fields

runRulesT :: ReaderT RulesEnv (StateT (Map RuleId Rule) (WriterT [MonitorFilePath] m)) a

data RulesEnv

The environment within the monadic API.

Constructors

RulesEnv

Fields

rulesEnvVerbosity :: !Verbosity

rulesEnvNameSpace :: !RulesNameSpace

computeRules :: Verbosity -> env -> Rules env -> IO (Map RuleId Rule, [MonitorFilePath])

Internal function: run the monadic Rules computations in order to obtain all the Rules with their RuleIds.

Internals

data Scope

Rules are defined with rich types by the package.

The build system only has a limited view of these; most data consists of opaque ByteStrings.

The Scope data-type describes which side of this divide we are on.

Constructors

User
User space (with rich types).

System
Build-system space (manipulation of raw data).

data SScope (scope :: Scope) where

Constructors

SUser :: SScope 'User

SSystem :: SScope 'System

data family Static (scope :: Scope) :: Type -> Type

A static pointer (in user scope) or its key (in system scope).

type RuleBinary = RuleData 'System

ruleBinary :: Rule -> RuleBinary