Distribution.Simple.Test

Description

This is the entry point into testing a built package. It performs the "./setup test" action. It runs test suites designated in the package description and reports on the results.

Documentation

test

Arguments

:: Args
positional command-line arguments

-> PackageDescription
information from the .cabal file

-> LocalBuildInfo
information from the configure step

-> TestFlags
flags sent to test

-> IO ()

Perform the "./setup test" action.