Distribution.Text

Description

This defines a Text class which is a bit like the Read and Show classes. The difference is that it uses a modern pretty printer and parser system and the format is not expected to be Haskell concrete syntax but rather the external human readable representation used by Cabal.

Documentation

class Text a where

Note: this class will soon be deprecated. It's not yet, so that we are -Wall clean.

Methods

disp :: a -> Doc

disp :: Pretty a => a -> Doc

parse :: ReadP r a

parse :: Parsec a => ReadP r a

defaultStyle :: Style

The default rendering style used in Cabal for console output. It has a fixed page width and adds line breaks automatically.

display :: Text a => a -> String

Pretty-prints with the default style.

flatStyle :: Style

A style for rendering all on one line.

simpleParse :: Text a => String -> Maybe a

stdParse :: Text ver => (ver -> String -> res) -> ReadP r res