Distribution.Types.ComponentId

Documentation

data ComponentId

A ComponentId uniquely identifies the transitive source code closure of a component (i.e. libraries, executables).

For non-Backpack components, this corresponds one to one with the UnitId, which serves as the basis for install paths, linker symbols, etc.

Use mkComponentId and unComponentId to convert from/to a String.

This type is opaque since Cabal-2.0

Since: 2.0.0.2

unComponentId :: ComponentId -> String

Convert ComponentId to String

Since: 2.0.0.2

mkComponentId :: String -> ComponentId

Construct a ComponentId from a String

mkComponentId is the inverse to unComponentId

Note: No validations are performed to ensure that the resulting ComponentId is valid

Since: 2.0.0.2