Distribution.Types.CondTree

Documentation

data CondTree v c a

A CondTree is used to represent the conditional structure of a Cabal file, reflecting a syntax element subject to constraints, and then any number of sub-elements which may be enabled subject to some condition. Both a and c are usually Monoids.

To be more concrete, consider the following fragment of a Cabal file:

build-depends: base >= 4.0
if flag(extra)
    build-depends: base >= 4.2

One way to represent this is to have CondTree ConfVar [Dependency] BuildInfo. Here, condTreeData represents the actual fields which are not behind any conditional, while condTreeComponents recursively records any further fields which are behind a conditional. condTreeConstraints records the constraints (in this case, base >= 4.0) which would be applied if you use this syntax; in general, this is derived off of targetBuildInfo (perhaps a good refactoring would be to convert this into an opaque type, with a smart constructor that pre-computes the dependencies.)

Constructors

CondNode

Fields

condTreeData :: a

condTreeConstraints :: c

condTreeComponents :: [CondBranch v c a]

data CondBranch v c a

A CondBranch represents a conditional branch, e.g., if flag(foo) on some syntax a. It also has an optional false branch.

Constructors

CondBranch

Fields

condBranchCondition :: Condition v

condBranchIfTrue :: CondTree v c a

condBranchIfFalse :: Maybe (CondTree v c a)

condIfThen :: Condition v -> CondTree v c a -> CondBranch v c a

condIfThenElse :: Condition v -> CondTree v c a -> CondTree v c a -> CondBranch v c a

mapCondTree :: (a -> b) -> (c -> d) -> (Condition v -> Condition w) -> CondTree v c a -> CondTree w d b

mapTreeConstrs :: (c -> d) -> CondTree v c a -> CondTree v d a

mapTreeConds :: (Condition v -> Condition w) -> CondTree v c a -> CondTree w c a

mapTreeData :: (a -> b) -> CondTree v c a -> CondTree v c b

traverseCondTreeV :: Traversal (CondTree v c a) (CondTree w c a) v w

@Traversal@ for the variables

traverseCondBranchV :: Traversal (CondBranch v c a) (CondBranch w c a) v w

@Traversal@ for the variables

traverseCondTreeC :: Traversal (CondTree v c a) (CondTree v d a) c d

@Traversal@ for the aggregated constraints

traverseCondBranchC :: Traversal (CondBranch v c a) (CondBranch v d a) c d

@Traversal@ for the aggregated constraints

extractCondition :: Eq v => (a -> Bool) -> CondTree v c a -> Condition v

Extract the condition matched by the given predicate from a cond tree.

We use this mainly for extracting buildable conditions (see the Note above), but the function is in fact more general.

simplifyCondTree :: (Semigroup a, Semigroup d) => (v -> Either v Bool) -> CondTree v d a -> (d, a)

Flattens a CondTree using a partial flag assignment. When a condition cannot be evaluated, both branches are ignored.

ignoreConditions :: (Semigroup a, Semigroup c) => CondTree v c a -> (a, c)

Flatten a CondTree. This will resolve the CondTree by taking all possible paths into account. Note that since branches represent exclusive choices this may not result in a "sane" result.