Distribution.Types.ForeignLibOption

Documentation

data ForeignLibOption

Constructors

ForeignLibStandalone
Merge in all dependent libraries (i.e., use ghc -shared -static rather than just record the dependencies, ala ghc -shared -dynamic). This option is compulsory on Windows and unsupported on other platforms.