Distribution.Types.GivenComponent

Documentation

data GivenComponent

A GivenComponent represents a library depended on and explicitly specified by the user/client with --dependency

It enables Cabal to know which ComponentId to associate with a library

Since: Cabal-2.3.0.0

Constructors

GivenComponent

Fields

givenComponentPackage :: PackageName

givenComponentName :: LibraryName

givenComponentId :: ComponentId

data PromisedComponent

A PromisedComponent represents a promised library depended on and explicitly specified by the user/client with --promised-dependency

It enables Cabal to know which ComponentId to associate with a library

Since: Cabal-3.14.0.0

Constructors

PromisedComponent

Fields

promisedComponentPackage :: PackageId

promisedComponentName :: LibraryName

promisedComponentId :: ComponentId