Distribution.Types.PackageDescription.Lens

Documentation

data PackageDescription

This data type is the internal representation of the file pkg.cabal. It contains two kinds of information about the package: information which is needed for all packages, such as the package name and version, and information which is needed for the simple build system only, such as the compiler options and library name.

package :: Lens' PackageDescription PackageIdentifier

licenseRaw :: Lens' PackageDescription (Either License License)

licenseFiles :: Lens' PackageDescription [String]

copyright :: Lens' PackageDescription String

maintainer :: Lens' PackageDescription String

author :: Lens' PackageDescription String

stability :: Lens' PackageDescription String

testedWith :: Lens' PackageDescription [(CompilerFlavor, VersionRange)]

homepage :: Lens' PackageDescription String

pkgUrl :: Lens' PackageDescription String

bugReports :: Lens' PackageDescription String

sourceRepos :: Lens' PackageDescription [SourceRepo]

synopsis :: Lens' PackageDescription String

description :: Lens' PackageDescription String

category :: Lens' PackageDescription String

customFieldsPD :: Lens' PackageDescription [(String, String)]

buildDepends :: Lens' PackageDescription [Dependency]

specVersionRaw :: Lens' PackageDescription (Either Version VersionRange)

buildTypeRaw :: Lens' PackageDescription (Maybe BuildType)

setupBuildInfo :: Lens' PackageDescription (Maybe SetupBuildInfo)

library :: Lens' PackageDescription (Maybe Library)

subLibraries :: Lens' PackageDescription [Library]

executables :: Lens' PackageDescription [Executable]

foreignLibs :: Lens' PackageDescription [ForeignLib]

testSuites :: Lens' PackageDescription [TestSuite]

benchmarks :: Lens' PackageDescription [Benchmark]

dataFiles :: Lens' PackageDescription [FilePath]

dataDir :: Lens' PackageDescription FilePath

extraSrcFiles :: Lens' PackageDescription [String]

extraTmpFiles :: Lens' PackageDescription [String]

extraDocFiles :: Lens' PackageDescription [String]