Distribution.Types.ParStrat

Documentation

data ParStratX sem

How to control parallelism, e.g. a fixed number of jobs or by using a system semaphore.

Constructors

NumJobs (Maybe Int)
Compile in parallel with the given number of jobs (`-jN` or `-j`).

UseSem sem
`--semaphore`: use a system semaphore to control parallelism.

Serial
No parallelism (neither `-jN` nor `--semaphore`, but could be `-j1`).

type ParStrat = ParStratX String

Used by Cabal to indicate that we want to use this specific semaphore (created by cabal-install)

type ParStratInstall = ParStratX Int

Used by cabal-install to say we want to create a semaphore with N slots.

isParallelBuild :: ParStratX n -> Bool

Determine if the parallelism strategy enables parallel builds.