Distribution.Types.TestSuite

Documentation

data TestSuite

A "test-suite" stanza in a cabal file.

Constructors

TestSuite

Fields

testName :: UnqualComponentName

testInterface :: TestSuiteInterface

testBuildInfo :: BuildInfo

emptyTestSuite :: TestSuite

testType :: TestSuite -> TestType

testModules :: TestSuite -> [ModuleName]

Get all the module names from a test suite.

testModulesAutogen :: TestSuite -> [ModuleName]

Get all the auto generated module names from a test suite. This are a subset of testModules.