Distribution.Utils.IOData

Description

Since: Cabal-2.2.0

IOData & IODataMode type

data IOData

Represents either textual or binary data passed via I/O functions which support binary/text mode

Since: Cabal-2.2

Constructors

IODataText String
How Text gets encoded is usually locale-dependent.

IODataBinary ByteString
Raw binary which gets read/written in binary mode.

data IODataMode mode where

Since: Cabal-3.2

Constructors

IODataModeText :: IODataMode String

IODataModeBinary :: IODataMode ByteString

class NFData mode => KnownIODataMode mode where

Since: Cabal-2.2

Methods

hGetIODataContents :: Handle -> IO mode

IOData Wrapper for hGetContents

Note: This operation uses lazy I/O. Use NFData to force all data to be read and consequently the internal file handle to be closed.

toIOData :: mode -> IOData

iodataMode :: IODataMode mode

withIOData :: IOData -> (forall mode. IODataMode mode -> mode -> r) -> r

null :: IOData -> Bool

Test whether IOData is empty

hPutContents :: Handle -> IOData -> IO ()

IOData Wrapper for hPutStr and hClose

This is the dual operation ot hGetIODataContents, and consequently the handle is closed with hClose.

Note: this performs lazy-IO.

Since: Cabal-2.2