Distribution.Backpack.ComponentsGraph

Description

See https://github.com/ezyang/ghc-proposals/blob/backpack/proposals/0000-backpack.rst

Documentation

type ComponentsGraph = [(Component, [ComponentName])]

A components graph is a source level graph tracking the dependencies between components in a package.

dispComponentsGraph :: ComponentsGraph -> Doc

Pretty-print a ComponentsGraph.

toComponentsGraph :: ComponentRequestedSpec -> PackageDescription -> Either [ComponentName] ComponentsGraph

Given the package description and a PackageDescription (used to determine if a package name is internal or not), create a graph of dependencies between the components. This is NOT necessarily the build order (although it is in the absence of Backpack.)

componentCycleMsg :: [ComponentName] -> Doc

Error message when there is a cycle; takes the SCC of components.