Distribution.Backpack.ModuleShape

Description

See https://github.com/ezyang/ghc-proposals/blob/backpack/proposals/0000-backpack.rst

Module shapes

data ModuleShape

A ModuleShape describes the provisions and requirements of a library. We can extract a ModuleShape from an InstalledPackageInfo.

Constructors

ModuleShape

Fields

modShapeProvides :: OpenModuleSubst

modShapeRequires :: Set ModuleName

emptyModuleShape :: ModuleShape

The default module shape, with no provisions and no requirements.

shapeInstalledPackage :: InstalledPackageInfo -> ModuleShape