Distribution.Compat.Stack

Documentation

type WithCallStack a = (?callStack :: CallStack) => a

data CallStack :: *

CallStacks are an alternate method of obtaining the call stack at a given point in the program.

When an implicit-parameter of type CallStack occurs in a program, GHC will solve it with the current location. If another CallStack implicit-parameter is in-scope (e.g. as a function argument), the new location will be appended to the one in-scope, creating an explicit call-stack. For example,

myerror :: (?loc :: CallStack) => String -> a
myerror msg = error (msg ++ "n" ++ showCallStack ?loc)

ghci> myerror "die" *** Exception: die ?loc, called at MyError.hs:7:51 in main:MyError myerror, called at interactive:2:1 in interactive:Ghci1

CallStacks do not interact with the RTS and do not require compilation with -prof. On the other hand, as they are built up explicitly using implicit-parameters, they will generally not contain as much information as the simulated call-stacks maintained by the RTS.

The CallStack type is abstract, but it can be converted into a [(String, SrcLoc)] via getCallStack. The String is the name of function that was called, the SrcLoc is the call-site. The list is ordered with the most recently called function at the head.

Since: 4.9.0.0

withFrozenCallStack :: WithCallStack (a -> a)

withLexicalCallStack :: (a -> WithCallStack (IO b)) -> WithCallStack (a -> IO b)

callStack :: (?callStack :: CallStack) => CallStack

prettyCallStack :: CallStack -> String

parentSrcLocPrefix :: WithCallStack String

Give the *parent* of the person who invoked this; so it's most suitable for being called from a utility function. You probably want to call this using withFrozenCallStack; otherwise it's not very useful. We didn't implement this for base-4.8.1 because we cannot rely on freezing to have taken place.