Distribution.Compat.Stack

Documentation

type WithCallStack a = HasCallStack => a

data CallStack :: *

CallStacks are a lightweight method of obtaining a partial call-stack at any point in the program.

A function can request its call-site with the HasCallStack constraint. For example, we can define

errorWithCallStack :: HasCallStack => String -> a

as a variant of error that will get its call-site. We can access the call-stack inside errorWithCallStack with callStack.

errorWithCallStack :: HasCallStack => String -> a
errorWithCallStack msg = error (msg ++ "n" ++ prettyCallStack callStack)

Thus, if we call errorWithCallStack we will get a formatted call-stack alongside our error message.

>>> errorWithCallStack "die"
*** Exception: die
CallStack (from HasCallStack):
  errorWithCallStack, called at <interactive>:2:1 in interactive:Ghci1

GHC solves HasCallStack constraints in three steps:

If there is a CallStack in scope -- i.e. the enclosing function has a HasCallStack constraint -- GHC will append the new call-site to the existing CallStack.

If there is no CallStack in scope -- e.g. in the GHCi session above -- and the enclosing definition does not have an explicit type signature, GHC will infer a HasCallStack constraint for the enclosing definition (subject to the monomorphism restriction).

If there is no CallStack in scope and the enclosing definition has an explicit type signature, GHC will solve the HasCallStack constraint for the singleton CallStack containing just the current call-site.

CallStacks do not interact with the RTS and do not require compilation with -prof. On the other hand, as they are built up explicitly via the HasCallStack constraints, they will generally not contain as much information as the simulated call-stacks maintained by the RTS.

A CallStack is a [(String, SrcLoc)]. The String is the name of function that was called, the SrcLoc is the call-site. The list is ordered with the most recently called function at the head.

NOTE: The intrepid user may notice that HasCallStack is just an alias for an implicit parameter ?callStack :: CallStack. This is an implementation detail and should not be considered part of the CallStack API, we may decide to change the implementation in the future.

Since: 4.8.1.0

annotateCallStackIO :: WithCallStack (IO a -> IO a)

This function is for when you *really* want to add a call stack to raised IO, but you don't have a Verbosity so you can't use annotateIO. If you have a Verbosity, please use that function instead.

withFrozenCallStack :: HasCallStack => (HasCallStack -> a) -> a

Perform some computation without adding new entries to the CallStack.

Since: 4.9.0.0

withLexicalCallStack :: (a -> WithCallStack (IO b)) -> WithCallStack (a -> IO b)

callStack :: HasCallStack -> CallStack

Return the current CallStack.

Does *not* include the call-site of callStack.

Since: 4.9.0.0

prettyCallStack :: CallStack -> String

Pretty print a CallStack.

Since: 4.9.0.0

parentSrcLocPrefix :: WithCallStack String

Give the *parent* of the person who invoked this; so it's most suitable for being called from a utility function. You probably want to call this using withFrozenCallStack; otherwise it's not very useful. We didn't implement this for base-4.8.1 because we cannot rely on freezing to have taken place.