Distribution.Compiler

Description

This has an enumeration of the various compilers that Cabal knows about. It also specifies the default compiler. Sadly you'll often see code that does case analysis on this compiler flavour enumeration like:

case compilerFlavor comp of
  GHC -> GHC.getInstalledPackages verbosity packageDb progconf
  JHC -> JHC.getInstalledPackages verbosity packageDb progconf

Obviously it would be better to use the proper Compiler abstraction because that would keep all the compiler-specific code together. Unfortunately we cannot make this change yet without breaking the UserHooks api, which would break all custom Setup.hs files, so for the moment we just have to live with this deficiency. If you're interested, see ticket #57.

Compiler flavor

data CompilerFlavor

Constructors

GHC

GHCJS

NHC

YHC

Hugs

HBC

Helium

JHC

LHC

UHC

HaskellSuite String

OtherCompiler String

buildCompilerId :: CompilerId

buildCompilerFlavor :: CompilerFlavor

defaultCompilerFlavor :: Maybe CompilerFlavor

The default compiler flavour to pick when compiling stuff. This defaults to the compiler used to build the Cabal lib.

However if it's not a recognised compiler then it's Nothing and the user will have to specify which compiler they want.

parseCompilerFlavorCompat :: ReadP r CompilerFlavor

Like classifyCompilerFlavor but compatible with the old ReadS parser.

It is compatible in the sense that it accepts only the same strings, eg GHC but not "ghc". However other strings get mapped to OtherCompiler. The point of this is that we do not allow extra valid values that would upset older Cabal versions that had a stricter parser however we cope with new values more gracefully so that we'll be able to introduce new value in future without breaking things so much.

Compiler id

data CompilerId

Constructors

CompilerId CompilerFlavor Version

Compiler info

data CompilerInfo

Compiler information used for resolving configurations. Some fields can be set to Nothing to indicate that the information is unknown.

Constructors

CompilerInfo

Fields

compilerInfoId :: CompilerId

Compiler flavour and version.

compilerInfoAbiTag :: AbiTag

Tag for distinguishing incompatible ABI's on the same architecture/os.

compilerInfoCompat :: Maybe [CompilerId]

Other implementations that this compiler claims to be compatible with, if known.

compilerInfoLanguages :: Maybe [Language]

Supported language standards, if known.

compilerInfoExtensions :: Maybe [Extension]

Supported extensions, if known.

unknownCompilerInfo :: CompilerId -> AbiTag -> CompilerInfo

Make a CompilerInfo of which only the known information is its CompilerId, its AbiTag and that it does not claim to be compatible with other compiler id's.

data AbiTag

Constructors

NoAbiTag

AbiTag String

abiTagString :: AbiTag -> String