Distribution.InstalledPackageInfo

Description

This is the information about an installed package that is communicated to the ghc-pkg program in order to register a package. ghc-pkg now consumes this package format (as of version 6.4). This is specific to GHC at the moment.

The .cabal file format is for describing a package that is not yet installed. It has a lot of flexibility, like conditionals and dependency ranges. As such, that format is not at all suitable for describing a package that has already been built and installed. By the time we get to that stage, we have resolved all conditionals and resolved dependency version constraints to exact versions of dependent packages. So, this module defines the InstalledPackageInfo data structure that contains all the info we keep about an installed package. There is a parser and pretty printer. The textual format is rather simpler than the .cabal format: there are no sections, for example.

Documentation

data InstalledPackageInfo_ m

Constructors

InstalledPackageInfo

Fields

installedPackageId :: InstalledPackageId

sourcePackageId :: PackageId

license :: License

copyright :: String

maintainer :: String

author :: String

stability :: String

homepage :: String

pkgUrl :: String

synopsis :: String

description :: String

category :: String

exposed :: Bool

exposedModules :: [m]

hiddenModules :: [m]

trusted :: Bool

importDirs :: [FilePath]

libraryDirs :: [FilePath]

hsLibraries :: [String]

extraLibraries :: [String]

extraGHCiLibraries :: [String]

includeDirs :: [FilePath]

includes :: [String]

depends :: [InstalledPackageId]

hugsOptions :: [String]

ccOptions :: [String]

ldOptions :: [String]

frameworkDirs :: [FilePath]

frameworks :: [String]

haddockInterfaces :: [FilePath]

haddockHTMLs :: [FilePath]

type InstalledPackageInfo = InstalledPackageInfo_ ModuleName

data ParseResult a

Constructors

ParseFailed PError

ParseOk [PWarning] a

data PError

Constructors

AmbigousParse String LineNo

NoParse String LineNo

TabsError LineNo

FromString String (Maybe LineNo)

data PWarning

emptyInstalledPackageInfo :: InstalledPackageInfo_ m

parseInstalledPackageInfo :: String -> ParseResult InstalledPackageInfo

showInstalledPackageInfo :: InstalledPackageInfo -> String

showInstalledPackageInfoField :: String -> Maybe (InstalledPackageInfo -> String)

fieldsInstalledPackageInfo :: [FieldDescr InstalledPackageInfo]