Distribution.Package

Description

Defines a package identifier along with a parser and pretty printer for it. PackageIdentifiers consist of a name and an exact version. It also defines a Dependency data type. A dependency is a package name and a version range, like "foo >= 1.2 && < 2".

Package ids

data PackageName

unPackageName :: PackageName -> String

Convert PackageName to String

mkPackageName :: String -> PackageName

data PackageIdentifier

The name and version of a package.

Constructors

PackageIdentifier

Fields

pkgName :: PackageName

The name of this package, eg. foo

pkgVersion :: Version

the version of this package, eg 1.2

type PackageId = PackageIdentifier

Type alias so we can use the shorter name PackageId.

data PkgconfigName

unPkgconfigName :: PkgconfigName -> String

mkPkgconfigName :: String -> PkgconfigName

Package keys/installed package IDs (used for linker symbols)

data ComponentId

unComponentId :: ComponentId -> String

mkComponentId :: String -> ComponentId

data UnitId

A unit identifier identifies a (possibly instantiated) package/component that can be installed the installed package database. There are several types of components that can be installed:

A traditional library with no holes, so that unitIdHash is Nothing. In the absence of Backpack, UnitId is the same as a ComponentId.

An indefinite, Backpack library with holes. In this case, unitIdHash is still Nothing, but in the install, there are only interfaces, no compiled objects.

An instantiated Backpack library with all the holes filled in. unitIdHash is a Just a hash of the instantiating mapping.

A unit is a component plus the additional information on how the holes are filled in. Thus there is a one to many relationship: for a particular component there are many different ways of filling in the holes, and each different combination is a unit (and has a separate UnitId).

UnitId is distinct from OpenUnitId, in that it is always installed, whereas OpenUnitId are intermediate unit identities that arise during mixin linking, and don't necessarily correspond to any actually installed unit. Since the mapping is not actually recorded in a UnitId, you can't actually substitute over them (but you can substitute over OpenUnitId). See also Distribution.Backpack.FullUnitId for a mechanism for expanding an instantiated UnitId to retrieve its mapping.

Backwards compatibility note: if you need to get the string representation of a UnitId to pass, e.g., as a -package-id flag, use the display function, which will work on all versions of Cabal.

unUnitId :: UnitId -> String

If you need backwards compatibility, consider using display instead, which is supported by all versions of Cabal.

mkUnitId :: String -> UnitId

data DefUnitId

A UnitId for a definite package. The DefUnitId invariant says that a UnitId identified this way is definite; i.e., it has no unfilled holes.

unsafeMkDefUnitId :: UnitId -> DefUnitId

Unsafely create a DefUnitId from a UnitId. Your responsibility is to ensure that the DefUnitId invariant holds.

unDefUnitId :: DefUnitId -> UnitId

newSimpleUnitId :: ComponentId -> UnitId

Create a unit identity with no associated hash directly from a ComponentId.

mkLegacyUnitId :: PackageId -> UnitId

Make an old-style UnitId from a package identifier

getHSLibraryName :: UnitId -> String

Returns library name prefixed with HS, suitable for filenames

type InstalledPackageId = UnitId

Deprecated: Use UnitId instead

Modules

data Module

A module identity uniquely identifies a Haskell module by qualifying a ModuleName with the UnitId which defined it. This type distinguishes between two packages which provide a module with the same name, or a module from the same package compiled with different dependencies. There are a few cases where Cabal needs to know about module identities, e.g., when writing out reexported modules in the InstalledPackageInfo.

Constructors

Module DefUnitId ModuleName

ABI hash

data AbiHash

unAbiHash :: AbiHash -> String

mkAbiHash :: String -> AbiHash

Package classes

class Package pkg where

Class of things that have a PackageIdentifier

Types in this class are all notions of a package. This allows us to have different types for the different phases that packages go though, from simple name/id, package description, configured or installed packages.

Not all kinds of packages can be uniquely identified by a PackageIdentifier. In particular, installed packages cannot, there may be many installed instances of the same source package.

Methods

packageId :: pkg -> PackageIdentifier

packageName :: Package pkg => pkg -> PackageName

packageVersion :: Package pkg => pkg -> Version

class Package pkg => HasUnitId pkg where

Packages that have an installed unit ID

Methods

installedUnitId :: pkg -> UnitId

installedPackageId :: HasUnitId pkg => pkg -> UnitId

Deprecated: Use installedUnitId instead

Compatibility wrapper for Cabal pre-1.24.

class HasUnitId pkg => PackageInstalled pkg where

Class of installed packages.

The primary data type which is an instance of this package is InstalledPackageInfo, but when we are doing install plans in Cabal install we may have other, installed package-like things which contain more metadata. Installed packages have exact dependencies installedDepends.

Methods

installedDepends :: pkg -> [UnitId]