Distribution.PackageDescription

Description

Backwards compatibility reexport of everything you need to know about .cabal files.

Package descriptions

data PackageDescription

This data type is the internal representation of the file pkg.cabal. It contains two kinds of information about the package: information which is needed for all packages, such as the package name and version, and information which is needed for the simple build system only, such as the compiler options and library name.

Constructors

PackageDescription

Fields

package :: PackageIdentifier

license :: License

licenseFiles :: [FilePath]

copyright :: String

maintainer :: String

author :: String

stability :: String

testedWith :: [(CompilerFlavor, VersionRange)]

homepage :: String

pkgUrl :: String

bugReports :: String

sourceRepos :: [SourceRepo]

synopsis :: String

A one-line summary of this package

description :: String

A more verbose description of this package

category :: String

customFieldsPD :: [(String, String)]

Custom fields starting with x-, stored in a simple assoc-list.

buildDepends :: [Dependency]

YOU PROBABLY DON'T WANT TO USE THIS FIELD. This field is special! Depending on how far along processing the PackageDescription we are, the contents of this field are either nonsense, or the collected dependencies of *all* the components in this package. buildDepends is initialized by finalizePD and flattenPackageDescription; prior to that, dependency info is stored in the CondTree built around a GenericPackageDescription. When this resolution is done, dependency info is written to the inner BuildInfo and this field. This is all horrible, and #2066 tracks progress to get rid of this field.

specVersionRaw :: Either Version VersionRange

The version of the Cabal spec that this package description uses. For historical reasons this is specified with a version range but only ranges of the form >= v make sense. We are in the process of transitioning to specifying just a single version, not a range.

buildType :: Maybe BuildType

setupBuildInfo :: Maybe SetupBuildInfo

library :: Maybe Library

subLibraries :: [Library]

executables :: [Executable]

foreignLibs :: [ForeignLib]

testSuites :: [TestSuite]

benchmarks :: [Benchmark]

dataFiles :: [FilePath]

dataDir :: FilePath

extraSrcFiles :: [FilePath]

extraTmpFiles :: [FilePath]

extraDocFiles :: [FilePath]

emptyPackageDescription :: PackageDescription

specVersion :: PackageDescription -> Version

The version of the Cabal spec that this package should be interpreted against.

Historically we used a version range but we are switching to using a single version. Currently we accept either. This function converts into a single version by ignoring upper bounds in the version range.

descCabalVersion :: PackageDescription -> VersionRange

Deprecated: Use specVersion instead

The range of versions of the Cabal tools that this package is intended to work with.

This function is deprecated and should not be used for new purposes, only to support old packages that rely on the old interpretation.

data BuildType

The type of build system used by this package.

Constructors

Simple
calls Distribution.Simple.defaultMain

Configure
calls Distribution.Simple.defaultMainWithHooks defaultUserHooks, which invokes configure to generate additional build information used by later phases.

Make
calls Distribution.Make.defaultMain

Custom
uses user-supplied Setup.hs or Setup.lhs (default)

UnknownBuildType String
a package that uses an unknown build type cannot actually be built. Doing it this way rather than just giving a parse error means we get better error messages and allows you to inspect the rest of the package description.

knownBuildTypes :: [BuildType]

allLibraries :: PackageDescription -> [Library]

Renaming (syntactic)

data ModuleRenaming

Renaming applied to the modules provided by a package. The boolean indicates whether or not to also include all of the original names of modules. Thus, ModuleRenaming False [] is "don't expose any modules, and ModuleRenaming True [(Data.Bool, Bool)] is, "expose all modules, but also expose Data.Bool as Bool". If a renaming is omitted you get the DefaultRenaming.

(NB: This is a list not a map so that we can preserve order.)

Constructors

ModuleRenaming [(ModuleName, ModuleName)]
A module renaming/thinning; e.g., (A as B, C as C) brings B and C into scope.

DefaultRenaming
The default renaming, bringing all exported modules into scope.

HidingRenaming [ModuleName]
Hiding renaming, e.g., hiding (A, B), bringing all exported modules into scope except the hidden ones.

defaultRenaming :: ModuleRenaming

The default renaming, if something is specified in build-depends only.

Libraries

data Library

Constructors

Library

Fields

libName :: Maybe UnqualComponentName

exposedModules :: [ModuleName]

reexportedModules :: [ModuleReexport]

signatures :: [ModuleName]

What sigs need implementations?

libExposed :: Bool

Is the lib to be exposed by default?

libBuildInfo :: BuildInfo

data ModuleReexport

Constructors

ModuleReexport

Fields

moduleReexportOriginalPackage :: Maybe PackageName

moduleReexportOriginalName :: ModuleName

moduleReexportName :: ModuleName

emptyLibrary :: Library

withLib :: PackageDescription -> (Library -> IO ()) -> IO ()

If the package description has a buildable library section, call the given function with the library build info as argument. You probably want withLibLBI if you have a LocalBuildInfo, see the note in Distribution.Types.ComponentRequestedSpec for more information.

hasPublicLib :: PackageDescription -> Bool

Does this package have a buildable PUBLIC library?

hasLibs :: PackageDescription -> Bool

Does this package have any libraries?

explicitLibModules :: Library -> [ModuleName]

Get all the module names from the library (exposed and internal modules) which are explicitly listed in the package description which would need to be compiled. (This does not include reexports, which do not need to be compiled.) This may not include all modules for which GHC generated interface files (i.e., implicit modules.)

libModulesAutogen :: Library -> [ModuleName]

Get all the auto generated module names from the library, exposed or not. This are a subset of libModules.

libModules :: Library -> [ModuleName]

Deprecated: If you want all modules that are built with a library, use allLibModules. Otherwise, use explicitLibModules for ONLY the modules explicitly mentioned in the package description.

Backwards-compatibility shim for explicitLibModules. In most cases, you actually want allLibModules, which returns all modules that will actually be compiled, as opposed to those which are explicitly listed in the package description (explicitLibModules); unfortunately, the type signature for allLibModules is incompatible since we need a ComponentLocalBuildInfo.

Executables

data Executable

Constructors

Executable

Fields

exeName :: UnqualComponentName

modulePath :: FilePath

buildInfo :: BuildInfo

emptyExecutable :: Executable

withExe :: PackageDescription -> (Executable -> IO ()) -> IO ()

Perform the action on each buildable Executable in the package description. You probably want withExeLBI if you have a LocalBuildInfo, see the note in Distribution.Types.ComponentRequestedSpec for more information.

hasExes :: PackageDescription -> Bool

does this package have any executables?

exeModules :: Executable -> [ModuleName]

Get all the module names from an exe

exeModulesAutogen :: Executable -> [ModuleName]

Get all the auto generated module names from an exe This are a subset of exeModules.

Tests

data TestSuite

A "test-suite" stanza in a cabal file.

Constructors

TestSuite

Fields

testName :: UnqualComponentName

testInterface :: TestSuiteInterface

testBuildInfo :: BuildInfo

data TestSuiteInterface

The test suite interfaces that are currently defined. Each test suite must specify which interface it supports.

More interfaces may be defined in future, either new revisions or totally new interfaces.

Constructors

TestSuiteExeV10 Version FilePath
Test interface "exitcode-stdio-1.0". The test-suite takes the form of an executable. It returns a zero exit code for success, non-zero for failure. The stdout and stderr channels may be logged. It takes no command line parameters and nothing on stdin.

TestSuiteLibV09 Version ModuleName
Test interface "detailed-0.9". The test-suite takes the form of a library containing a designated module that exports "tests :: [Test]".

TestSuiteUnsupported TestType
A test suite that does not conform to one of the above interfaces for the given reason (e.g. unknown test type).

data TestType

The "test-type" field in the test suite stanza.

Constructors

TestTypeExe Version
"type: exitcode-stdio-x.y"

TestTypeLib Version
"type: detailed-x.y"

TestTypeUnknown String Version
Some unknown test type e.g. "type: foo"

testType :: TestSuite -> TestType

knownTestTypes :: [TestType]

emptyTestSuite :: TestSuite

hasTests :: PackageDescription -> Bool

Does this package have any test suites?

withTest :: PackageDescription -> (TestSuite -> IO ()) -> IO ()

Perform an action on each buildable TestSuite in a package. You probably want withTestLBI if you have a LocalBuildInfo, see the note in Distribution.Types.ComponentRequestedSpec for more information.

testModules :: TestSuite -> [ModuleName]

Get all the module names from a test suite.

testModulesAutogen :: TestSuite -> [ModuleName]

Get all the auto generated module names from a test suite. This are a subset of testModules.

Benchmarks

data Benchmark

A "benchmark" stanza in a cabal file.

Constructors

Benchmark

Fields

benchmarkName :: UnqualComponentName

benchmarkInterface :: BenchmarkInterface

benchmarkBuildInfo :: BuildInfo

data BenchmarkInterface

The benchmark interfaces that are currently defined. Each benchmark must specify which interface it supports.

More interfaces may be defined in future, either new revisions or totally new interfaces.

Constructors

BenchmarkExeV10 Version FilePath
Benchmark interface "exitcode-stdio-1.0". The benchmark takes the form of an executable. It returns a zero exit code for success, non-zero for failure. The stdout and stderr channels may be logged. It takes no command line parameters and nothing on stdin.

BenchmarkUnsupported BenchmarkType
A benchmark that does not conform to one of the above interfaces for the given reason (e.g. unknown benchmark type).

data BenchmarkType

The "benchmark-type" field in the benchmark stanza.

Constructors

BenchmarkTypeExe Version
"type: exitcode-stdio-x.y"

BenchmarkTypeUnknown String Version
Some unknown benchmark type e.g. "type: foo"

benchmarkType :: Benchmark -> BenchmarkType

knownBenchmarkTypes :: [BenchmarkType]

emptyBenchmark :: Benchmark

hasBenchmarks :: PackageDescription -> Bool

Does this package have any benchmarks?

withBenchmark :: PackageDescription -> (Benchmark -> IO ()) -> IO ()

Perform an action on each buildable Benchmark in a package. You probably want withBenchLBI if you have a LocalBuildInfo, see the note in Distribution.Types.ComponentRequestedSpec for more information.

benchmarkModules :: Benchmark -> [ModuleName]

Get all the module names from a benchmark.

benchmarkModulesAutogen :: Benchmark -> [ModuleName]

Get all the auto generated module names from a benchmark. This are a subset of benchmarkModules.

Build information

data BuildInfo

Constructors

BuildInfo

Fields

buildable :: Bool

component is buildable here | Tools needed to build this bit.

This is a legacy field that "build-tool-depends" larely supersedes.

Unless use are very sure what you are doing, use the functions in BuildToolDepends rather than accessing this field directly.

buildTools :: [LegacyExeDependency]

toolDepends :: [ExeDependency]

Haskell tools needed to build this bit

This field is better than "build-tools" because it allows one to precisely specify an executable in a package.

Unless use are very sure what you are doing, use the functions in BuildToolDepends rather than accessing this field directly.

cppOptions :: [String]

options for pre-processing Haskell code

ccOptions :: [String]

options for C compiler

ldOptions :: [String]

options for linker

pkgconfigDepends :: [PkgconfigDependency]

pkg-config packages that are used

frameworks :: [String]

support frameworks for Mac OS X

extraFrameworkDirs :: [String]

extra locations to find frameworks.

cSources :: [FilePath]

jsSources :: [FilePath]

hsSourceDirs :: [FilePath]

where to look for the Haskell module hierarchy

otherModules :: [ModuleName]

non-exposed or non-main modules

autogenModules :: [ModuleName]

not present on sdist, Paths_* or user-generated with a custom Setup.hs

defaultLanguage :: Maybe Language

language used when not explicitly specified

otherLanguages :: [Language]

other languages used within the package

defaultExtensions :: [Extension]

language extensions used by all modules

otherExtensions :: [Extension]

other language extensions used within the package

oldExtensions :: [Extension]

the old extensions field, treated same as defaultExtensions

extraLibs :: [String]

what libraries to link with when compiling a program that uses your package

extraGHCiLibs :: [String]

if present, overrides extraLibs when package is loaded with GHCi.

extraLibDirs :: [String]

includeDirs :: [FilePath]

directories to find .h files

includes :: [FilePath]

The .h files to be found in includeDirs

installIncludes :: [FilePath]

.h files to install with the package

options :: [(CompilerFlavor, [String])]

profOptions :: [(CompilerFlavor, [String])]

sharedOptions :: [(CompilerFlavor, [String])]

customFieldsBI :: [(String, String)]

Custom fields starting with x-, stored in a simple assoc-list.

targetBuildDepends :: [Dependency]

Dependencies specific to a library or executable target

mixins :: [Mixin]

emptyBuildInfo :: BuildInfo

allBuildInfo :: PackageDescription -> [BuildInfo]

The BuildInfo for the library (if there is one and it's buildable), and all buildable executables, test suites and benchmarks. Useful for gathering dependencies.

allLanguages :: BuildInfo -> [Language]

The Languages used by this component

allExtensions :: BuildInfo -> [Extension]

The Extensions that are used somewhere by this component

usedExtensions :: BuildInfo -> [Extension]

The Extensions that are used by all modules in this component

hcOptions :: CompilerFlavor -> BuildInfo -> [String]

Select options for a particular Haskell compiler.

hcProfOptions :: CompilerFlavor -> BuildInfo -> [String]

hcSharedOptions :: CompilerFlavor -> BuildInfo -> [String]

Supplementary build information

data ComponentName

Constructors

CLibName

CSubLibName UnqualComponentName

CFLibName UnqualComponentName

CExeName UnqualComponentName

CTestName UnqualComponentName

CBenchName UnqualComponentName

defaultLibName :: ComponentName

type HookedBuildInfo = (Maybe BuildInfo, [(UnqualComponentName, BuildInfo)])

HookedBuildInfo is mechanism that hooks can use to override the BuildInfos inside packages. One example use-case (which is used in core libraries today) is as a way of passing flags which are computed by a configure script into Cabal. In this case, the autoconf build type adds hooks to read in a textual HookedBuildInfo format prior to doing any operations.

Quite honestly, this mechanism is a massive hack since we shouldn't be editing the PackageDescription data structure (it's easy to assume that this data structure shouldn't change and run into bugs, see for example 1c20a6328579af9e37677d507e2e9836ef70ab9d). But it's a bit convenient, because there isn't another data structure that allows adding extra BuildInfo style things.

In any case, a lot of care has to be taken to make sure the HookedBuildInfo is applied to the PackageDescription. In general this process occurs in Distribution.Simple, which is responsible for orchestrating the hooks mechanism. The general strategy:

We run the pre-hook, which produces a HookedBuildInfo (e.g., in the Autoconf case, it reads it out from a file).

We sanity-check the hooked build info with sanityCheckHookedBuildInfo.

We update our PackageDescription (either freshly read or cached from LocalBuildInfo) with updatePackageDescription.

In principle, we are also supposed to update the copy of the PackageDescription stored in LocalBuildInfo at localPkgDescr. Unfortunately, in practice, there are lots of Custom setup scripts which fail to update localPkgDescr so you really shouldn't rely on it. It's not DEPRECATED because there are legitimate uses for it, but... yeah. Sharp knife. See https://github.com/haskell/cabal/issues/3606 for more information on the issue.

It is not well-specified whether or not a HookedBuildInfo applied at configure time is persistent to the LocalBuildInfo. The fact that HookedBuildInfo is passed to confHook MIGHT SUGGEST that the HookedBuildInfo is applied at this time, but actually since 9317b67e6122ab14e53f81b573bd0ecb388eca5a it has been ONLY used to create a modified package description that we check for problems: it is never actually saved to the LBI. Since HookedBuildInfo is applied monoidally to the existing build infos (and it is not an idempotent monoid), it could break things to save it, since we are obligated to apply any new HookedBuildInfo and then we'd get the effect twice. But this does mean we have to re-apply it every time. Hey, it's more flexibility.

emptyHookedBuildInfo :: HookedBuildInfo

updatePackageDescription :: HookedBuildInfo -> PackageDescription -> PackageDescription

package configuration

data GenericPackageDescription

Constructors

GenericPackageDescription

Fields

packageDescription :: PackageDescription

genPackageFlags :: [Flag]

condLibrary :: Maybe (CondTree ConfVar [Dependency] Library)

condSubLibraries :: [(UnqualComponentName, CondTree ConfVar [Dependency] Library)]

condForeignLibs :: [(UnqualComponentName, CondTree ConfVar [Dependency] ForeignLib)]

condExecutables :: [(UnqualComponentName, CondTree ConfVar [Dependency] Executable)]

condTestSuites :: [(UnqualComponentName, CondTree ConfVar [Dependency] TestSuite)]

condBenchmarks :: [(UnqualComponentName, CondTree ConfVar [Dependency] Benchmark)]

data Flag

A flag can represent a feature to be included, or a way of linking a target against its dependencies, or in fact whatever you can think of.

Constructors

MkFlag

Fields

flagName :: FlagName

flagDescription :: String

flagDefault :: Bool

flagManual :: Bool

emptyFlag :: FlagName -> Flag

A Flag initialized with default parameters.

data FlagName

A FlagName is the name of a user-defined configuration flag

Use mkFlagName and unFlagName to convert from/to a String.

This type is opaque since Cabal-2.0

Since: 2.0

mkFlagName :: String -> FlagName

Construct a FlagName from a String

mkFlagName is the inverse to unFlagName

Note: No validations are performed to ensure that the resulting FlagName is valid

Since: 2.0

unFlagName :: FlagName -> String

Convert FlagName to String

Since: 2.0

type FlagAssignment = [(FlagName, Bool)]

A FlagAssignment is a total or partial mapping of FlagNames to Bool flag values. It represents the flags chosen by the user or discovered during configuration. For example --flags=foo --flags=-bar becomes [("foo", True), ("bar", False)]

showFlagValue :: (FlagName, Bool) -> String

String representation of a flag-value pair.

dispFlagAssignment :: FlagAssignment -> Doc

Pretty-prints a flag assignment.

parseFlagAssignment :: ReadP r FlagAssignment

Parses a flag assignment.

data CondTree v c a

A CondTree is used to represent the conditional structure of a Cabal file, reflecting a syntax element subject to constraints, and then any number of sub-elements which may be enabled subject to some condition. Both a and c are usually Monoids.

To be more concrete, consider the following fragment of a Cabal file:

build-depends: base >= 4.0
if flag(extra)
    build-depends: base >= 4.2

One way to represent this is to have CondTree ConfVar [Dependency] BuildInfo. Here, condTreeData represents the actual fields which are not behind any conditional, while condTreeComponents recursively records any further fields which are behind a conditional. condTreeConstraints records the constraints (in this case, base >= 4.0) which would be applied if you use this syntax; in general, this is derived off of targetBuildInfo (perhaps a good refactoring would be to convert this into an opaque type, with a smart constructor that pre-computes the dependencies.)

Constructors

CondNode

Fields

condTreeData :: a

condTreeConstraints :: c

condTreeComponents :: [CondBranch v c a]

data ConfVar

A ConfVar represents the variable type used.

Constructors

OS OS

Arch Arch

Flag FlagName

Impl CompilerFlavor VersionRange

data Condition c

A boolean expression parameterized over the variable type used.

Constructors

Var c

Lit Bool

CNot (Condition c)

COr (Condition c) (Condition c)

CAnd (Condition c) (Condition c)

cNot :: Condition a -> Condition a

Boolean negation of a Condition value.

cAnd :: Condition a -> Condition a -> Condition a

Boolean AND of two Condtion values.

cOr :: Eq v => Condition v -> Condition v -> Condition v

Boolean OR of two Condition values.

Source repositories

data SourceRepo

Information about the source revision control system for a package.

When specifying a repo it is useful to know the meaning or intention of the information as doing so enables automation. There are two obvious common purposes: one is to find the repo for the latest development version, the other is to find the repo for this specific release. The ReopKind specifies which one we mean (or another custom one).

A package can specify one or the other kind or both. Most will specify just a head repo but some may want to specify a repo to reconstruct the sources for this package release.

The required information is the RepoType which tells us if it's using Darcs, Git for example. The repoLocation and other details are interpreted according to the repo type.

Constructors

SourceRepo

Fields

repoKind :: RepoKind

The kind of repo. This field is required.

repoType :: Maybe RepoType

The type of the source repository system for this repo, eg Darcs or Git. This field is required.

repoLocation :: Maybe String

The location of the repository. For most RepoTypes this is a URL. This field is required.

repoModule :: Maybe String

CVS can put multiple "modules" on one server and requires a module name in addition to the location to identify a particular repo. Logically this is part of the location but unfortunately has to be specified separately. This field is required for the CVS RepoType and should not be given otherwise.

repoBranch :: Maybe String

The name or identifier of the branch, if any. Many source control systems have the notion of multiple branches in a repo that exist in the same location. For example Git and CVS use this while systems like Darcs use different locations for different branches. This field is optional but should be used if necessary to identify the sources, especially for the RepoThis repo kind.

repoTag :: Maybe String

The tag identify a particular state of the repository. This should be given for the RepoThis repo kind and not for RepoHead kind.

repoSubdir :: Maybe FilePath

Some repositories contain multiple projects in different subdirectories This field specifies the subdirectory where this packages sources can be found, eg the subdirectory containing the .cabal file. It is interpreted relative to the root of the repository. This field is optional. If not given the default is "." ie no subdirectory.

data RepoKind

What this repo info is for, what it represents.

Constructors

RepoHead
The repository for the "head" or development version of the project. This repo is where we should track the latest development activity or the usual repo people should get to contribute patches.

RepoThis
The repository containing the sources for this exact package version or release. For this kind of repo a tag should be given to give enough information to re-create the exact sources.

RepoKindUnknown String

data RepoType

An enumeration of common source control systems. The fields used in the SourceRepo depend on the type of repo. The tools and methods used to obtain and track the repo depend on the repo type.

Constructors

Darcs

Git

SVN

CVS

Mercurial

GnuArch

Bazaar

Monotone

OtherRepoType String

knownRepoTypes :: [RepoType]

emptySourceRepo :: RepoKind -> SourceRepo

Custom setup build information

data SetupBuildInfo

Constructors

SetupBuildInfo

Fields

setupDepends :: [Dependency]

defaultSetupDepends :: Bool

Is this a default 'custom-setup' section added by the cabal-install code (as opposed to user-provided)? This field is only used internally, and doesn't correspond to anything in the .cabal file. See #3199.