Distribution.Pretty

Documentation

class Pretty a where

Minimal complete definition

pretty

Methods

pretty :: a -> Doc

prettyShow :: Pretty a => a -> String

defaultStyle :: Style

The default rendering style used in Cabal for console output. It has a fixed page width and adds line breaks automatically.

flatStyle :: Style

A style for rendering all on one line.

Utilities

showFilePath :: FilePath -> Doc

showToken :: String -> Doc

showFreeText :: String -> Doc

Pretty-print free-format text, ensuring that it is vertically aligned, and with blank lines replaced by dots for correct re-parsing.

indentWith :: Int

the indentation used for pretty printing

Deprecated

type Separator = [Doc] -> Doc