Distribution.Simple

Description

This is the command line front end to the Simple build system. When given the parsed command-line args and package information, is able to perform basic commands like configure, build, install, register, etc.

This module exports the main functions that Setup.hs scripts use. It re-exports the UserHooks type, the standard entry points like defaultMain and defaultMainWithHooks and the predefined sets of UserHooks that custom Setup.hs scripts can extend to add their own behaviour.

This module isn't called "Simple" because it's simple. Far from it. It's called "Simple" because it does complicated things to simple software.

The original idea was that there could be different build systems that all presented the same compatible command line interfaces. There is still a Distribution.Make system but in practice no packages use it.

Documentation

module Distribution.Package

module Distribution.Version

module Distribution.License

data AbiTag

Constructors

NoAbiTag

AbiTag String

data CompilerInfo

Compiler information used for resolving configurations. Some fields can be set to Nothing to indicate that the information is unknown.

Constructors

CompilerInfo

Fields

compilerInfoId :: CompilerId

Compiler flavour and version.

compilerInfoAbiTag :: AbiTag

Tag for distinguishing incompatible ABI's on the same architecture/os.

compilerInfoCompat :: Maybe [CompilerId]

Other implementations that this compiler claims to be compatible with, if known.

compilerInfoLanguages :: Maybe [Language]

Supported language standards, if known.

compilerInfoExtensions :: Maybe [Extension]

Supported extensions, if known.

data CompilerId

Constructors

CompilerId CompilerFlavor Version

data PerCompilerFlavor v

PerCompilerFlavor carries only info per GHC and GHCJS

Cabal parses only ghc-options and ghcjs-options, others are omitted.

Constructors

PerCompilerFlavor v v

data CompilerFlavor

Constructors

GHC

GHCJS

NHC

YHC

Hugs

HBC

Helium

JHC

LHC

UHC

Eta

HaskellSuite String

OtherCompiler String

knownCompilerFlavors :: [CompilerFlavor]

classifyCompilerFlavor :: String -> CompilerFlavor

buildCompilerFlavor :: CompilerFlavor

buildCompilerId :: CompilerId

defaultCompilerFlavor :: Maybe CompilerFlavor

The default compiler flavour to pick when compiling stuff. This defaults to the compiler used to build the Cabal lib.

However if it's not a recognised compiler then it's Nothing and the user will have to specify which compiler they want.

perCompilerFlavorToList :: PerCompilerFlavor v -> [(CompilerFlavor, v)]

abiTagString :: AbiTag -> String

unknownCompilerInfo :: CompilerId -> AbiTag -> CompilerInfo

Make a CompilerInfo of which only the known information is its CompilerId, its AbiTag and that it does not claim to be compatible with other compiler id's.

data ProfDetailLevel

Some compilers (notably GHC) support profiling and can instrument programs so the system can account costs to different functions. There are different levels of detail that can be used for this accounting. For compilers that do not support this notion or the particular detail levels, this is either ignored or just capped to some similar level they do support.

Constructors

ProfDetailNone

ProfDetailDefault

ProfDetailExportedFunctions

ProfDetailToplevelFunctions

ProfDetailAllFunctions

ProfDetailOther String

data DebugInfoLevel

Some compilers support emitting debug info. Some have different levels. For compilers that do not the level is just capped to the level they do support.

Constructors

NoDebugInfo

MinimalDebugInfo

NormalDebugInfo

MaximalDebugInfo

data OptimisationLevel

Some compilers support optimising. Some have different levels. For compilers that do not the level is just capped to the level they do support.

Constructors

NoOptimisation

NormalOptimisation

MaximumOptimisation

type PackageDBStack = [PackageDB]

We typically get packages from several databases, and stack them together. This type lets us be explicit about that stacking. For example typical stacks include:

[GlobalPackageDB]
[GlobalPackageDB, UserPackageDB]
[GlobalPackageDB, SpecificPackageDB "package.conf.inplace"]

Note that the GlobalPackageDB is invariably at the bottom since it contains the rts, base and other special compiler-specific packages.

We are not restricted to using just the above combinations. In particular we can use several custom package dbs and the user package db together.

When it comes to writing, the top most (last) package is used.

data PackageDB

Some compilers have a notion of a database of available packages. For some there is just one global db of packages, other compilers support a per-user or an arbitrary db specified at some location in the file system. This can be used to build isloated environments of packages, for example to build a collection of related packages without installing them globally.

Constructors

GlobalPackageDB

UserPackageDB

SpecificPackageDB FilePath

data Compiler

Constructors

Compiler

Fields

compilerId :: CompilerId

Compiler flavour and version.

compilerAbiTag :: AbiTag

Tag for distinguishing incompatible ABI's on the same architecture/os.

compilerCompat :: [CompilerId]

Other implementations that this compiler claims to be compatible with.

compilerLanguages :: [(Language, Flag)]

Supported language standards.

compilerExtensions :: [(Extension, Maybe Flag)]

Supported extensions.

compilerProperties :: Map String String

A key-value map for properties not covered by the above fields.

showCompilerId :: Compiler -> String

showCompilerIdWithAbi :: Compiler -> String

compilerFlavor :: Compiler -> CompilerFlavor

compilerVersion :: Compiler -> Version

compilerCompatFlavor :: CompilerFlavor -> Compiler -> Bool

Is this compiler compatible with the compiler flavour we're interested in?

For example this checks if the compiler is actually GHC or is another compiler that claims to be compatible with some version of GHC, e.g. GHCJS.

if compilerCompatFlavor GHC compiler then ... else ...

compilerCompatVersion :: CompilerFlavor -> Compiler -> Maybe Version

Is this compiler compatible with the compiler flavour we're interested in, and if so what version does it claim to be compatible with.

For example this checks if the compiler is actually GHC-7.x or is another compiler that claims to be compatible with some GHC-7.x version.

case compilerCompatVersion GHC compiler of
  Just (Version (7:_)) -> ...
  _                    -> ...

compilerInfo :: Compiler -> CompilerInfo

registrationPackageDB :: PackageDBStack -> PackageDB

Return the package that we should register into. This is the package db at the top of the stack.

absolutePackageDBPaths :: PackageDBStack -> NoCallStackIO PackageDBStack

Make package paths absolute

absolutePackageDBPath :: PackageDB -> NoCallStackIO PackageDB

flagToOptimisationLevel :: Maybe String -> OptimisationLevel

flagToDebugInfoLevel :: Maybe String -> DebugInfoLevel

unsupportedLanguages :: Compiler -> [Language] -> [Language]

languageToFlags :: Compiler -> Maybe Language -> [Flag]

unsupportedExtensions :: Compiler -> [Extension] -> [Extension]

For the given compiler, return the extensions it does not support.

extensionsToFlags :: Compiler -> [Extension] -> [Flag]

For the given compiler, return the flags for the supported extensions.

parmakeSupported :: Compiler -> Bool

Does this compiler support parallel --make mode?

reexportedModulesSupported :: Compiler -> Bool

Does this compiler support reexported-modules?

renamingPackageFlagsSupported :: Compiler -> Bool

Does this compiler support thinning/renaming on package flags?

unifiedIPIDRequired :: Compiler -> Bool

Does this compiler have unified IPIDs (so no package keys)

packageKeySupported :: Compiler -> Bool

Does this compiler support package keys?

unitIdSupported :: Compiler -> Bool

Does this compiler support unit IDs?

backpackSupported :: Compiler -> Bool

Does this compiler support Backpack?

libraryDynDirSupported :: Compiler -> Bool

Does this compiler support a package database entry with: "dynamic-library-dirs"?

arResponseFilesSupported :: Compiler -> Bool

Does this compiler's "ar" command supports response file arguments (i.e. @file-style arguments).

coverageSupported :: Compiler -> Bool

Does this compiler support Haskell program coverage?

profilingSupported :: Compiler -> Bool

Does this compiler support profiling?

flagToProfDetailLevel :: String -> ProfDetailLevel

knownProfDetailLevels :: [(String, [String], ProfDetailLevel)]

showProfDetailLevel :: ProfDetailLevel -> String

module Language.Haskell.Extension

Simple interface

defaultMain :: IO ()

A simple implementation of main for a Cabal setup script. It reads the package description file using IO, and performs the action specified on the command line.

defaultMainNoRead :: GenericPackageDescription -> IO ()

Like defaultMain, but accepts the package description as input rather than using IO to read it.

defaultMainArgs :: [String] -> IO ()

A version of defaultMain that is passed the command line arguments, rather than getting them from the environment.

Customization

data UserHooks

Hooks allow authors to add specific functionality before and after a command is run, and also to specify additional preprocessors.

WARNING: The hooks interface is under rather constant flux as we try to understand users needs. Setup files that depend on this interface may break in future releases.

Constructors

UserHooks

Fields

readDesc :: IO (Maybe GenericPackageDescription)

Read the description file

hookedPreProcessors :: [PPSuffixHandler]

Custom preprocessors in addition to and overriding knownSuffixHandlers.

hookedPrograms :: [Program]

These programs are detected at configure time. Arguments for them are added to the configure command.

preConf :: Args -> ConfigFlags -> IO HookedBuildInfo

Hook to run before configure command

confHook :: (GenericPackageDescription, HookedBuildInfo) -> ConfigFlags -> IO LocalBuildInfo

Over-ride this hook to get different behavior during configure.

postConf :: Args -> ConfigFlags -> PackageDescription -> LocalBuildInfo -> IO ()

Hook to run after configure command

preBuild :: Args -> BuildFlags -> IO HookedBuildInfo

Hook to run before build command. Second arg indicates verbosity level.

buildHook :: PackageDescription -> LocalBuildInfo -> UserHooks -> BuildFlags -> IO ()

Over-ride this hook to get different behavior during build.

postBuild :: Args -> BuildFlags -> PackageDescription -> LocalBuildInfo -> IO ()

Hook to run after build command. Second arg indicates verbosity level.

preRepl :: Args -> ReplFlags -> IO HookedBuildInfo

Hook to run before repl command. Second arg indicates verbosity level.

replHook :: PackageDescription -> LocalBuildInfo -> UserHooks -> ReplFlags -> [String] -> IO ()

Over-ride this hook to get different behavior during interpretation.

postRepl :: Args -> ReplFlags -> PackageDescription -> LocalBuildInfo -> IO ()

Hook to run after repl command. Second arg indicates verbosity level.

preClean :: Args -> CleanFlags -> IO HookedBuildInfo

Hook to run before clean command. Second arg indicates verbosity level.

cleanHook :: PackageDescription -> () -> UserHooks -> CleanFlags -> IO ()

Over-ride this hook to get different behavior during clean.

postClean :: Args -> CleanFlags -> PackageDescription -> () -> IO ()

Hook to run after clean command. Second arg indicates verbosity level.

preCopy :: Args -> CopyFlags -> IO HookedBuildInfo

Hook to run before copy command

copyHook :: PackageDescription -> LocalBuildInfo -> UserHooks -> CopyFlags -> IO ()

Over-ride this hook to get different behavior during copy.

postCopy :: Args -> CopyFlags -> PackageDescription -> LocalBuildInfo -> IO ()

Hook to run after copy command

preInst :: Args -> InstallFlags -> IO HookedBuildInfo

Hook to run before install command

instHook :: PackageDescription -> LocalBuildInfo -> UserHooks -> InstallFlags -> IO ()

Over-ride this hook to get different behavior during install.

postInst :: Args -> InstallFlags -> PackageDescription -> LocalBuildInfo -> IO ()

Hook to run after install command. postInst should be run on the target, not on the build machine.

preReg :: Args -> RegisterFlags -> IO HookedBuildInfo

Hook to run before register command

regHook :: PackageDescription -> LocalBuildInfo -> UserHooks -> RegisterFlags -> IO ()

Over-ride this hook to get different behavior during registration.

postReg :: Args -> RegisterFlags -> PackageDescription -> LocalBuildInfo -> IO ()

Hook to run after register command

preUnreg :: Args -> RegisterFlags -> IO HookedBuildInfo

Hook to run before unregister command

unregHook :: PackageDescription -> LocalBuildInfo -> UserHooks -> RegisterFlags -> IO ()

Over-ride this hook to get different behavior during unregistration.

postUnreg :: Args -> RegisterFlags -> PackageDescription -> LocalBuildInfo -> IO ()

Hook to run after unregister command

preHscolour :: Args -> HscolourFlags -> IO HookedBuildInfo

Hook to run before hscolour command. Second arg indicates verbosity level.

hscolourHook :: PackageDescription -> LocalBuildInfo -> UserHooks -> HscolourFlags -> IO ()

Over-ride this hook to get different behavior during hscolour.

postHscolour :: Args -> HscolourFlags -> PackageDescription -> LocalBuildInfo -> IO ()

Hook to run after hscolour command. Second arg indicates verbosity level.

preDoctest :: Args -> DoctestFlags -> IO HookedBuildInfo

Hook to run before doctest command. Second arg indicates verbosity level.

doctestHook :: PackageDescription -> LocalBuildInfo -> UserHooks -> DoctestFlags -> IO ()

Over-ride this hook to get different behavior during doctest.

postDoctest :: Args -> DoctestFlags -> PackageDescription -> LocalBuildInfo -> IO ()

Hook to run after doctest command. Second arg indicates verbosity level.

preHaddock :: Args -> HaddockFlags -> IO HookedBuildInfo

Hook to run before haddock command. Second arg indicates verbosity level.

haddockHook :: PackageDescription -> LocalBuildInfo -> UserHooks -> HaddockFlags -> IO ()

Over-ride this hook to get different behavior during haddock.

postHaddock :: Args -> HaddockFlags -> PackageDescription -> LocalBuildInfo -> IO ()

Hook to run after haddock command. Second arg indicates verbosity level.

preTest :: Args -> TestFlags -> IO HookedBuildInfo

Hook to run before test command.

testHook :: Args -> PackageDescription -> LocalBuildInfo -> UserHooks -> TestFlags -> IO ()

Over-ride this hook to get different behavior during test.

postTest :: Args -> TestFlags -> PackageDescription -> LocalBuildInfo -> IO ()

Hook to run after test command.

preBench :: Args -> BenchmarkFlags -> IO HookedBuildInfo

Hook to run before bench command.

benchHook :: Args -> PackageDescription -> LocalBuildInfo -> UserHooks -> BenchmarkFlags -> IO ()

Over-ride this hook to get different behavior during bench.

postBench :: Args -> BenchmarkFlags -> PackageDescription -> LocalBuildInfo -> IO ()

Hook to run after bench command.

type Args = [String]

defaultMainWithHooks :: UserHooks -> IO ()

A customizable version of defaultMain.

defaultMainWithHooksArgs :: UserHooks -> [String] -> IO ()

A customizable version of defaultMain that also takes the command line arguments.

defaultMainWithHooksNoRead :: UserHooks -> GenericPackageDescription -> IO ()

A customizable version of defaultMainNoRead.

defaultMainWithHooksNoReadArgs :: UserHooks -> GenericPackageDescription -> [String] -> IO ()

A customizable version of defaultMainNoRead that also takes the command line arguments.

Since: Cabal-2.2.0.0

Standard sets of hooks

simpleUserHooks :: UserHooks

Hooks that correspond to a plain instantiation of the "simple" build system

autoconfUserHooks :: UserHooks

Basic autoconf UserHooks:

postConf runs ./configure, if present.

the pre-hooks preBuild, preClean, preCopy, preInst, preReg and preUnreg read additional build information from package.buildinfo, if present.

Thus configure can use local system information to generate package.buildinfo and possibly other files.

emptyUserHooks :: UserHooks

Empty UserHooks which do nothing.