Distribution.Simple.Build

Description

This is the entry point to actually building the modules in a package. It doesn't actually do much itself, most of the work is delegated to compiler-specific actions. It does do some non-compiler specific bits like running pre-processors.

Documentation

build

Arguments

:: PackageDescription
Mostly information from the .cabal file

-> LocalBuildInfo
Configuration information

-> BuildFlags
Flags that the user passed to build

-> [PPSuffixHandler]
preprocessors to run before compiling

-> IO ()

Build the libraries and executables in this package.

repl

Arguments

:: PackageDescription
Mostly information from the .cabal file

-> LocalBuildInfo
Configuration information

-> ReplFlags
Flags that the user passed to build

-> [PPSuffixHandler]
preprocessors to run before compiling

-> [String]

-> IO ()

startInterpreter :: Verbosity -> ProgramDb -> Compiler -> PackageDBStack -> IO ()

Start an interpreter without loading any package files.

initialBuildSteps

Arguments

:: FilePath
"dist" prefix

-> PackageDescription
mostly information from the .cabal file

-> LocalBuildInfo
Configuration information

-> Verbosity
The verbosity to use

-> IO ()

writeAutogenFiles :: Verbosity -> PackageDescription -> LocalBuildInfo -> IO ()

Generate and write out the Paths_pkg.hs and cabal_macros.h files