Distribution.Simple.Build.PathsModule

Description

Generating the Paths_pkgname module.

This is a module that Cabal generates for the benefit of packages. It enables them to find their version number and find any installed data files at runtime. This code should probably be split off into another module.

Documentation

generatePathsModule :: PackageDescription -> LocalBuildInfo -> ComponentLocalBuildInfo -> String

pkgPathEnvVar

Arguments

:: PackageDescription

-> String
path component; one of "bindir", "libdir", "datadir", "libexecdir", or "sysconfdir"

-> String
environment variable name

Generates the name of the environment variable controlling the path component of interest.

Note: The format of these strings is part of Cabal's public API; changing this function constitutes a *backwards-compatibility* break.