Distribution.Simple.BuildPaths

Description

A bunch of dirs, paths and file names used for intermediate build steps.

Documentation

defaultDistPref :: FilePath

srcPref :: FilePath -> FilePath

buildInfoPref :: FilePath -> FilePath

Build info json file, generated in every build

haddockDirName :: HaddockTarget -> PackageDescription -> FilePath

This is the name of the directory in which the generated haddocks should be stored. It does not include the distdochtml prefix.

hscolourPref :: HaddockTarget -> FilePath -> PackageDescription -> FilePath

haddockPref :: HaddockTarget -> FilePath -> PackageDescription -> FilePath

The directory to which generated haddock documentation should be written.

autogenPackageModulesDir :: LocalBuildInfo -> String

The directory in which we put auto-generated modules for EVERY component in the package.

autogenComponentModulesDir :: LocalBuildInfo -> ComponentLocalBuildInfo -> String

The directory in which we put auto-generated modules for a particular component.

autogenPathsModuleName :: PackageDescription -> ModuleName

The name of the auto-generated Paths_* module associated with a package

autogenPackageInfoModuleName :: PackageDescription -> ModuleName

The name of the auto-generated PackageInfo_* module associated with a package

cppHeaderName :: String

haddockName :: PackageDescription -> FilePath

mkGenericStaticLibName :: String -> String

Create a library name for a static library from a given name. Prepends lib and appends the static library extension (.a).

mkLibName :: UnitId -> String

mkProfLibName :: UnitId -> String

mkGenericSharedLibName :: Platform -> CompilerId -> String -> String

Create a library name for a shared library from a given name. Prepends lib and appends the -<compilerFlavour><compilerVersion> as well as the shared library extension.

mkSharedLibName :: Platform -> CompilerId -> UnitId -> String

mkStaticLibName :: Platform -> CompilerId -> UnitId -> String

mkGenericSharedBundledLibName :: Platform -> CompilerId -> String -> String

Create a library name for a bundled shared library from a given name. This matches the naming convention for shared libraries as implemented in GHC's packageHsLibs function in the Packages module. If the given name is prefixed with HS, then this prepends lib and appends the compiler flavour/version and shared library extension e.g.: "HSrts-1.0" -> "libHSrts-1.0-ghc8.7.20190109.so" Otherwise the given name should be prefixed with C, then this strips the C, prepends lib and appends the shared library extension e.g.: Cffi -> "libffi.so"

exeExtension :: Platform -> String

Default extension for executable files on the current platform. (typically "" on Unix and "exe" on Windows or OS/2)

objExtension :: String

Extension for object files. For GHC the extension is "o".

dllExtension :: Platform -> String

Extension for dynamically linked (or shared) libraries (typically "so" on Unix and "dll" on Windows)

staticLibExtension :: Platform -> String

Extension for static libraries

TODO: Here, as well as in dllExtension, it's really the target OS that we're interested in, not the build OS.

Source files & build directories

getSourceFiles :: Verbosity -> [FilePath] -> [ModuleName] -> IO [(ModuleName, FilePath)]

getLibSourceFiles :: Verbosity -> LocalBuildInfo -> Library -> ComponentLocalBuildInfo -> IO [(ModuleName, FilePath)]

getExeSourceFiles :: Verbosity -> LocalBuildInfo -> Executable -> ComponentLocalBuildInfo -> IO [(ModuleName, FilePath)]

getFLibSourceFiles :: Verbosity -> LocalBuildInfo -> ForeignLib -> ComponentLocalBuildInfo -> IO [(ModuleName, FilePath)]

exeBuildDir :: LocalBuildInfo -> Executable -> FilePath

The directory where we put build results for an executable

flibBuildDir :: LocalBuildInfo -> ForeignLib -> FilePath

The directory where we put build results for a foreign library