Distribution.Simple.BuildPaths

Description

A bunch of dirs, paths and file names used for intermediate build steps.

Documentation

defaultDistPref :: FilePath

srcPref :: FilePath -> FilePath

haddockDirName :: HaddockTarget -> PackageDescription -> FilePath

This is the name of the directory in which the generated haddocks should be stored. It does not include the distdochtml prefix.

hscolourPref :: FilePath -> PackageDescription -> FilePath

hscolourPref' :: HaddockTarget -> FilePath -> PackageDescription -> FilePath

haddockPref :: FilePath -> PackageDescription -> FilePath

haddockPref' :: HaddockTarget -> FilePath -> PackageDescription -> FilePath

The directory to which generated haddock documentation should be written.

autogenModulesDir :: LocalBuildInfo -> String

The directory in which we put auto-generated modules

autogenModuleName :: PackageDescription -> ModuleName

The name of the auto-generated module associated with a package

cppHeaderName :: String

haddockName :: PackageDescription -> FilePath

mkLibName :: UnitId -> String

mkProfLibName :: UnitId -> String

mkSharedLibName :: CompilerId -> UnitId -> String

exeExtension :: String

Default extension for executable files on the current platform. (typically "" on Unix and "exe" on Windows or OS/2)

objExtension :: String

Extension for object files. For GHC the extension is "o".

dllExtension :: String

Extension for dynamically linked (or shared) libraries (typically "so" on Unix and "dll" on Windows)