Distribution.Simple.BuildTarget

Description

Handling for user-specified build targets

Main interface

readTargetInfos :: Verbosity -> PackageDescription -> LocalBuildInfo -> [String] -> IO [TargetInfo]

Take a list of String build targets, and parse and validate them into actual TargetInfos to be builtregisteredwhatever.

readBuildTargets :: PackageDescription -> [String] -> IO [BuildTarget]

Read a list of user-supplied build target strings and resolve them to BuildTargets according to a PackageDescription. If there are problems with any of the targets e.g. they don't exist or are misformatted, throw an IOException.

Build targets

data BuildTarget

A fully resolved build target.

Constructors

BuildTargetComponent ComponentName
A specific component

BuildTargetModule ComponentName ModuleName
A specific module within a specific component.

BuildTargetFile ComponentName FilePath
A specific file within a specific component.

showBuildTarget :: PackageId -> BuildTarget -> String

Unambiguously render a BuildTarget, so that it can be parsed in all situations.

data QualLevel

Constructors

QL1

QL2

QL3

buildTargetComponentName :: BuildTarget -> ComponentName

Parsing user build targets

data UserBuildTarget

Various ways that a user may specify a build target.

readUserBuildTargets :: [String] -> ([UserBuildTargetProblem], [UserBuildTarget])

showUserBuildTarget :: UserBuildTarget -> String

data UserBuildTargetProblem

Constructors

UserBuildTargetUnrecognised String

reportUserBuildTargetProblems :: [UserBuildTargetProblem] -> IO ()

Resolving build targets

resolveBuildTargets :: PackageDescription -> [(UserBuildTarget, Bool)] -> ([BuildTargetProblem], [BuildTarget])

Given a bunch of user-specified targets, try to resolve what it is they refer to.

data BuildTargetProblem

Constructors

BuildTargetExpected UserBuildTarget [String] String

expected thing

(actually got)

BuildTargetNoSuch UserBuildTarget [(String, String)]

(no such thing, actually got)

BuildTargetAmbiguous UserBuildTarget [(UserBuildTarget, BuildTarget)]

reportBuildTargetProblems :: [BuildTargetProblem] -> IO ()