Distribution.Simple.BuildTarget

Description

Handling for user-specified build targets

Build targets

data BuildTarget

A fully resolved build target.

Constructors

BuildTargetComponent ComponentName
A specific component

BuildTargetModule ComponentName ModuleName
A specific module within a specific component.

BuildTargetFile ComponentName FilePath
A specific file within a specific component.

readBuildTargets :: PackageDescription -> [String] -> IO [BuildTarget]

Parsing user build targets

data UserBuildTarget

Various ways that a user may specify a build target.

readUserBuildTargets :: [String] -> ([UserBuildTargetProblem], [UserBuildTarget])

data UserBuildTargetProblem

Constructors

UserBuildTargetUnrecognised String

reportUserBuildTargetProblems :: [UserBuildTargetProblem] -> IO ()

Resolving build targets

resolveBuildTargets :: PackageDescription -> [(UserBuildTarget, Bool)] -> ([BuildTargetProblem], [BuildTarget])

Given a bunch of user-specified targets, try to resolve what it is they refer to.

data BuildTargetProblem

Constructors

BuildTargetExpected UserBuildTarget [String] String

expected thing

(actually got)

BuildTargetNoSuch UserBuildTarget [(String, String)]

(no such thing, actually got)

BuildTargetAmbigious UserBuildTarget [(UserBuildTarget, BuildTarget)]

reportBuildTargetProblems :: [BuildTargetProblem] -> IO ()