Distribution.Simple.Glob

Description

Simple file globbing.

Documentation

data GlobSyntaxError

Constructors

StarInDirectory

StarInFileName

StarInExtension

NoExtensionOnStar

EmptyGlob

LiteralFileNameGlobStar

VersionDoesNotSupportGlobStar

VersionDoesNotSupportGlob

data GlobResult a

Constructors

GlobMatch a
The glob matched the value supplied.

GlobWarnMultiDot a
The glob did not match the value supplied because the cabal-version is too low and the extensions on the file did not precisely match the glob's extensions, but rather the glob was a proper suffix of the file's extensions; i.e., if not for the low cabal-version, it would have matched.

GlobMissingDirectory FilePath
The glob couldn't match because the directory named doesn't exist. The directory will be as it appears in the glob (i.e., relative to the directory passed to matchDirFileGlob, and, for 'data-files', relative to 'data-dir').

matchDirFileGlob :: Verbosity -> CabalSpecVersion -> FilePath -> FilePath -> IO [FilePath]

This will die' when the glob matches no files, or if the glob refers to a missing directory, or if the glob fails to parse.

The Version argument must be the spec version of the package description being processed, as globs behave slightly differently in different spec versions.

The first FilePath argument is the directory that the glob is relative to. It must be a valid directory (and hence it can't be the empty string). The returned values will not include this prefix.

The second FilePath is the glob itself.

matchDirFileGlobWithDie :: Verbosity -> (Verbosity -> String -> IO [FilePath]) -> CabalSpecVersion -> FilePath -> FilePath -> IO [FilePath]

Like matchDirFileGlob but with customizable die

Since: Cabal-3.6.0.0

runDirFileGlob :: Verbosity -> FilePath -> Glob -> IO [GlobResult FilePath]

Match files against a pre-parsed glob, starting in a directory.

The Version argument must be the spec version of the package description being processed, as globs behave slightly differently in different spec versions.

The FilePath argument is the directory that the glob is relative to. It must be a valid directory (and hence it can't be the empty string). The returned values will not include this prefix.

fileGlobMatches :: Glob -> FilePath -> Maybe (GlobResult FilePath)

Returns Nothing if the glob didn't match at all, or Just the result if the glob matched (or would have matched with a higher cabal-version).

parseFileGlob :: CabalSpecVersion -> FilePath -> Either GlobSyntaxError Glob

explainGlobSyntaxError :: FilePath -> GlobSyntaxError -> String

isRecursiveInRoot :: Glob -> Bool

data Glob