Distribution.Simple.Haddock

Description

This module deals with the haddock and hscolour commands. Sadly this is a rather complicated module. It deals with two versions of haddock (0.x and 2.x). It has to do pre-processing for haddock 0.x which involves 'unlit'ing and using -DHADDOCK for any source code that uses cpp. It uses information about installed packages (from ghc-pkg) to find the locations of documentation for dependent packages, so it can create links.

The hscolour support allows generating html versions of the original source, with coloured syntax highlighting.

Documentation

haddock :: PackageDescription -> LocalBuildInfo -> [PPSuffixHandler] -> HaddockFlags -> IO ()

hscolour :: PackageDescription -> LocalBuildInfo -> [PPSuffixHandler] -> HscolourFlags -> IO ()