Distribution.Simple.Haddock

Description

This module deals with the haddock and hscolour commands. It uses information about installed packages (from ghc-pkg) to find the locations of documentation for dependent packages, so it can create links.

The hscolour support allows generating HTML versions of the original source, with coloured syntax highlighting.

Documentation

haddock :: PackageDescription -> LocalBuildInfo -> [PPSuffixHandler] -> HaddockFlags -> IO ()

hscolour :: PackageDescription -> LocalBuildInfo -> [PPSuffixHandler] -> HscolourFlags -> IO ()

haddockPackagePaths :: [InstalledPackageInfo] -> Maybe (InstalledPackageInfo -> FilePath) -> IO ([(FilePath, Maybe FilePath)], Maybe String)

Given a list of InstalledPackageInfos, return a list of interfaces and HTML paths, and an optional warning for packages with missing documentation.