Distribution.Simple.Hpc

Description

This module provides functions for locating various HPC-related paths and a function for adding the necessary options to a PackageDescription to build test suites with HPC enabled.

Documentation

data Way

Constructors

Vanilla

Prof

Dyn

guessWay :: LocalBuildInfo -> Way

Attempt to guess the way the test suites in this package were compiled and linked with the library so the correct module interfaces are found.

htmlDir

Arguments

:: FilePath
"dist/" prefix

-> Way

-> FilePath
Component name

-> FilePath
Path to test suite's HTML markup directory

mixDir

Arguments

:: FilePath
"dist/" prefix

-> Way

-> FilePath
Component name

-> FilePath
Directory containing test suite's .mix files

tixDir

Arguments

:: FilePath
"dist/" prefix

-> Way

-> FilePath
Component name

-> FilePath
Directory containing test suite's .tix files

tixFilePath

Arguments

:: FilePath
"dist/" prefix

-> Way

-> FilePath
Component name

-> FilePath
Path to test suite's .tix file

Path to the .tix file containing a test suite's sum statistics.

markupPackage

Arguments

:: Verbosity

-> LocalBuildInfo

-> FilePath
"dist/" prefix

-> String
Library name

-> [TestSuite]

-> IO ()

Generate the HTML markup for all of a package's test suites.

markupTest

Arguments

:: Verbosity

-> LocalBuildInfo

-> FilePath
"dist/" prefix

-> String
Library name

-> TestSuite

-> IO ()

Generate the HTML markup for a test suite.