Distribution.Simple.Hpc

Description

This module provides functions for locating various HPC-related paths and a function for adding the necessary options to a PackageDescription to build test suites with HPC enabled.

Documentation

enableCoverage

Arguments

:: Bool
Enable coverage?

-> String
"dist/" prefix

-> PackageDescription

-> PackageDescription

Conditionally enable Haskell Program Coverage by adding the necessary GHC options to a PackageDescription.

TODO: do this differently in the build stage by constructing local build info, not by modifying the original PackageDescription.

htmlDir

Arguments

:: FilePath
"dist/" prefix

-> FilePath
Component name

-> FilePath
Path to test suite's HTML markup directory

tixDir

Arguments

:: FilePath
"dist/" prefix

-> FilePath
Component name

-> FilePath
Directory containing test suite's .tix files

tixFilePath

Arguments

:: FilePath
"dist/" prefix

-> FilePath
Component name

-> FilePath
Path to test suite's .tix file

Path to the .tix file containing a test suite's sum statistics.

markupPackage

Arguments

:: Verbosity

-> LocalBuildInfo

-> FilePath
"dist/" prefix

-> String
Library name

-> [TestSuite]

-> IO ()

Generate the HTML markup for all of a package's test suites.

markupTest

Arguments

:: Verbosity

-> LocalBuildInfo

-> FilePath
"dist/" prefix

-> String
Library name

-> TestSuite

-> IO ()

Generate the HTML markup for a test suite.