Distribution.Simple.LocalBuildInfo

Description

Once a package has been configured we have resolved conditionals and dependencies, configured the compiler and other needed external programs. The LocalBuildInfo is used to hold all this information. It holds the install dirs, the compiler, the exact package dependencies, the configured programs, the package database to use and a bunch of miscellaneous configure flags. It gets saved and reloaded from a file (dist/setup-config). It gets passed in to very many subsequent build actions.

Documentation

data LocalBuildInfo

Data cached after configuration step. See also ConfigFlags.

Constructors

LocalBuildInfo

Fields

configFlags :: ConfigFlags

Options passed to the configuration step. Needed to re-run configuration when .cabal is out of date

flagAssignment :: FlagAssignment

The final set of flags which were picked for this package

componentEnabledSpec :: ComponentRequestedSpec

What components were enabled during configuration, and why.

extraConfigArgs :: [String]

Extra args on the command line for the configuration step. Needed to re-run configuration when .cabal is out of date

installDirTemplates :: InstallDirTemplates

The installation directories for the various different kinds of files TODO: inplaceDirTemplates :: InstallDirs FilePath

compiler :: Compiler

The compiler we're building with

hostPlatform :: Platform

The platform we're building for

buildDir :: FilePath

Where to build the package.

componentGraph :: Graph ComponentLocalBuildInfo

All the components to build, ordered by topological sort, and with their INTERNAL dependencies over the intrapackage dependency graph. TODO: this is assumed to be short; otherwise we want some sort of ordered map.

componentNameMap :: Map ComponentName [ComponentLocalBuildInfo]

A map from component name to all matching components. These coincide with componentGraph

installedPkgs :: InstalledPackageIndex

All the info about the installed packages that the current package depends on (directly or indirectly). The copy saved on disk does NOT include internal dependencies (because we just don't have enough information at this point to have an InstalledPackageInfo for an internal dep), but we will often update it with the internal dependencies; see for example build. (This admonition doesn't apply for per-component builds.)

pkgDescrFile :: Maybe FilePath

the filename containing the .cabal file, if available

localPkgDescr :: PackageDescription

WARNING WARNING WARNING Be VERY careful about using this function; we haven't deprecated it but using it could introduce subtle bugs related to HookedBuildInfo.

In principle, this is supposed to contain the resolved package description, that does not contain any conditionals. However, it MAY NOT contain the description wtih a HookedBuildInfo applied to it; see HookedBuildInfo for the whole sordid saga. As much as possible, Cabal library should avoid using this parameter.

withPrograms :: ProgramDb

Location and args for all programs

withPackageDB :: PackageDBStack

What package database to use, global/user

withVanillaLib :: Bool

Whether to build normal libs.

withProfLib :: Bool

Whether to build profiling versions of libs.

withSharedLib :: Bool

Whether to build shared versions of libs.

withDynExe :: Bool

Whether to link executables dynamically

withProfExe :: Bool

Whether to build executables for profiling.

withProfLibDetail :: ProfDetailLevel

Level of automatic profile detail.

withProfExeDetail :: ProfDetailLevel

Level of automatic profile detail.

withOptimization :: OptimisationLevel

Whether to build with optimization (if available).

withDebugInfo :: DebugInfoLevel

Whether to emit debug info (if available).

withGHCiLib :: Bool

Whether to build libs suitable for use with GHCi.

splitObjs :: Bool

Use -split-objs with GHC, if available

stripExes :: Bool

Whether to strip executables during install

stripLibs :: Bool

Whether to strip libraries during install

exeCoverage :: Bool

Whether to enable executable program coverage

libCoverage :: Bool

Whether to enable library program coverage

progPrefix :: PathTemplate

Prefix to be prepended to installed executables

progSuffix :: PathTemplate

Suffix to be appended to installed executables

relocatable :: Bool

externalPackageDeps :: LocalBuildInfo -> [(UnitId, PackageId)]

External package dependencies for the package as a whole. This is the union of the individual componentPackageDeps, less any internal deps.

localComponentId :: LocalBuildInfo -> ComponentId

Extract the ComponentId from the public library component of a LocalBuildInfo if it exists, or make a fake component ID based on the package ID.

localUnitId :: LocalBuildInfo -> UnitId

Extract the UnitId from the library component of a LocalBuildInfo if it exists, or make a fake unit ID based on the package ID.

localCompatPackageKey :: LocalBuildInfo -> String

Extract the compatibility package key from the public library component of a LocalBuildInfo if it exists, or make a fake package key based on the package ID.

Buildable package components

data Component

Constructors

CLib Library

CFLib ForeignLib

CExe Executable

CTest TestSuite

CBench Benchmark

data ComponentName

Constructors

CLibName

CSubLibName UnqualComponentName

CFLibName UnqualComponentName

CExeName UnqualComponentName

CTestName UnqualComponentName

CBenchName UnqualComponentName

defaultLibName :: ComponentName

showComponentName :: ComponentName -> String

componentNameString :: ComponentName -> Maybe UnqualComponentName

This gets the underlying unqualified component name. In fact, it is guaranteed to uniquely identify a component, returning Nothing if the ComponentName was for the public library.

data ComponentLocalBuildInfo

The first five fields are common across all algebraic variants.

Constructors

LibComponentLocalBuildInfo

Fields

componentLocalName :: ComponentName

It would be very convenient to store the literal Library here, but if we do that, it will get serialized (via the Binary) instance twice. So instead we just provide the ComponentName, which can be used to find the Component in the PackageDescription. NB: eventually, this will NOT uniquely identify the ComponentLocalBuildInfo.

componentComponentId :: ComponentId

The computed ComponentId of this component.

componentUnitId :: UnitId

The computed UnitId which uniquely identifies this component. Might be hashed.

componentIsIndefinite_ :: Bool

Is this an indefinite component (i.e. has unfilled holes)?

componentInstantiatedWith :: [(ModuleName, OpenModule)]

How the component was instantiated

componentPackageDeps :: [(UnitId, PackageId)]

Resolved internal and external package dependencies for this component. The BuildInfo specifies a set of build dependencies that must be satisfied in terms of version ranges. This field fixes those dependencies to the specific versions available on this machine for this compiler.

componentIncludes :: [(OpenUnitId, ModuleRenaming)]

The set of packages that are brought into scope during compilation, including a ModuleRenaming which may used to hide or rename modules. This is what gets translated into -package-id arguments. This is a modernized version of componentPackageDeps, which is kept around for BC purposes.

componentExeDeps :: [UnitId]

componentInternalDeps :: [UnitId]

The internal dependencies which induce a graph on the ComponentLocalBuildInfo of this package. This does NOT coincide with componentPackageDeps because it ALSO records 'build-tool' dependencies on executables. Maybe one day cabal-install will also handle these correctly too!

componentCompatPackageKey :: String

Compatibility package key that we pass to older versions of GHC.

componentCompatPackageName :: PackageName

Compatibility package name that we register this component as.

componentExposedModules :: [ExposedModule]

A list of exposed modules (either defined in this component, or reexported from another component.)

componentIsPublic :: Bool

Convenience field, specifying whether or not this is the public library that has the same name as the package.

FLibComponentLocalBuildInfo

Fields

componentLocalName :: ComponentName

It would be very convenient to store the literal Library here, but if we do that, it will get serialized (via the Binary) instance twice. So instead we just provide the ComponentName, which can be used to find the Component in the PackageDescription. NB: eventually, this will NOT uniquely identify the ComponentLocalBuildInfo.

componentComponentId :: ComponentId

The computed ComponentId of this component.

componentUnitId :: UnitId

The computed UnitId which uniquely identifies this component. Might be hashed.

componentPackageDeps :: [(UnitId, PackageId)]

Resolved internal and external package dependencies for this component. The BuildInfo specifies a set of build dependencies that must be satisfied in terms of version ranges. This field fixes those dependencies to the specific versions available on this machine for this compiler.

componentIncludes :: [(OpenUnitId, ModuleRenaming)]

The set of packages that are brought into scope during compilation, including a ModuleRenaming which may used to hide or rename modules. This is what gets translated into -package-id arguments. This is a modernized version of componentPackageDeps, which is kept around for BC purposes.

componentExeDeps :: [UnitId]

componentInternalDeps :: [UnitId]

The internal dependencies which induce a graph on the ComponentLocalBuildInfo of this package. This does NOT coincide with componentPackageDeps because it ALSO records 'build-tool' dependencies on executables. Maybe one day cabal-install will also handle these correctly too!

ExeComponentLocalBuildInfo

Fields

componentLocalName :: ComponentName

It would be very convenient to store the literal Library here, but if we do that, it will get serialized (via the Binary) instance twice. So instead we just provide the ComponentName, which can be used to find the Component in the PackageDescription. NB: eventually, this will NOT uniquely identify the ComponentLocalBuildInfo.

componentComponentId :: ComponentId

The computed ComponentId of this component.

componentUnitId :: UnitId

The computed UnitId which uniquely identifies this component. Might be hashed.

componentPackageDeps :: [(UnitId, PackageId)]

Resolved internal and external package dependencies for this component. The BuildInfo specifies a set of build dependencies that must be satisfied in terms of version ranges. This field fixes those dependencies to the specific versions available on this machine for this compiler.

componentIncludes :: [(OpenUnitId, ModuleRenaming)]

The set of packages that are brought into scope during compilation, including a ModuleRenaming which may used to hide or rename modules. This is what gets translated into -package-id arguments. This is a modernized version of componentPackageDeps, which is kept around for BC purposes.

componentExeDeps :: [UnitId]

componentInternalDeps :: [UnitId]

The internal dependencies which induce a graph on the ComponentLocalBuildInfo of this package. This does NOT coincide with componentPackageDeps because it ALSO records 'build-tool' dependencies on executables. Maybe one day cabal-install will also handle these correctly too!

TestComponentLocalBuildInfo

Fields

componentLocalName :: ComponentName

It would be very convenient to store the literal Library here, but if we do that, it will get serialized (via the Binary) instance twice. So instead we just provide the ComponentName, which can be used to find the Component in the PackageDescription. NB: eventually, this will NOT uniquely identify the ComponentLocalBuildInfo.

componentComponentId :: ComponentId

The computed ComponentId of this component.

componentUnitId :: UnitId

The computed UnitId which uniquely identifies this component. Might be hashed.

componentPackageDeps :: [(UnitId, PackageId)]

Resolved internal and external package dependencies for this component. The BuildInfo specifies a set of build dependencies that must be satisfied in terms of version ranges. This field fixes those dependencies to the specific versions available on this machine for this compiler.

componentIncludes :: [(OpenUnitId, ModuleRenaming)]

The set of packages that are brought into scope during compilation, including a ModuleRenaming which may used to hide or rename modules. This is what gets translated into -package-id arguments. This is a modernized version of componentPackageDeps, which is kept around for BC purposes.

componentExeDeps :: [UnitId]

componentInternalDeps :: [UnitId]

The internal dependencies which induce a graph on the ComponentLocalBuildInfo of this package. This does NOT coincide with componentPackageDeps because it ALSO records 'build-tool' dependencies on executables. Maybe one day cabal-install will also handle these correctly too!

BenchComponentLocalBuildInfo

Fields

componentLocalName :: ComponentName

It would be very convenient to store the literal Library here, but if we do that, it will get serialized (via the Binary) instance twice. So instead we just provide the ComponentName, which can be used to find the Component in the PackageDescription. NB: eventually, this will NOT uniquely identify the ComponentLocalBuildInfo.

componentComponentId :: ComponentId

The computed ComponentId of this component.

componentUnitId :: UnitId

The computed UnitId which uniquely identifies this component. Might be hashed.

componentPackageDeps :: [(UnitId, PackageId)]

Resolved internal and external package dependencies for this component. The BuildInfo specifies a set of build dependencies that must be satisfied in terms of version ranges. This field fixes those dependencies to the specific versions available on this machine for this compiler.

componentIncludes :: [(OpenUnitId, ModuleRenaming)]

The set of packages that are brought into scope during compilation, including a ModuleRenaming which may used to hide or rename modules. This is what gets translated into -package-id arguments. This is a modernized version of componentPackageDeps, which is kept around for BC purposes.

componentExeDeps :: [UnitId]

componentInternalDeps :: [UnitId]

The internal dependencies which induce a graph on the ComponentLocalBuildInfo of this package. This does NOT coincide with componentPackageDeps because it ALSO records 'build-tool' dependencies on executables. Maybe one day cabal-install will also handle these correctly too!

componentBuildDir :: LocalBuildInfo -> ComponentLocalBuildInfo -> FilePath

foldComponent :: (Library -> a) -> (ForeignLib -> a) -> (Executable -> a) -> (TestSuite -> a) -> (Benchmark -> a) -> Component -> a

componentName :: Component -> ComponentName

componentBuildInfo :: Component -> BuildInfo

componentBuildable :: Component -> Bool

pkgComponents :: PackageDescription -> [Component]

All the components in the package.

pkgBuildableComponents :: PackageDescription -> [Component]

lookupComponent :: PackageDescription -> ComponentName -> Maybe Component

getComponent :: PackageDescription -> ComponentName -> Component

getComponentLocalBuildInfo :: LocalBuildInfo -> ComponentName -> ComponentLocalBuildInfo

Deprecated: This function is not well-defined, because a 'ComponentName' does not uniquely identify a 'ComponentLocalBuildInfo'. If you have a 'TargetInfo', you should use 'targetCLBI' to get the 'ComponentLocalBuildInfo'. Otherwise, use 'componentNameTargets' to get all possible 'ComponentLocalBuildInfo's. This will be removed in Cabal 2.2.

allComponentsInBuildOrder :: LocalBuildInfo -> [ComponentLocalBuildInfo]

componentsInBuildOrder :: LocalBuildInfo -> [ComponentName] -> [ComponentLocalBuildInfo]

Deprecated: You've got 'TargetInfo' right? Use 'neededTargetsInBuildOrder' on the 'UnitId's you can 'nodeKey' out.

depLibraryPaths

Arguments

:: Bool
Building for inplace?

-> Bool
Generate prefix-relative library paths

-> LocalBuildInfo

-> ComponentLocalBuildInfo
Component that is being built

-> NoCallStackIO [FilePath]

Determine the directories containing the dynamic libraries of the transitive dependencies of the component we are building.

When wanted, and possible, returns paths relative to the installDirs prefix

allLibModules :: Library -> ComponentLocalBuildInfo -> [ModuleName]

Get all module names that needed to be built by GHC; i.e., all of these ModuleNames have interface files associated with them that need to be installed.

withAllComponentsInBuildOrder :: PackageDescription -> LocalBuildInfo -> (Component -> ComponentLocalBuildInfo -> IO ()) -> IO ()

Perform the action on each buildable Library or Executable (Component) in the PackageDescription, subject to the build order specified by the compBuildOrder field of the given LocalBuildInfo

withComponentsInBuildOrder :: PackageDescription -> LocalBuildInfo -> [ComponentName] -> (Component -> ComponentLocalBuildInfo -> IO ()) -> IO ()

Deprecated: You have got a 'TargetInfo' right? Use 'withNeededTargetsInBuildOrder' on the 'UnitId's you can 'nodeKey' out.

withComponentsLBI :: PackageDescription -> LocalBuildInfo -> (Component -> ComponentLocalBuildInfo -> IO ()) -> IO ()

Deprecated: Use withAllComponentsInBuildOrder

withLibLBI :: PackageDescription -> LocalBuildInfo -> (Library -> ComponentLocalBuildInfo -> IO ()) -> IO ()

Perform the action on each enabled library in the package description with the ComponentLocalBuildInfo.

withExeLBI :: PackageDescription -> LocalBuildInfo -> (Executable -> ComponentLocalBuildInfo -> IO ()) -> IO ()

Perform the action on each enabled Executable in the package description. Extended version of withExe that also gives corresponding build info.

withBenchLBI :: PackageDescription -> LocalBuildInfo -> (Benchmark -> ComponentLocalBuildInfo -> IO ()) -> IO ()

Perform the action on each enabled Benchmark in the package description.

withTestLBI :: PackageDescription -> LocalBuildInfo -> (TestSuite -> ComponentLocalBuildInfo -> IO ()) -> IO ()

enabledTestLBIs :: PackageDescription -> LocalBuildInfo -> [(TestSuite, ComponentLocalBuildInfo)]

enabledBenchLBIs :: PackageDescription -> LocalBuildInfo -> [(Benchmark, ComponentLocalBuildInfo)]

Installation directories

module Distribution.Simple.InstallDirs

absoluteInstallDirs :: PackageDescription -> LocalBuildInfo -> CopyDest -> InstallDirs FilePath

Backwards compatibility function which computes the InstallDirs assuming that $libname points to the public library (or some fake package identifier if there is no public library.) IF AT ALL POSSIBLE, please use absoluteComponentInstallDirs instead.

prefixRelativeInstallDirs :: PackageId -> LocalBuildInfo -> InstallDirs (Maybe FilePath)

Backwards compatibility function which computes the InstallDirs assuming that $libname points to the public library (or some fake package identifier if there is no public library.) IF AT ALL POSSIBLE, please use prefixRelativeComponentInstallDirs instead.

absoluteComponentInstallDirs :: PackageDescription -> LocalBuildInfo -> UnitId -> CopyDest -> InstallDirs FilePath

See absoluteInstallDirs.

prefixRelativeComponentInstallDirs :: PackageId -> LocalBuildInfo -> UnitId -> InstallDirs (Maybe FilePath)

See prefixRelativeInstallDirs

substPathTemplate :: PackageId -> LocalBuildInfo -> UnitId -> PathTemplate -> FilePath