Distribution.Simple.LocalBuildInfo

Description

Once a package has been configured we have resolved conditionals and dependencies, configured the compiler and other needed external programs. The LocalBuildInfo is used to hold all this information. It holds the install dirs, the compiler, the exact package dependencies, the configured programs, the package database to use and a bunch of miscellaneous configure flags. It gets saved and reloaded from a file (dist/setup-config). It gets passed in to very many subsequent build actions.

Documentation

data LocalBuildInfo

Data cached after configuration step. See also ConfigFlags.

Constructors

LocalBuildInfo

Fields

configFlags :: ConfigFlags

Options passed to the configuration step. Needed to re-run configuration when .cabal is out of date

flagAssignment :: FlagAssignment

The final set of flags which were picked for this package

componentEnabledSpec :: ComponentRequestedSpec

What components were enabled during configuration, and why.

extraConfigArgs :: [String]

Extra args on the command line for the configuration step. Needed to re-run configuration when .cabal is out of date

installDirTemplates :: InstallDirTemplates

The installation directories for the various different kinds of files TODO: inplaceDirTemplates :: InstallDirs FilePath

compiler :: Compiler

The compiler we're building with

hostPlatform :: Platform

The platform we're building for

buildDir :: FilePath

Where to build the package.

cabalFilePath :: Maybe FilePath

Path to the cabal file, if given during configuration.

componentGraph :: Graph ComponentLocalBuildInfo

All the components to build, ordered by topological sort, and with their INTERNAL dependencies over the intrapackage dependency graph. TODO: this is assumed to be short; otherwise we want some sort of ordered map.

componentNameMap :: Map ComponentName [ComponentLocalBuildInfo]

A map from component name to all matching components. These coincide with componentGraph

installedPkgs :: InstalledPackageIndex

All the info about the installed packages that the current package depends on (directly or indirectly). The copy saved on disk does NOT include internal dependencies (because we just don't have enough information at this point to have an InstalledPackageInfo for an internal dep), but we will often update it with the internal dependencies; see for example build. (This admonition doesn't apply for per-component builds.)

pkgDescrFile :: Maybe FilePath

the filename containing the .cabal file, if available

localPkgDescr :: PackageDescription

WARNING WARNING WARNING Be VERY careful about using this function; we haven't deprecated it but using it could introduce subtle bugs related to HookedBuildInfo.

In principle, this is supposed to contain the resolved package description, that does not contain any conditionals. However, it MAY NOT contain the description with a HookedBuildInfo applied to it; see HookedBuildInfo for the whole sordid saga. As much as possible, Cabal library should avoid using this parameter.

withPrograms :: ProgramDb

Location and args for all programs

withPackageDB :: PackageDBStack

What package database to use, global/user

withVanillaLib :: Bool

Whether to build normal libs.

withProfLib :: Bool

Whether to build profiling versions of libs.

withSharedLib :: Bool

Whether to build shared versions of libs.

withStaticLib :: Bool

Whether to build static versions of libs (with all other libs rolled in)

withDynExe :: Bool

Whether to link executables dynamically

withFullyStaticExe :: Bool

Whether to link executables fully statically

withProfExe :: Bool

Whether to build executables for profiling.

withProfLibDetail :: ProfDetailLevel

Level of automatic profile detail.

withProfExeDetail :: ProfDetailLevel

Level of automatic profile detail.

withOptimization :: OptimisationLevel

Whether to build with optimization (if available).

withDebugInfo :: DebugInfoLevel

Whether to emit debug info (if available).

withGHCiLib :: Bool

Whether to build libs suitable for use with GHCi.

splitSections :: Bool

Use -split-sections with GHC, if available

splitObjs :: Bool

Use -split-objs with GHC, if available

stripExes :: Bool

Whether to strip executables during install

stripLibs :: Bool

Whether to strip libraries during install

exeCoverage :: Bool

Whether to enable executable program coverage

libCoverage :: Bool

Whether to enable library program coverage

progPrefix :: PathTemplate

Prefix to be prepended to installed executables

progSuffix :: PathTemplate

Suffix to be appended to installed executables

relocatable :: Bool

localComponentId :: LocalBuildInfo -> ComponentId

Extract the ComponentId from the public library component of a LocalBuildInfo if it exists, or make a fake component ID based on the package ID.

localUnitId :: LocalBuildInfo -> UnitId

Extract the UnitId from the library component of a LocalBuildInfo if it exists, or make a fake unit ID based on the package ID.

localCompatPackageKey :: LocalBuildInfo -> String

Extract the compatibility package key from the public library component of a LocalBuildInfo if it exists, or make a fake package key based on the package ID.

Buildable package components

data Component

Constructors

CLib Library

CFLib ForeignLib

CExe Executable

CTest TestSuite

CBench Benchmark

data ComponentName

Constructors

CLibName LibraryName

CNotLibName NotLibComponentName

Bundled Patterns

pattern CBenchName :: UnqualComponentName -> ComponentName

pattern CTestName :: UnqualComponentName -> ComponentName

pattern CExeName :: UnqualComponentName -> ComponentName

pattern CFLibName :: UnqualComponentName -> ComponentName

data LibraryName

Constructors

LMainLibName

LSubLibName UnqualComponentName

defaultLibName :: LibraryName

showComponentName :: ComponentName -> String

componentNameString :: ComponentName -> Maybe UnqualComponentName

This gets the underlying unqualified component name. In fact, it is guaranteed to uniquely identify a component, returning Nothing if the ComponentName was for the public library.

data ComponentLocalBuildInfo

The first five fields are common across all algebraic variants.

Constructors

LibComponentLocalBuildInfo

Fields

componentLocalName :: ComponentName

It would be very convenient to store the literal Library here, but if we do that, it will get serialized (via the Binary) instance twice. So instead we just provide the ComponentName, which can be used to find the Component in the PackageDescription. NB: eventually, this will NOT uniquely identify the ComponentLocalBuildInfo.

componentComponentId :: ComponentId

The computed ComponentId of this component.

componentUnitId :: UnitId

The computed UnitId which uniquely identifies this component. Might be hashed.

componentIsIndefinite_ :: Bool

Is this an indefinite component (i.e. has unfilled holes)?

componentInstantiatedWith :: [(ModuleName, OpenModule)]

How the component was instantiated

componentPackageDeps :: [(UnitId, MungedPackageId)]

Resolved internal and external package dependencies for this component. The BuildInfo specifies a set of build dependencies that must be satisfied in terms of version ranges. This field fixes those dependencies to the specific versions available on this machine for this compiler.

componentIncludes :: [(OpenUnitId, ModuleRenaming)]

The set of packages that are brought into scope during compilation, including a ModuleRenaming which may used to hide or rename modules. This is what gets translated into -package-id arguments. This is a modernized version of componentPackageDeps, which is kept around for BC purposes.

componentExeDeps :: [UnitId]

componentInternalDeps :: [UnitId]

The internal dependencies which induce a graph on the ComponentLocalBuildInfo of this package. This does NOT coincide with componentPackageDeps because it ALSO records 'build-tool' dependencies on executables. Maybe one day cabal-install will also handle these correctly too!

componentCompatPackageKey :: String

Compatibility "package key" that we pass to older versions of GHC.

componentCompatPackageName :: MungedPackageName

Compatibility "package name" that we register this component as.

componentExposedModules :: [ExposedModule]

A list of exposed modules (either defined in this component, or reexported from another component.)

componentIsPublic :: Bool

Convenience field, specifying whether or not this is the "public library" that has the same name as the package.

FLibComponentLocalBuildInfo

Fields

componentLocalName :: ComponentName

It would be very convenient to store the literal Library here, but if we do that, it will get serialized (via the Binary) instance twice. So instead we just provide the ComponentName, which can be used to find the Component in the PackageDescription. NB: eventually, this will NOT uniquely identify the ComponentLocalBuildInfo.

componentComponentId :: ComponentId

The computed ComponentId of this component.

componentUnitId :: UnitId

The computed UnitId which uniquely identifies this component. Might be hashed.

componentPackageDeps :: [(UnitId, MungedPackageId)]

Resolved internal and external package dependencies for this component. The BuildInfo specifies a set of build dependencies that must be satisfied in terms of version ranges. This field fixes those dependencies to the specific versions available on this machine for this compiler.

componentIncludes :: [(OpenUnitId, ModuleRenaming)]

The set of packages that are brought into scope during compilation, including a ModuleRenaming which may used to hide or rename modules. This is what gets translated into -package-id arguments. This is a modernized version of componentPackageDeps, which is kept around for BC purposes.

componentExeDeps :: [UnitId]

componentInternalDeps :: [UnitId]

The internal dependencies which induce a graph on the ComponentLocalBuildInfo of this package. This does NOT coincide with componentPackageDeps because it ALSO records 'build-tool' dependencies on executables. Maybe one day cabal-install will also handle these correctly too!

ExeComponentLocalBuildInfo

Fields

componentLocalName :: ComponentName

It would be very convenient to store the literal Library here, but if we do that, it will get serialized (via the Binary) instance twice. So instead we just provide the ComponentName, which can be used to find the Component in the PackageDescription. NB: eventually, this will NOT uniquely identify the ComponentLocalBuildInfo.

componentComponentId :: ComponentId

The computed ComponentId of this component.

componentUnitId :: UnitId

The computed UnitId which uniquely identifies this component. Might be hashed.

componentPackageDeps :: [(UnitId, MungedPackageId)]

Resolved internal and external package dependencies for this component. The BuildInfo specifies a set of build dependencies that must be satisfied in terms of version ranges. This field fixes those dependencies to the specific versions available on this machine for this compiler.

componentIncludes :: [(OpenUnitId, ModuleRenaming)]

The set of packages that are brought into scope during compilation, including a ModuleRenaming which may used to hide or rename modules. This is what gets translated into -package-id arguments. This is a modernized version of componentPackageDeps, which is kept around for BC purposes.

componentExeDeps :: [UnitId]

componentInternalDeps :: [UnitId]

The internal dependencies which induce a graph on the ComponentLocalBuildInfo of this package. This does NOT coincide with componentPackageDeps because it ALSO records 'build-tool' dependencies on executables. Maybe one day cabal-install will also handle these correctly too!

TestComponentLocalBuildInfo

Fields

componentLocalName :: ComponentName

It would be very convenient to store the literal Library here, but if we do that, it will get serialized (via the Binary) instance twice. So instead we just provide the ComponentName, which can be used to find the Component in the PackageDescription. NB: eventually, this will NOT uniquely identify the ComponentLocalBuildInfo.

componentComponentId :: ComponentId

The computed ComponentId of this component.

componentUnitId :: UnitId

The computed UnitId which uniquely identifies this component. Might be hashed.

componentPackageDeps :: [(UnitId, MungedPackageId)]

Resolved internal and external package dependencies for this component. The BuildInfo specifies a set of build dependencies that must be satisfied in terms of version ranges. This field fixes those dependencies to the specific versions available on this machine for this compiler.

componentIncludes :: [(OpenUnitId, ModuleRenaming)]

The set of packages that are brought into scope during compilation, including a ModuleRenaming which may used to hide or rename modules. This is what gets translated into -package-id arguments. This is a modernized version of componentPackageDeps, which is kept around for BC purposes.

componentExeDeps :: [UnitId]

componentInternalDeps :: [UnitId]

The internal dependencies which induce a graph on the ComponentLocalBuildInfo of this package. This does NOT coincide with componentPackageDeps because it ALSO records 'build-tool' dependencies on executables. Maybe one day cabal-install will also handle these correctly too!

BenchComponentLocalBuildInfo

Fields

componentLocalName :: ComponentName

It would be very convenient to store the literal Library here, but if we do that, it will get serialized (via the Binary) instance twice. So instead we just provide the ComponentName, which can be used to find the Component in the PackageDescription. NB: eventually, this will NOT uniquely identify the ComponentLocalBuildInfo.

componentComponentId :: ComponentId

The computed ComponentId of this component.

componentUnitId :: UnitId

The computed UnitId which uniquely identifies this component. Might be hashed.

componentPackageDeps :: [(UnitId, MungedPackageId)]

Resolved internal and external package dependencies for this component. The BuildInfo specifies a set of build dependencies that must be satisfied in terms of version ranges. This field fixes those dependencies to the specific versions available on this machine for this compiler.

componentIncludes :: [(OpenUnitId, ModuleRenaming)]

The set of packages that are brought into scope during compilation, including a ModuleRenaming which may used to hide or rename modules. This is what gets translated into -package-id arguments. This is a modernized version of componentPackageDeps, which is kept around for BC purposes.

componentExeDeps :: [UnitId]

componentInternalDeps :: [UnitId]

The internal dependencies which induce a graph on the ComponentLocalBuildInfo of this package. This does NOT coincide with componentPackageDeps because it ALSO records 'build-tool' dependencies on executables. Maybe one day cabal-install will also handle these correctly too!

componentBuildDir :: LocalBuildInfo -> ComponentLocalBuildInfo -> FilePath

foldComponent :: (Library -> a) -> (ForeignLib -> a) -> (Executable -> a) -> (TestSuite -> a) -> (Benchmark -> a) -> Component -> a

componentName :: Component -> ComponentName

componentBuildInfo :: Component -> BuildInfo

componentBuildable :: Component -> Bool

Is a component buildable (i.e., not marked with buildable: False)? See also this note in Distribution.Types.ComponentRequestedSpec.

Since: Cabal-syntax-2.0.0.2

pkgComponents :: PackageDescription -> [Component]

All the components in the package.

pkgBuildableComponents :: PackageDescription -> [Component]

A list of all components in the package that are buildable, i.e., were not marked with buildable: False. This does NOT indicate if we are actually going to build the component, see enabledComponents instead.

Since: Cabal-syntax-2.0.0.2

lookupComponent :: PackageDescription -> ComponentName -> Maybe Component

getComponent :: PackageDescription -> ComponentName -> Component

allComponentsInBuildOrder :: LocalBuildInfo -> [ComponentLocalBuildInfo]

depLibraryPaths

Arguments

:: Bool
Building for inplace?

-> Bool
Generate prefix-relative library paths

-> LocalBuildInfo

-> ComponentLocalBuildInfo
Component that is being built

-> IO [FilePath]

Determine the directories containing the dynamic libraries of the transitive dependencies of the component we are building.

When wanted, and possible, returns paths relative to the installDirs prefix

allLibModules :: Library -> ComponentLocalBuildInfo -> [ModuleName]

Get all module names that needed to be built by GHC; i.e., all of these ModuleNames have interface files associated with them that need to be installed.

withAllComponentsInBuildOrder :: PackageDescription -> LocalBuildInfo -> (Component -> ComponentLocalBuildInfo -> IO ()) -> IO ()

Perform the action on each buildable Library or Executable (Component) in the PackageDescription, subject to the build order specified by the compBuildOrder field of the given LocalBuildInfo

withLibLBI :: PackageDescription -> LocalBuildInfo -> (Library -> ComponentLocalBuildInfo -> IO ()) -> IO ()

Perform the action on each enabled library in the package description with the ComponentLocalBuildInfo.

withExeLBI :: PackageDescription -> LocalBuildInfo -> (Executable -> ComponentLocalBuildInfo -> IO ()) -> IO ()

Perform the action on each enabled Executable in the package description. Extended version of withExe that also gives corresponding build info.

withBenchLBI :: PackageDescription -> LocalBuildInfo -> (Benchmark -> ComponentLocalBuildInfo -> IO ()) -> IO ()

Perform the action on each enabled Benchmark in the package description.

withTestLBI :: PackageDescription -> LocalBuildInfo -> (TestSuite -> ComponentLocalBuildInfo -> IO ()) -> IO ()

enabledTestLBIs :: PackageDescription -> LocalBuildInfo -> [(TestSuite, ComponentLocalBuildInfo)]

enabledBenchLBIs :: PackageDescription -> LocalBuildInfo -> [(Benchmark, ComponentLocalBuildInfo)]

Installation directories

data InstallDirs dir

The directories where we will install files for packages.

We have several different directories for different types of files since many systems have conventions whereby different types of files in a package are installed in different directories. This is particularly the case on Unix style systems.

Constructors

InstallDirs

Fields

prefix :: dir

bindir :: dir

libdir :: dir

libsubdir :: dir

dynlibdir :: dir

flibdir :: dir

foreign libraries

libexecdir :: dir

libexecsubdir :: dir

includedir :: dir

datadir :: dir

datasubdir :: dir

docdir :: dir

mandir :: dir

htmldir :: dir

haddockdir :: dir

sysconfdir :: dir

type InstallDirTemplates = InstallDirs PathTemplate

The installation directories in terms of PathTemplates that contain variables.

The defaults for most of the directories are relative to each other, in particular they are all relative to a single prefix. This makes it convenient for the user to override the default installation directory by only having to specify --prefix=... rather than overriding each individually. This is done by allowing $-style variables in the dirs. These are expanded by textual substitution (see substPathTemplate).

A few of these installation directories are split into two components, the dir and subdir. The full installation path is formed by combining the two together with /. The reason for this is compatibility with other Unix build systems which also support --libdir and --datadir. We would like users to be able to configure --libdir=/usr/lib64 for example but because by default we want to support installing multiple versions of packages and building the same package for multiple compilers we append the libsubdir to get: /usr/lib64/$libname/$compiler.

An additional complication is the need to support relocatable packages on systems which support such things, like Windows.

data CopyDest

The location prefix for the copy command.

Constructors

NoCopyDest

CopyTo FilePath

CopyToDb FilePath
when using the ${pkgroot} as prefix. The CopyToDb will adjust the paths to be relative to the provided package database when copying / installing.

data PathTemplate

An abstract path, possibly containing variables that need to be substituted for to get a real FilePath.

data PathTemplateVariable

Constructors

PrefixVar
The $prefix path variable

BindirVar
The $bindir path variable

LibdirVar
The $libdir path variable

LibsubdirVar
The $libsubdir path variable

DynlibdirVar
The $dynlibdir path variable

DatadirVar
The $datadir path variable

DatasubdirVar
The $datasubdir path variable

DocdirVar
The $docdir path variable

HtmldirVar
The $htmldir path variable

PkgNameVar
The $pkg package name path variable

PkgVerVar
The $version package version path variable

PkgIdVar
The $pkgid package Id path variable, eg foo-1.0

LibNameVar
The $libname path variable

CompilerVar
The compiler name and version, eg ghc-6.6.1

OSVar
The operating system name, eg windows or linux

ArchVar
The CPU architecture name, eg i386 or x86_64

AbiVar
The compiler's ABI identifier,

AbiTagVar
The optional ABI tag for the compiler

ExecutableNameVar
The executable name; used in shell wrappers

TestSuiteNameVar
The name of the test suite being run

TestSuiteResultVar
The result of the test suite being run, eg pass, fail, or error.

BenchmarkNameVar
The name of the benchmark being run

type PathTemplateEnv = [(PathTemplateVariable, PathTemplate)]

defaultInstallDirs :: CompilerFlavor -> Bool -> Bool -> IO InstallDirTemplates

defaultInstallDirs' :: Bool -> CompilerFlavor -> Bool -> Bool -> IO InstallDirTemplates

combineInstallDirs :: (a -> b -> c) -> InstallDirs a -> InstallDirs b -> InstallDirs c

substituteInstallDirTemplates :: PathTemplateEnv -> InstallDirTemplates -> InstallDirTemplates

Substitute the install dir templates into each other.

To prevent cyclic substitutions, only some variables are allowed in particular dir templates. If out of scope vars are present, they are not substituted for. Checking for any remaining unsubstituted vars can be done as a subsequent operation.

The reason it is done this way is so that in prefixRelativeInstallDirs we can replace prefix with the PrefixVar and get resulting PathTemplates that still have the PrefixVar in them. Doing this makes it each to check which paths are relative to the $prefix.

toPathTemplate :: FilePath -> PathTemplate

Convert a FilePath to a PathTemplate including any template vars.

fromPathTemplate :: PathTemplate -> FilePath

Convert back to a path, any remaining vars are included

combinePathTemplate :: PathTemplate -> PathTemplate -> PathTemplate

initialPathTemplateEnv :: PackageIdentifier -> UnitId -> CompilerInfo -> Platform -> PathTemplateEnv

The initial environment has all the static stuff but no paths

platformTemplateEnv :: Platform -> PathTemplateEnv

compilerTemplateEnv :: CompilerInfo -> PathTemplateEnv

packageTemplateEnv :: PackageIdentifier -> UnitId -> PathTemplateEnv

abiTemplateEnv :: CompilerInfo -> Platform -> PathTemplateEnv

installDirsTemplateEnv :: InstallDirs PathTemplate -> PathTemplateEnv

absoluteInstallDirs :: PackageDescription -> LocalBuildInfo -> CopyDest -> InstallDirs FilePath

Backwards compatibility function which computes the InstallDirs assuming that $libname points to the public library (or some fake package identifier if there is no public library.) IF AT ALL POSSIBLE, please use absoluteComponentInstallDirs instead.

prefixRelativeInstallDirs :: PackageId -> LocalBuildInfo -> InstallDirs (Maybe FilePath)

Backwards compatibility function which computes the InstallDirs assuming that $libname points to the public library (or some fake package identifier if there is no public library.) IF AT ALL POSSIBLE, please use prefixRelativeComponentInstallDirs instead.

absoluteInstallCommandDirs :: PackageDescription -> LocalBuildInfo -> UnitId -> CopyDest -> InstallDirs FilePath

absoluteComponentInstallDirs :: PackageDescription -> LocalBuildInfo -> UnitId -> CopyDest -> InstallDirs FilePath

See absoluteInstallDirs.

prefixRelativeComponentInstallDirs :: PackageId -> LocalBuildInfo -> UnitId -> InstallDirs (Maybe FilePath)

See prefixRelativeInstallDirs

substPathTemplate :: PackageId -> LocalBuildInfo -> UnitId -> PathTemplate -> FilePath