Distribution.Simple.PackageIndex

Description

An index of packages.

Package index data type

data PackageIndex

The collection of information about packages from one or more PackageDBs.

Packages are uniquely identified in by their InstalledPackageId, they can also be effeciently looked up by package name or by name and version.

Creating an index

fromList :: [InstalledPackageInfo] -> PackageIndex

Build an index out of a bunch of packages.

If there are duplicates by InstalledPackageId then later ones mask earlier ones.

Updates

merge :: PackageIndex -> PackageIndex -> PackageIndex

Merge two indexes.

Packages from the second mask packages from the first if they have the exact same InstalledPackageId.

For packages with the same source PackageId, packages from the second are "preferred" over those from the first. Being preferred means they are top result when we do a lookup by source PackageId. This is the mechanism we use to prefer user packages over global packages.

insert :: InstalledPackageInfo -> PackageIndex -> PackageIndex

Inserts a single package into the index.

This is equivalent to (but slightly quicker than) using mappend or merge with a singleton index.

deleteInstalledPackageId :: InstalledPackageId -> PackageIndex -> PackageIndex

Removes a single installed package from the index.

deleteSourcePackageId :: PackageId -> PackageIndex -> PackageIndex

Removes all packages with this source PackageId from the index.

deletePackageName :: PackageName -> PackageIndex -> PackageIndex

Removes all packages with this (case-sensitive) name from the index.

Queries

Precise lookups

lookupInstalledPackageId :: PackageIndex -> InstalledPackageId -> Maybe InstalledPackageInfo

Does a lookup by source package id (name & version).

Since multiple package DBs mask each other by InstalledPackageId, then we get back at most one package.

lookupSourcePackageId :: PackageIndex -> PackageId -> [InstalledPackageInfo]

Does a lookup by source package id (name & version).

There can be multiple installed packages with the same source PackageId but different InstalledPackageId. They are returned in order of preference, with the most preferred first.

lookupPackageName :: PackageIndex -> PackageName -> [(Version, [InstalledPackageInfo])]

Does a lookup by source package name.

lookupDependency :: PackageIndex -> Dependency -> [(Version, [InstalledPackageInfo])]

Does a lookup by source package name and a range of versions.

We get back any number of versions of the specified package name, all satisfying the version range constraint.

Case-insensitive searches

searchByName :: PackageIndex -> String -> SearchResult [InstalledPackageInfo]

Does a case-insensitive search by package name.

If there is only one package that compares case-insentiviely to this name then the search is unambiguous and we get back all versions of that package. If several match case-insentiviely but one matches exactly then it is also unambiguous.

If however several match case-insentiviely and none match exactly then we have an ambiguous result, and we get back all the versions of all the packages. The list of ambiguous results is split by exact package name. So it is a non-empty list of non-empty lists.

data SearchResult a

Constructors

None

Unambiguous a

Ambiguous [a]

searchByNameSubstring :: PackageIndex -> String -> [InstalledPackageInfo]

Does a case-insensitive substring search by package name.

That is, all packages that contain the given string in their name.

Bulk queries

allPackages :: PackageIndex -> [InstalledPackageInfo]

Get all the packages from the index.

allPackagesByName :: PackageIndex -> [(PackageName, [InstalledPackageInfo])]

Get all the packages from the index.

They are grouped by package name (case-sensitively).

allPackagesBySourcePackageId :: PackageIndex -> [(PackageId, [InstalledPackageInfo])]

Get all the packages from the index.

They are grouped by source package id (package name and version).

Special queries

brokenPackages :: PackageIndex -> [(InstalledPackageInfo, [InstalledPackageId])]

All packages that have immediate dependencies that are not in the index.

Returns such packages along with the dependencies that they're missing.

dependencyClosure :: PackageIndex -> [InstalledPackageId] -> Either PackageIndex [(InstalledPackageInfo, [InstalledPackageId])]

Tries to take the transitive closure of the package dependencies.

If the transitive closure is complete then it returns that subset of the index. Otherwise it returns the broken packages as in brokenPackages.

Note that if the result is Right [] it is because at least one of the original given PackageIds do not occur in the index.

reverseDependencyClosure :: PackageIndex -> [InstalledPackageId] -> [InstalledPackageInfo]

Takes the transitive closure of the packages reverse dependencies.

The given PackageIds must be in the index.

topologicalOrder :: PackageIndex -> [InstalledPackageInfo]

reverseTopologicalOrder :: PackageIndex -> [InstalledPackageInfo]

dependencyInconsistencies :: PackageIndex -> [(PackageName, [(PackageId, Version)])]

Given a package index where we assume we want to use all the packages (use dependencyClosure if you need to get such a index subset) find out if the dependencies within it use consistent versions of each package. Return all cases where multiple packages depend on different versions of some other package.

Each element in the result is a package name along with the packages that depend on it and the versions they require. These are guaranteed to be distinct.

dependencyCycles :: PackageIndex -> [[InstalledPackageInfo]]

Find if there are any cycles in the dependency graph. If there are no cycles the result is [].

This actually computes the strongly connected components. So it gives us a list of groups of packages where within each group they all depend on each other, directly or indirectly.

dependencyGraph :: PackageIndex -> (Graph, Vertex -> InstalledPackageInfo, InstalledPackageId -> Maybe Vertex)

Builds a graph of the package dependencies.

Dependencies on other packages that are not in the index are discarded. You can check if there are any such dependencies with brokenPackages.

moduleNameIndex :: PackageIndex -> Map ModuleName [InstalledPackageInfo]