Distribution.Simple.Program.Internal

Description

Internal utilities used by Distribution.Simple.Program.*.

Documentation

stripExtractVersion :: String -> String

Extract the version number from the output of 'strip --version'.

Invoking strip --version gives very inconsistent results. We ignore everything in parentheses (see #2497), look for the first word that starts with a number, and try parsing out the first two components of it. Non-GNU strip doesn't appear to have a version flag.