Distribution.Simple.Program.Script

Description

This module provides an library interface to the hc-pkg program. Currently only GHC and LHC have hc-pkg programs.

Documentation

invocationAsSystemScript :: OS -> ProgramInvocation -> String

Generate a system script, either POSIX shell script or Windows batch file as appropriate for the given system.

invocationAsShellScript :: ProgramInvocation -> String

Generate a POSIX shell script that invokes a program.

invocationAsBatchFile :: ProgramInvocation -> String

Generate a Windows batch file that invokes a program.