Distribution.Simple.Register

Description

This module deals with registering and unregistering packages. There are a couple ways it can do this, one is to do it directly. Another is to generate a script that can be run later to do it. The idea here being that the user is shielded from the details of what command to use for package registration for a particular compiler. In practice this aspect was not especially popular so we also provide a way to simply generate the package registration file which then must be manually passed to ghc-pkg. It is possible to generate registration information for where the package is to be installed, or alternatively to register the package in place in the build tree. The latter is occasionally handy, and will become more important when we try to build multi-package systems.

This module does not delegate anything to the per-compiler modules but just mixes it all in in this module, which is rather unsatisfactory. The script generation and the unregister feature are not well used or tested.

Documentation

register

Arguments

:: PackageDescription

-> LocalBuildInfo

-> RegisterFlags
Install in the user's database?; verbose

-> IO ()

unregister :: PackageDescription -> LocalBuildInfo -> RegisterFlags -> IO ()

initPackageDB :: Verbosity -> Compiler -> ProgramConfiguration -> FilePath -> IO ()

doesPackageDBExist :: FilePath -> IO Bool

createPackageDB :: Verbosity -> Compiler -> ProgramConfiguration -> Bool -> FilePath -> IO ()

Create an empty package DB at the specified location.

deletePackageDB :: FilePath -> IO ()

invokeHcPkg :: Verbosity -> Compiler -> ProgramConfiguration -> PackageDBStack -> [String] -> IO ()

Run hc-pkg using a given package DB stack, directly forwarding the provided command-line arguments to it.

registerPackage :: Verbosity -> Compiler -> ProgramConfiguration -> Bool -> PackageDBStack -> InstalledPackageInfo -> IO ()

generateRegistrationInfo :: Verbosity -> PackageDescription -> Library -> LocalBuildInfo -> ComponentLocalBuildInfo -> Bool -> Bool -> FilePath -> PackageDB -> IO InstalledPackageInfo

inplaceInstalledPackageInfo

Arguments

:: FilePath
top of the build tree

-> FilePath
location of the dist tree

-> PackageDescription

-> AbiHash

-> Library

-> LocalBuildInfo

-> ComponentLocalBuildInfo

-> InstalledPackageInfo

Construct InstalledPackageInfo for a library that is in place in the build tree.

This function knows about the layout of in place packages.

absoluteInstalledPackageInfo :: PackageDescription -> AbiHash -> Library -> LocalBuildInfo -> ComponentLocalBuildInfo -> InstalledPackageInfo

Construct InstalledPackageInfo for the final install location of a library package.

This function knows about the layout of installed packages.

generalInstalledPackageInfo

Arguments

:: ([FilePath] -> [FilePath])
Translate relative include dir paths to absolute paths.

-> PackageDescription

-> AbiHash

-> Library

-> LocalBuildInfo

-> ComponentLocalBuildInfo

-> InstallDirs FilePath

-> InstalledPackageInfo

Construct InstalledPackageInfo for a library in a package, given a set of installation directories.