Distribution.Simple.Setup

Description

This is a big module, but not very complicated. The code is very regular and repetitive. It defines the command line interface for all the Cabal commands. For each command (like configure, build etc) it defines a type that holds all the flags, the default set of flags and a CommandUI that maps command line flags to and from the corresponding flags type.

All the flags types are instances of Monoid, see http://www.haskell.org/pipermail/cabal-devel/2007-December/001509.html for an explanation.

The types defined here get used in the front end and especially in cabal-install which has to do quite a bit of manipulating sets of command line flags.

This is actually relatively nice, it works quite well. The main change it needs is to unify it with the code for managing sets of fields that can be read and written from files. This would allow us to save configure flags in config files.

Documentation

data GlobalFlags

Flags that apply at the top level, not to any sub-command.

Constructors

GlobalFlags

Fields

globalVersion :: Flag Bool

globalNumericVersion :: Flag Bool

emptyGlobalFlags :: GlobalFlags

defaultGlobalFlags :: GlobalFlags

globalCommand :: [Command action] -> CommandUI GlobalFlags

data ConfigFlags

Flags to configure command.

IMPORTANT: every time a new flag is added, filterConfigureFlags should be updated. IMPORTANT: every time a new flag is added, it should be added to the Eq instance

Constructors

ConfigFlags

Fields

configArgs :: [String]

configPrograms_ :: Option' (Last' ProgramDb)

All programs that cabal may run

configProgramPaths :: [(String, FilePath)]

user specified programs paths

configProgramArgs :: [(String, [String])]

user specified programs args

configProgramPathExtra :: NubList FilePath

Extend the $PATH

configHcFlavor :: Flag CompilerFlavor

The "flavor" of the compiler, e.g. GHC.

configHcPath :: Flag FilePath

given compiler location

configHcPkg :: Flag FilePath

given hc-pkg location

configVanillaLib :: Flag Bool

Enable vanilla library

configProfLib :: Flag Bool

Enable profiling in the library

configSharedLib :: Flag Bool

Build shared library

configStaticLib :: Flag Bool

Build static library

configDynExe :: Flag Bool

Enable dynamic linking of the executables.

configFullyStaticExe :: Flag Bool

Enable fully static linking of the executables.

configProfExe :: Flag Bool

Enable profiling in the executables.

configProf :: Flag Bool

Enable profiling in the library and executables.

configProfDetail :: Flag ProfDetailLevel

Profiling detail level in the library and executables.

configProfLibDetail :: Flag ProfDetailLevel

Profiling detail level in the library

configConfigureArgs :: [String]

Extra arguments to configure

configOptimization :: Flag OptimisationLevel

Enable optimization.

configProgPrefix :: Flag PathTemplate

Installed executable prefix.

configProgSuffix :: Flag PathTemplate

Installed executable suffix.

configInstallDirs :: InstallDirs (Flag PathTemplate)

Installation paths

configScratchDir :: Flag FilePath

configExtraLibDirs :: [FilePath]

path to search for extra libraries

configExtraLibDirsStatic :: [FilePath]

path to search for extra libraries when linking fully static executables

configExtraFrameworkDirs :: [FilePath]

path to search for extra frameworks (OS X only)

configExtraIncludeDirs :: [FilePath]

path to search for header files

configIPID :: Flag String

explicit IPID to be used

configCID :: Flag ComponentId

explicit CID to be used

configDeterministic :: Flag Bool

be as deterministic as possible (e.g., invariant over GHC, database, etc). Used by the test suite

configDistPref :: Flag FilePath

"dist" prefix

configCabalFilePath :: Flag FilePath

Cabal file to use

configVerbosity :: Flag Verbosity

verbosity level

configUserInstall :: Flag Bool

The --user/--global flag

configPackageDBs :: [Maybe PackageDB]

Which package DBs to use

configGHCiLib :: Flag Bool

Enable compiling library for GHCi

configSplitSections :: Flag Bool

Enable -split-sections with GHC

configSplitObjs :: Flag Bool

Enable -split-objs with GHC

configStripExes :: Flag Bool

Enable executable stripping

configStripLibs :: Flag Bool

Enable library stripping

configConstraints :: [PackageVersionConstraint]

Additional constraints for dependencies.

configDependencies :: [GivenComponent]

The packages depended on.

configInstantiateWith :: [(ModuleName, Module)]

The requested Backpack instantiation. If empty, either this package does not use Backpack, or we just want to typecheck the indefinite package.

configConfigurationsFlags :: FlagAssignment

configTests :: Flag Bool

Enable test suite compilation

configBenchmarks :: Flag Bool

Enable benchmark compilation

configCoverage :: Flag Bool

Enable program coverage

configLibCoverage :: Flag Bool

Enable program coverage (deprecated)

configExactConfiguration :: Flag Bool

All direct dependencies and flags are provided on the command line by the user via the '--dependency' and '--flags' options.

configFlagError :: Flag String

Halt and show an error message indicating an error in flag assignment

configRelocatable :: Flag Bool

Enable relocatable package built

configDebugInfo :: Flag DebugInfoLevel

Emit debug info.

configDumpBuildInfo :: Flag DumpBuildInfo

Should we dump available build information on build? Dump build information to disk before attempting to build, tooling can parse these files and use them to compile the source files themselves.

configUseResponseFiles :: Flag Bool

Whether to use response files at all. They're used for such tools as haddock, or ld.

configAllowDependingOnPrivateLibs :: Flag Bool

Allow depending on private sublibraries. This is used by external tools (like cabal-install) so they can add multiple-public-libraries compatibility to older ghcs by checking visibility externally.

emptyConfigFlags :: ConfigFlags

defaultConfigFlags :: ProgramDb -> ConfigFlags

configureCommand :: ProgramDb -> CommandUI ConfigFlags

configPrograms :: WithCallStack (ConfigFlags -> ProgramDb)

More convenient version of configPrograms. Results in an error if internal invariant is violated.

configAbsolutePaths :: ConfigFlags -> IO ConfigFlags

readPackageDb :: String -> Maybe PackageDB

Parse a PackageDB stack entry

Since: Cabal-3.7.0.0

readPackageDbList :: String -> [Maybe PackageDB]

showPackageDb :: Maybe PackageDB -> String

Show a PackageDB stack entry

Since: Cabal-3.7.0.0

showPackageDbList :: [Maybe PackageDB] -> [String]

data CopyFlags

Flags to copy: (destdir, copy-prefix (backwards compat), verbosity)

Constructors

CopyFlags

Fields

copyDest :: Flag CopyDest

copyDistPref :: Flag FilePath

copyVerbosity :: Flag Verbosity

copyArgs :: [String]

copyCabalFilePath :: Flag FilePath

emptyCopyFlags :: CopyFlags

defaultCopyFlags :: CopyFlags

copyCommand :: CommandUI CopyFlags

data InstallFlags

Flags to install: (package db, verbosity)

Constructors

InstallFlags

Fields

installPackageDB :: Flag PackageDB

installDest :: Flag CopyDest

installDistPref :: Flag FilePath

installUseWrapper :: Flag Bool

installInPlace :: Flag Bool

installVerbosity :: Flag Verbosity

installCabalFilePath :: Flag FilePath

emptyInstallFlags :: InstallFlags

defaultInstallFlags :: InstallFlags

installCommand :: CommandUI InstallFlags

data HaddockTarget

When we build haddock documentation, there are two cases:

We build haddocks only for the current development version, intended for local use and not for distribution. In this case, we store the generated documentation in distdochtml/name.

We build haddocks for intended for uploading them to hackage. In this case, we need to follow the layout that hackage expects from documentation tarballs, and we might also want to use different flags than for development builds, so in this case we store the generated documentation in distdochtml/id-docs.

Constructors

ForHackage

ForDevelopment

data HaddockFlags

Constructors

HaddockFlags

Fields

haddockProgramPaths :: [(String, FilePath)]

haddockProgramArgs :: [(String, [String])]

haddockHoogle :: Flag Bool

haddockHtml :: Flag Bool

haddockHtmlLocation :: Flag String

haddockForHackage :: Flag HaddockTarget

haddockExecutables :: Flag Bool

haddockTestSuites :: Flag Bool

haddockBenchmarks :: Flag Bool

haddockForeignLibs :: Flag Bool

haddockInternal :: Flag Bool

haddockCss :: Flag FilePath

haddockLinkedSource :: Flag Bool

haddockQuickJump :: Flag Bool

haddockHscolourCss :: Flag FilePath

haddockContents :: Flag PathTemplate

haddockIndex :: Flag PathTemplate

haddockDistPref :: Flag FilePath

haddockKeepTempFiles :: Flag Bool

haddockVerbosity :: Flag Verbosity

haddockCabalFilePath :: Flag FilePath

haddockBaseUrl :: Flag String

haddockLib :: Flag String

haddockArgs :: [String]

emptyHaddockFlags :: HaddockFlags

defaultHaddockFlags :: HaddockFlags

haddockCommand :: CommandUI HaddockFlags

data Visibility

Governs whether modules from a given interface should be visible or hidden in the Haddock generated content page. We don't expose this functionality to the user, but simply use Visible for only local packages. Visibility of modules is available since haddock-2.26.1.

Constructors

Visible

Hidden

data HaddockProjectFlags

Constructors

HaddockProjectFlags

Fields

haddockProjectHackage :: Flag Bool

a shortcut option which builds documentation linked to hackage. It implies: * `--html-location='https://hackage.haskell.org/package/$prg-$version/docs' * `--quickjump` * `--gen-index` * `--gen-contents` * `--hyperlinked-source`

haddockProjectLocal :: Flag Bool

a shortcut option which builds self contained directory which contains all the documentation, it implies: * `--quickjump` * `--gen-index` * `--gen-contents` * `--hyperlinked-source`

And it will also pass `--base-url` option to haddock.

haddockProjectDir :: Flag String

output directory of combined haddocks, the default is './haddocks'

haddockProjectPrologue :: Flag String

haddockProjectGenIndex :: Flag Bool

haddockProjectGenContents :: Flag Bool

haddockProjectInterfaces :: Flag [(FilePath, Maybe FilePath, Maybe FilePath, Visibility)]

haddocksInterfaces is inferred by the haddocksAction; currently not exposed to the user.

haddockProjectProgramPaths :: [(String, FilePath)]

haddockProjectProgramArgs :: [(String, [String])]

haddockProjectHoogle :: Flag Bool

haddockProjectHtmlLocation :: Flag String

haddockProjectExecutables :: Flag Bool

haddockProjectTestSuites :: Flag Bool

haddockProjectBenchmarks :: Flag Bool

haddockProjectForeignLibs :: Flag Bool

haddockProjectInternal :: Flag Bool

haddockProjectCss :: Flag FilePath

haddockProjectLinkedSource :: Flag Bool

haddockProjectQuickJump :: Flag Bool

haddockProjectHscolourCss :: Flag FilePath

haddockProjectKeepTempFiles :: Flag Bool

haddockProjectVerbosity :: Flag Verbosity

haddockProjectLib :: Flag String

emptyHaddockProjectFlags :: HaddockProjectFlags

defaultHaddockProjectFlags :: HaddockProjectFlags

haddockProjectCommand :: CommandUI HaddockProjectFlags

data HscolourFlags

Constructors

HscolourFlags

Fields

hscolourCSS :: Flag FilePath

hscolourExecutables :: Flag Bool

hscolourTestSuites :: Flag Bool

hscolourBenchmarks :: Flag Bool

hscolourForeignLibs :: Flag Bool

hscolourDistPref :: Flag FilePath

hscolourVerbosity :: Flag Verbosity

hscolourCabalFilePath :: Flag FilePath

emptyHscolourFlags :: HscolourFlags

defaultHscolourFlags :: HscolourFlags

hscolourCommand :: CommandUI HscolourFlags

data BuildFlags

Constructors

BuildFlags

Fields

buildProgramPaths :: [(String, FilePath)]

buildProgramArgs :: [(String, [String])]

buildDistPref :: Flag FilePath

buildVerbosity :: Flag Verbosity

buildNumJobs :: Flag (Maybe Int)

buildArgs :: [String]

buildCabalFilePath :: Flag FilePath

emptyBuildFlags :: BuildFlags

defaultBuildFlags :: BuildFlags

buildCommand :: ProgramDb -> CommandUI BuildFlags

data DumpBuildInfo

Constructors

NoDumpBuildInfo

DumpBuildInfo

data ReplFlags

Constructors

ReplFlags

Fields

replProgramPaths :: [(String, FilePath)]

replProgramArgs :: [(String, [String])]

replDistPref :: Flag FilePath

replVerbosity :: Flag Verbosity

replReload :: Flag Bool

replReplOptions :: ReplOptions

defaultReplFlags :: ReplFlags

replCommand :: ProgramDb -> CommandUI ReplFlags

data ReplOptions

Constructors

ReplOptions

Fields

replOptionsFlags :: [String]

replOptionsNoLoad :: Flag Bool

data CleanFlags

Constructors

CleanFlags

Fields

cleanSaveConf :: Flag Bool

cleanDistPref :: Flag FilePath

cleanVerbosity :: Flag Verbosity

cleanCabalFilePath :: Flag FilePath

emptyCleanFlags :: CleanFlags

defaultCleanFlags :: CleanFlags

cleanCommand :: CommandUI CleanFlags

data RegisterFlags

Flags to register and unregister: (user package, gen-script, in-place, verbosity)

Constructors

RegisterFlags

Fields

regPackageDB :: Flag PackageDB

regGenScript :: Flag Bool

regGenPkgConf :: Flag (Maybe FilePath)

regInPlace :: Flag Bool

regDistPref :: Flag FilePath

regPrintId :: Flag Bool

regVerbosity :: Flag Verbosity

regArgs :: [String]

regCabalFilePath :: Flag FilePath

emptyRegisterFlags :: RegisterFlags

defaultRegisterFlags :: RegisterFlags

registerCommand :: CommandUI RegisterFlags

unregisterCommand :: CommandUI RegisterFlags

data SDistFlags

Flags to sdist: (snapshot, verbosity)

Constructors

SDistFlags

Fields

sDistSnapshot :: Flag Bool

sDistDirectory :: Flag FilePath

sDistDistPref :: Flag FilePath

sDistListSources :: Flag FilePath

sDistVerbosity :: Flag Verbosity

emptySDistFlags :: SDistFlags

defaultSDistFlags :: SDistFlags

sdistCommand :: CommandUI SDistFlags

data TestFlags

Constructors

TestFlags

Fields

testDistPref :: Flag FilePath

testVerbosity :: Flag Verbosity

testHumanLog :: Flag PathTemplate

testMachineLog :: Flag PathTemplate

testShowDetails :: Flag TestShowDetails

testKeepTix :: Flag Bool

testWrapper :: Flag FilePath

testFailWhenNoTestSuites :: Flag Bool

testOptions :: [PathTemplate]

emptyTestFlags :: TestFlags

defaultTestFlags :: TestFlags

testCommand :: CommandUI TestFlags

data TestShowDetails

Constructors

Never

Failures

Always

Streaming

Direct

data BenchmarkFlags

Constructors

BenchmarkFlags

Fields

benchmarkDistPref :: Flag FilePath

benchmarkVerbosity :: Flag Verbosity

benchmarkOptions :: [PathTemplate]

emptyBenchmarkFlags :: BenchmarkFlags

defaultBenchmarkFlags :: BenchmarkFlags

benchmarkCommand :: CommandUI BenchmarkFlags

data CopyDest

The location prefix for the copy command.

Constructors

NoCopyDest

CopyTo FilePath

CopyToDb FilePath
when using the ${pkgroot} as prefix. The CopyToDb will adjust the paths to be relative to the provided package database when copying / installing.

configureArgs :: Bool -> ConfigFlags -> [String]

Arguments to pass to a configure script, e.g. generated by autoconf.

configureOptions :: ShowOrParseArgs -> [OptionField ConfigFlags]

configureCCompiler :: Verbosity -> ProgramDb -> IO (FilePath, [String])

configureLinker :: Verbosity -> ProgramDb -> IO (FilePath, [String])

buildOptions :: ProgramDb -> ShowOrParseArgs -> [OptionField BuildFlags]

haddockOptions :: ShowOrParseArgs -> [OptionField HaddockFlags]

haddockProjectOptions :: ShowOrParseArgs -> [OptionField HaddockProjectFlags]

installDirsOptions :: [OptionField (InstallDirs (Flag PathTemplate))]

testOptions' :: ShowOrParseArgs -> [OptionField TestFlags]

benchmarkOptions' :: ShowOrParseArgs -> [OptionField BenchmarkFlags]

programDbOptions :: ProgramDb -> ShowOrParseArgs -> (flags -> [(String, [String])]) -> ([(String, [String])] -> flags -> flags) -> [OptionField flags]

For each known program PROG in progDb, produce a PROG-options OptionField.

programDbPaths' :: (String -> String) -> ProgramDb -> ShowOrParseArgs -> (flags -> [(String, FilePath)]) -> ([(String, FilePath)] -> flags -> flags) -> [OptionField flags]

Like programDbPaths, but allows to customise the option name.

programFlagsDescription :: ProgramDb -> String

replOptions :: ShowOrParseArgs -> [OptionField ReplOptions]

splitArgs :: String -> [String]

Helper function to split a string into a list of arguments. It's supposed to handle quoted things sensibly, eg:

splitArgs "--foo=\"C:/Program Files/Bar/" --baz"
  = ["--foo=C:/Program Files/Bar", "--baz"]

splitArgs "\"-DMSGSTR=\\\"foo bar\\\"\" --baz"
  = ["-DMSGSTR=\"foo bar\"","--baz"]

defaultDistPref :: FilePath

optionDistPref :: (flags -> Flag FilePath) -> (Flag FilePath -> flags -> flags) -> ShowOrParseArgs -> OptionField flags

data Flag a

All flags are monoids, they come in two flavours:

list flags eg

--ghc-option=foo --ghc-option=bar

gives us all the values ["foo", "bar"]

singular value flags, eg:

--enable-foo --disable-foo

gives us Just False So this Flag type is for the latter singular kind of flag. Its monoid instance gives us the behaviour where it starts out as NoFlag and later flags override earlier ones.

Constructors

Flag a

NoFlag

toFlag :: a -> Flag a

fromFlag :: WithCallStack (Flag a -> a)

fromFlagOrDefault :: a -> Flag a -> a

flagToMaybe :: Flag a -> Maybe a

flagToList :: Flag a -> [a]

maybeToFlag :: Maybe a -> Flag a

class BooleanFlag a where

Types that represent boolean flags.

Methods

asBool :: a -> Bool

boolOpt :: SFlags -> SFlags -> MkOptDescr (a -> Flag Bool) (Flag Bool -> a -> a) a

boolOpt' :: OptFlags -> OptFlags -> MkOptDescr (a -> Flag Bool) (Flag Bool -> a -> a) a

trueArg :: MkOptDescr (a -> Flag Bool) (Flag Bool -> a -> a) a

falseArg :: MkOptDescr (a -> Flag Bool) (Flag Bool -> a -> a) a

optionVerbosity :: (flags -> Flag Verbosity) -> (Flag Verbosity -> flags -> flags) -> OptionField flags

optionNumJobs :: (flags -> Flag (Maybe Int)) -> (Flag (Maybe Int) -> flags -> flags) -> OptionField flags