Distribution.Simple.Test.LibV09

Documentation

runTest :: PackageDescription -> LocalBuildInfo -> TestFlags -> TestSuite -> IO TestSuiteLog

simpleTestStub :: ModuleName -> String

Source code for library test suite stub executable

stubFilePath :: TestSuite -> FilePath

The filename of the source file for the stub executable associated with a library TestSuite.

stubMain :: IO [Test] -> IO ()

Main function for test stubs. Once, it was written directly into the stub, but minimizing the amount of code actually in the stub maximizes the number of detectable errors when Cabal is compiled.

stubName :: TestSuite -> FilePath

The name of the stub executable associated with a library TestSuite.

stubWriteLog :: FilePath -> String -> TestLogs -> IO ()

From a test stub, write the TestSuiteLog to temporary file for the calling Cabal process to read.

writeSimpleTestStub

Arguments

:: TestSuite
library TestSuite for which a stub is being created

-> FilePath
path to directory where stub source should be located

-> IO ()

Write the source file for a library TestSuite stub executable.