Distribution.System

Description

Cabal often needs to do slightly different things on specific platforms. You probably know about the os however using that is very inconvenient because it is a string and different Haskell implementations do not agree on using the same strings for the same platforms! (In particular see the controversy over "windows" vs "ming32"). So to make it more consistent and easy to use we have an OS enumeration.

Operating System

data OS

Constructors

Linux

Windows

OSX

FreeBSD

OpenBSD

NetBSD

DragonFly

Solaris

AIX

HPUX

IRIX

HaLVM

IOS

Ghcjs

OtherOS String

buildOS :: OS

Machine Architecture

data Arch

Constructors

I386

X86_64

PPC

PPC64

Sparc

Arm

Mips

SH

IA64

S390

Alpha

Hppa

Rs6000

M68k

Vax

JavaScript

OtherArch String

buildArch :: Arch

Platform is a pair of arch and OS

data Platform

Constructors

Platform Arch OS

buildPlatform :: Platform

The platform Cabal was compiled on. In most cases, LocalBuildInfo.hostPlatform should be used instead (the platform we're targeting).

platformFromTriple :: String -> Maybe Platform