Distribution.Text

Description

This defines a Text class which is a bit like the Read and Show classes. The difference is that is uses a modern pretty printer and parser system and the format is not expected to be Haskell concrete syntax but rather the external human readable representation used by Cabal.

Documentation

class Text a where

Minimal complete definition

disp, parse

Methods

disp :: a -> Doc

parse :: ReadP r a

defaultStyle :: Style

The default rendering style used in Cabal for console output.

display :: Text a => a -> String

simpleParse :: Text a => String -> Maybe a