Distribution.Types.BuildInfo

Documentation

data BuildInfo

Constructors

BuildInfo

Fields

buildable :: Bool

component is buildable here | Tools needed to build this bit.

This is a legacy field that "build-tool-depends" larely supersedes.

Unless use are very sure what you are doing, use the functions in BuildToolDepends rather than accessing this field directly.

buildTools :: [LegacyExeDependency]

toolDepends :: [ExeDependency]

Haskell tools needed to build this bit

This field is better than "build-tools" because it allows one to precisely specify an executable in a package.

Unless use are very sure what you are doing, use the functions in BuildToolDepends rather than accessing this field directly.

cppOptions :: [String]

options for pre-processing Haskell code

ccOptions :: [String]

options for C compiler

ldOptions :: [String]

options for linker

pkgconfigDepends :: [PkgconfigDependency]

pkg-config packages that are used

frameworks :: [String]

support frameworks for Mac OS X

extraFrameworkDirs :: [String]

extra locations to find frameworks.

cSources :: [FilePath]

jsSources :: [FilePath]

hsSourceDirs :: [FilePath]

where to look for the Haskell module hierarchy

otherModules :: [ModuleName]

non-exposed or non-main modules

autogenModules :: [ModuleName]

not present on sdist, Paths_* or user-generated with a custom Setup.hs

defaultLanguage :: Maybe Language

language used when not explicitly specified

otherLanguages :: [Language]

other languages used within the package

defaultExtensions :: [Extension]

language extensions used by all modules

otherExtensions :: [Extension]

other language extensions used within the package

oldExtensions :: [Extension]

the old extensions field, treated same as defaultExtensions

extraLibs :: [String]

what libraries to link with when compiling a program that uses your package

extraGHCiLibs :: [String]

if present, overrides extraLibs when package is loaded with GHCi.

extraLibDirs :: [String]

includeDirs :: [FilePath]

directories to find .h files

includes :: [FilePath]

The .h files to be found in includeDirs

installIncludes :: [FilePath]

.h files to install with the package

options :: [(CompilerFlavor, [String])]

profOptions :: [(CompilerFlavor, [String])]

sharedOptions :: [(CompilerFlavor, [String])]

customFieldsBI :: [(String, String)]

Custom fields starting with x-, stored in a simple assoc-list.

targetBuildDepends :: [Dependency]

Dependencies specific to a library or executable target

mixins :: [Mixin]

emptyBuildInfo :: BuildInfo

allLanguages :: BuildInfo -> [Language]

The Languages used by this component

allExtensions :: BuildInfo -> [Extension]

The Extensions that are used somewhere by this component

usedExtensions :: BuildInfo -> [Extension]

The Extensions that are used by all modules in this component

hcOptions :: CompilerFlavor -> BuildInfo -> [String]

Select options for a particular Haskell compiler.

hcProfOptions :: CompilerFlavor -> BuildInfo -> [String]

hcSharedOptions :: CompilerFlavor -> BuildInfo -> [String]