Distribution.Types.ForeignLib

Documentation

data ForeignLib

A foreign library stanza is like a library stanza, except that the built code is intended for consumption by a non-Haskell client.

Constructors

ForeignLib

Fields

foreignLibName :: UnqualComponentName

Name of the foreign library

foreignLibType :: ForeignLibType

What kind of foreign library is this (static or dynamic).

foreignLibOptions :: [ForeignLibOption]

What options apply to this foreign library (e.g., are we merging in all foreign dependencies.)

foreignLibBuildInfo :: BuildInfo

Build information for this foreign library.

foreignLibVersionInfo :: Maybe LibVersionInfo

Libtool-style version-info data to compute library version. Refer to the libtool documentation on the current:revision:age versioning scheme.

foreignLibVersionLinux :: Maybe Version

Linux library version

foreignLibModDefFile :: [FilePath]

(Windows-specific) module definition files

This is a list rather than a maybe field so that we can flatten the condition trees (for instance, when creating an sdist)

emptyForeignLib :: ForeignLib

An empty foreign library.

foreignLibModules :: ForeignLib -> [ModuleName]

Modules defined by a foreign library.

foreignLibIsShared :: ForeignLib -> Bool

Is the foreign library shared?

foreignLibVersion :: ForeignLib -> OS -> [Int]

Get a version number for a foreign library. If we're on Linux, and a Linux version is specified, use that. If we're on Linux, and libtool-style version-info is specified, translate that field into appropriate version numbers. Otherwise, this feature is unsupported so we don't return any version data.

data LibVersionInfo

mkLibVersionInfo :: (Int, Int, Int) -> LibVersionInfo

Construct LibVersionInfo from (current, revision, age) numbers.

For instance, mkLibVersionInfo (3,0,0) constructs a LibVersionInfo representing the version-info 3:0:0.

All version components must be non-negative.

libVersionInfoCRA :: LibVersionInfo -> (Int, Int, Int)

From a given LibVersionInfo, extract the (current, revision, age) numbers.

libVersionNumber :: LibVersionInfo -> (Int, Int, Int)

Given a version-info field, produce a major.minor.build version

libVersionNumberShow :: LibVersionInfo -> String

Given a version-info field, return major.minor.build as a String

libVersionMajor :: LibVersionInfo -> Int

Return the major version of a version-info field.