Distribution.Types.GenericPackageDescription

Documentation

data GenericPackageDescription

Constructors

GenericPackageDescription

Fields

packageDescription :: PackageDescription

genPackageFlags :: [Flag]

condLibrary :: Maybe (CondTree ConfVar [Dependency] Library)

condSubLibraries :: [(UnqualComponentName, CondTree ConfVar [Dependency] Library)]

condForeignLibs :: [(UnqualComponentName, CondTree ConfVar [Dependency] ForeignLib)]

condExecutables :: [(UnqualComponentName, CondTree ConfVar [Dependency] Executable)]

condTestSuites :: [(UnqualComponentName, CondTree ConfVar [Dependency] TestSuite)]

condBenchmarks :: [(UnqualComponentName, CondTree ConfVar [Dependency] Benchmark)]

data Flag

A flag can represent a feature to be included, or a way of linking a target against its dependencies, or in fact whatever you can think of.

Constructors

MkFlag

Fields

flagName :: FlagName

flagDescription :: String

flagDefault :: Bool

flagManual :: Bool

emptyFlag :: FlagName -> Flag

A Flag initialized with default parameters.

data FlagName

A FlagName is the name of a user-defined configuration flag

Use mkFlagName and unFlagName to convert from/to a String.

This type is opaque since Cabal-2.0

Since: 2.0.0.2

mkFlagName :: String -> FlagName

Construct a FlagName from a String

mkFlagName is the inverse to unFlagName

Note: No validations are performed to ensure that the resulting FlagName is valid

Since: 2.0.0.2

unFlagName :: FlagName -> String

Convert FlagName to String

Since: 2.0.0.2

type FlagAssignment = [(FlagName, Bool)]

A FlagAssignment is a total or partial mapping of FlagNames to Bool flag values. It represents the flags chosen by the user or discovered during configuration. For example --flags=foo --flags=-bar becomes [("foo", True), ("bar", False)]

showFlagValue :: (FlagName, Bool) -> String

String representation of a flag-value pair.

dispFlagAssignment :: FlagAssignment -> Doc

Pretty-prints a flag assignment.

parseFlagAssignment :: ReadP r FlagAssignment

Parses a flag assignment.

data ConfVar

A ConfVar represents the variable type used.

Constructors

OS OS

Arch Arch

Flag FlagName

Impl CompilerFlavor VersionRange