Distribution.Types.Library

Documentation

data Library

Constructors

Library

Fields

libName :: Maybe UnqualComponentName

exposedModules :: [ModuleName]

reexportedModules :: [ModuleReexport]

signatures :: [ModuleName]

What sigs need implementations?

libExposed :: Bool

Is the lib to be exposed by default?

libBuildInfo :: BuildInfo

emptyLibrary :: Library

explicitLibModules :: Library -> [ModuleName]

Get all the module names from the library (exposed and internal modules) which are explicitly listed in the package description which would need to be compiled. (This does not include reexports, which do not need to be compiled.) This may not include all modules for which GHC generated interface files (i.e., implicit modules.)

libModulesAutogen :: Library -> [ModuleName]

Get all the auto generated module names from the library, exposed or not. This are a subset of libModules.

libModules :: Library -> [ModuleName]

Deprecated: If you want all modules that are built with a library, use allLibModules. Otherwise, use explicitLibModules for ONLY the modules explicitly mentioned in the package description.

Backwards-compatibility shim for explicitLibModules. In most cases, you actually want allLibModules, which returns all modules that will actually be compiled, as opposed to those which are explicitly listed in the package description (explicitLibModules); unfortunately, the type signature for allLibModules is incompatible since we need a ComponentLocalBuildInfo.