Distribution.Types.SourceRepo.Lens

Documentation

data SourceRepo

Information about the source revision control system for a package.

When specifying a repo it is useful to know the meaning or intention of the information as doing so enables automation. There are two obvious common purposes: one is to find the repo for the latest development version, the other is to find the repo for this specific release. The ReopKind specifies which one we mean (or another custom one).

A package can specify one or the other kind or both. Most will specify just a head repo but some may want to specify a repo to reconstruct the sources for this package release.

The required information is the RepoType which tells us if it's using Darcs, Git for example. The repoLocation and other details are interpreted according to the repo type.

repoKind :: Lens' SourceRepo RepoKind

repoType :: Lens' SourceRepo (Maybe RepoType)

repoLocation :: Lens' SourceRepo (Maybe String)

repoModule :: Lens' SourceRepo (Maybe String)

repoBranch :: Lens' SourceRepo (Maybe String)

repoTag :: Lens' SourceRepo (Maybe String)

repoSubdir :: Lens' SourceRepo (Maybe FilePath)