Distribution.Types.TargetInfo

Documentation

data TargetInfo

The TargetInfo contains all the information necessary to build a specific target (e.g., componentmodulefile) in a package. In principle, one can get the Component from a ComponentLocalBuildInfo and LocalBuildInfo, but it is much more convenient to have the component in hand.

Constructors

TargetInfo

Fields

targetCLBI :: ComponentLocalBuildInfo

targetComponent :: Component