Distribution.Types.UnqualComponentName

Documentation

data UnqualComponentName

An unqualified component name, for any kind of component.

This is distinguished from a ComponentName and ComponentId. The former also states which of a library, executable, etc the name refers too. The later uniquely identifiers a component and its closure.

Since: 2.0

unUnqualComponentName :: UnqualComponentName -> String

Convert UnqualComponentName to String

Since: 2.0

mkUnqualComponentName :: String -> UnqualComponentName

Construct a UnqualComponentName from a String

mkUnqualComponentName is the inverse to unUnqualComponentName

Note: No validations are performed to ensure that the resulting UnqualComponentName is valid

Since: 2.0

packageNameToUnqualComponentName :: PackageName -> UnqualComponentName

Converts a package name to an unqualified component name

Useful in legacy situations where a package name may refer to an internal component, if one is defined with that name.

Since: 2.0

unqualComponentNameToPackageName :: UnqualComponentName -> PackageName

Converts an unqualified component name to a package name

packageNameToUnqualComponentName is the inverse of unqualComponentNameToPackageName.

Useful in legacy situations where a package name may refer to an internal component, if one is defined with that name.

Since: 2.0