Distribution.Utils.IOData

Description

Since: 2.2.0

IOData & IODataMode type

data IOData

Represents either textual or binary data passed via I/O functions which support binary/text mode

Since: 2.2.0

Constructors

IODataText String
How Text gets encoded is usually locale-dependent.

IODataBinary ByteString
Raw binary which gets read/written in binary mode.

data IODataMode

Constructors

IODataModeText

IODataModeBinary

null :: IOData -> Bool

Test whether IOData is empty

Since: 2.2.0

hGetContents :: Handle -> IODataMode -> IO IOData

IOData Wrapper for hGetContents

Note: This operation uses lazy I/O. Use NFData to force all data to be read and consequently the internal file handle to be closed.

Since: 2.2.0

hPutContents :: Handle -> IOData -> IO ()

IOData Wrapper for hPutStr and hClose

This is the dual operation ot ioDataHGetContents, and consequently the handle is closed with hClose.

Since: 2.2.0