Distribution.Utils.LogProgress

Documentation

type LogProgress a = Progress LogMsg Doc a

The Progress monad with specialized logging and error messages.

data LogMsg

A tracing message which will be output at some verbosity.

Constructors

LogMsg Verbosity Doc

runLogProgress :: Verbosity -> LogProgress a -> IO a

Run LogProgress, outputting traces according to Verbosity, die if there is an error.

warnProgress :: Doc -> LogProgress ()

Output a warning trace message in LogProgress.

infoProgress :: Doc -> LogProgress ()

Output an informational trace message in LogProgress.