Distribution.Utils.LogProgress

Documentation

data LogProgress a

The Progress monad with specialized logging and error messages.

runLogProgress :: Verbosity -> LogProgress a -> NoCallStackIO a

Run LogProgress, outputting traces according to Verbosity, die if there is an error.

warnProgress :: Doc -> LogProgress ()

Output a warning trace message in LogProgress.

infoProgress :: Doc -> LogProgress ()

Output an informational trace message in LogProgress.

dieProgress :: Doc -> LogProgress a

Fail the computation with an error message.

addProgressCtx :: CtxMsg -> LogProgress a -> LogProgress a

Add a message to the error/warning context.