Distribution.Utils.NubList

Documentation

data NubList a

NubList : A de-duplicated list that maintains the original order.

toNubList :: Ord a => [a] -> NubList a

Smart constructor for the NubList type.

fromNubList :: NubList a -> [a]

overNubList :: Ord a => ([a] -> [a]) -> NubList a -> NubList a

Lift a function over lists to a function over NubLists.

data NubListR a

NubListR : A right-biased version of NubList. That is toNubListR [-XNoFoo, -XFoo, -XNoFoo] will result in [-XFoo, -XNoFoo], unlike the normal NubList, which is left-biased. Built on top of ordNubRight and listUnionRight.

toNubListR :: Ord a => [a] -> NubListR a

Smart constructor for the NubListR type.

fromNubListR :: NubListR a -> [a]

overNubListR :: Ord a => ([a] -> [a]) -> NubListR a -> NubListR a

Lift a function over lists to a function over NubListRs.