Distribution.Utils.ShortText

ShortText type

data ShortText

Compact representation of short Strings

The data is stored internally as UTF8 in an ShortByteString when compiled against bytestring >= 0.10.4, and otherwise the fallback is to use plain old non-compat '[Char]'.

Note: This type is for internal uses (such as e.g. PackageName) and shall not be exposed in Cabal's API

Since: 2.0.0

toShortText :: String -> ShortText

Construct ShortText from String

fromShortText :: ShortText -> String

Convert ShortText to String

internal utilities

decodeStringUtf8 :: [Word8] -> String

Decode String from UTF8-encoded octets.

Invalid data will be decoded as the replacement character (U+FFFD)

See also encodeStringUtf8

encodeStringUtf8 :: String -> [Word8]

Encode String to a list of UTF8-encoded octets

See also decodeUtf8