Distribution.Verbosity

Description

A Verbosity type with associated utilities.

There are 4 standard verbosity levels from silent, normal, verbose up to deafening. This is used for deciding what logging messages to print.

Verbosity also is equipped with some internal settings which can be used to control at a fine granularity the verbosity of specific settings (e.g., so that you can trace only particular things you are interested in.) It's important to note that the instances for Verbosity assume that this does not exist.

Verbosity

data Verbosity

silent :: Verbosity

normal :: Verbosity

verbose :: Verbosity

deafening :: Verbosity

moreVerbose :: Verbosity -> Verbosity

lessVerbose :: Verbosity -> Verbosity

intToVerbosity :: Int -> Maybe Verbosity

flagToVerbosity :: ReadE Verbosity

showForCabal :: Verbosity -> String

showForGHC :: Verbosity -> String

Call stacks

verboseCallSite :: Verbosity -> Verbosity

Turn on verbose call-site printing when we log.

verboseCallStack :: Verbosity -> Verbosity

Turn on verbose call-stack printing when we log.

isVerboseCallSite :: Verbosity -> Bool

Test if we should output call sites when we log.

isVerboseCallStack :: Verbosity -> Bool

Test if we should output call stacks when we log.

line-wrapping

verboseNoWrap :: Verbosity -> Verbosity

Disable line-wrapping for log messages.

isVerboseNoWrap :: Verbosity -> Bool

Test if line-wrapping is disabled for log messages.