Distribution.Verbosity

Description

A Verbosity type with associated utilities.

There are 4 standard verbosity levels from silent, normal, verbose up to deafening. This is used for deciding what logging messages to print.

Verbosity also is equipped with some internal settings which can be used to control at a fine granularity the verbosity of specific settings (e.g., so that you can trace only particular things you are interested in.) It's important to note that the instances for Verbosity assume that this does not exist.

Verbosity

data Verbosity

silent :: Verbosity

In silent mode, we should not print anything unless an error occurs.

normal :: Verbosity

Print stuff we want to see by default.

verbose :: Verbosity

Be more verbose about what's going on.

deafening :: Verbosity

Not only are we verbose ourselves (perhaps even noisier than when being verbose), but we tell everything we run to be verbose too.

moreVerbose :: Verbosity -> Verbosity

Increase verbosity level, but stay silent if we are.

lessVerbose :: Verbosity -> Verbosity

Decrease verbosity level, but stay deafening if we are.

isVerboseQuiet :: Verbosity -> Bool

Test if we had called lessVerbose on the verbosity.

intToVerbosity :: Int -> Maybe Verbosity

Numeric verbosity level 0..3: 0 is silent, 3 is deafening.

flagToVerbosity :: ReadE Verbosity

showForCabal :: Verbosity -> String

showForGHC :: Verbosity -> String

verboseNoFlags :: Verbosity -> Verbosity

Turn off all flags.

verboseHasFlags :: Verbosity -> Bool

modifyVerbosity :: (Verbosity -> Verbosity) -> Verbosity -> Verbosity

Combinator for transforming verbosity level while retaining the original hidden state.

For instance, the following property holds

isVerboseNoWrap (modifyVerbosity (max verbose) v) == isVerboseNoWrap v

Note: you can use modifyVerbosity (const v1) v0 to overwrite v1's flags with v0's flags.

Since: Cabal-2.0.1.0

Call stacks

verboseCallSite :: Verbosity -> Verbosity

Turn on verbose call-site printing when we log.

verboseCallStack :: Verbosity -> Verbosity

Turn on verbose call-stack printing when we log.

isVerboseCallSite :: Verbosity -> Bool

Test if we should output call sites when we log.

isVerboseCallStack :: Verbosity -> Bool

Test if we should output call stacks when we log.

Output markets

verboseMarkOutput :: Verbosity -> Verbosity

Turn on -----BEGIN CABAL OUTPUT----- markers for output from Cabal (as opposed to GHC, or system dependent).

isVerboseMarkOutput :: Verbosity -> Bool

Test if we should output markets.

verboseUnmarkOutput :: Verbosity -> Verbosity

Turn off marking; useful for suppressing nondeterministic output.

Line wrapping

verboseNoWrap :: Verbosity -> Verbosity

Disable line-wrapping for log messages.

isVerboseNoWrap :: Verbosity -> Bool

Test if line-wrapping is disabled for log messages.

Time stamps

verboseTimestamp :: Verbosity -> Verbosity

Turn on timestamps for log messages.

isVerboseTimestamp :: Verbosity -> Bool

Test if we should output timestamps when we log.

verboseNoTimestamp :: Verbosity -> Verbosity

Turn off timestamps for log messages.

Stderr

verboseStderr :: Verbosity -> Verbosity

Switch logging to stderr.

Since: Cabal-3.4.0.0

isVerboseStderr :: Verbosity -> Bool

Test if we should output to stderr when we log.

Since: Cabal-3.4.0.0

verboseNoStderr :: Verbosity -> Verbosity

Switch logging to stdout.

Since: Cabal-3.4.0.0

No warnings

verboseNoWarn :: Verbosity -> Verbosity

Turn off warnings for log messages.

isVerboseNoWarn :: Verbosity -> Bool

Test if we should output warnings when we log.