Distribution.Verbosity

Description

A simple Verbosity type with associated utilities. There are 4 standard verbosity levels from silent, normal, verbose up to deafening. This is used for deciding what logging messages to print.

Verbosity

data Verbosity

silent :: Verbosity

normal :: Verbosity

verbose :: Verbosity

deafening :: Verbosity

moreVerbose :: Verbosity -> Verbosity

lessVerbose :: Verbosity -> Verbosity

intToVerbosity :: Int -> Maybe Verbosity

flagToVerbosity :: ReadE Verbosity

showForCabal :: Verbosity -> String

showForGHC :: Verbosity -> String