Language.Haskell.Extension

Description

Haskell language dialects and extensions

Documentation

data Language

This represents a Haskell language dialect.

Language Extensions are interpreted relative to one of these base languages.

Constructors

Haskell98
The Haskell 98 language as defined by the Haskell 98 report. http://haskell.org/onlinereport/

Haskell2010
The Haskell 2010 language as defined by the Haskell 2010 report. http://www.haskell.org/onlinereport/haskell2010

UnknownLanguage String
An unknown language, identified by its name.

knownLanguages :: [Language]

data Extension

This represents language extensions beyond a base Language definition (such as Haskell98) that are supported by some implementations, usually in some special mode.

Where applicable, references are given to an implementation's official documentation.

Constructors

EnableExtension KnownExtension
Enable a known extension

DisableExtension KnownExtension
Disable a known extension

UnknownExtension String
An unknown extension, identified by the name of its LANGUAGE pragma.

data KnownExtension

Constructors

OverlappingInstances
Allow overlapping class instances, provided there is a unique most specific instance for each use.

http://www.haskell.org/ghc/docs/latest/html/users_guide/type-class-extensions.html#instance-overlap

UndecidableInstances
Ignore structural rules guaranteeing the termination of class instance resolution. Termination is guaranteed by a fixed-depth recursion stack, and compilation may fail if this depth is exceeded.

http://www.haskell.org/ghc/docs/latest/html/users_guide/type-class-extensions.html#undecidable-instances

IncoherentInstances
Implies OverlappingInstances. Allow the implementation to choose an instance even when it is possible that further instantiation of types will lead to a more specific instance being applicable.

http://www.haskell.org/ghc/docs/latest/html/users_guide/type-class-extensions.html#instance-overlap

DoRec
(deprecated) Allow recursive bindings in do blocks, using the rec keyword. See also RecursiveDo.

RecursiveDo
Allow recursive bindings using mdo, a variant of do. DoRec provides a different, preferred syntax.

http://www.haskell.org/ghc/docs/latest/html/users_guide/syntax-extns.html#recursive-do-notation

ParallelListComp
Provide syntax for writing list comprehensions which iterate over several lists together, like the zipWith family of functions.

http://www.haskell.org/ghc/docs/latest/html/users_guide/syntax-extns.html#parallel-list-comprehensions

MultiParamTypeClasses
Allow multiple parameters in a type class.

http://www.haskell.org/ghc/docs/latest/html/users_guide/type-class-extensions.html#multi-param-type-classes

MonomorphismRestriction
Enable the dreaded monomorphism restriction.

http://www.haskell.org/ghc/docs/latest/html/users_guide/monomorphism.html

FunctionalDependencies
Allow a specification attached to a multi-parameter type class which indicates that some parameters are entirely determined by others. The implementation will check that this property holds for the declared instances, and will use this property to reduce ambiguity in instance resolution.

http://www.haskell.org/ghc/docs/latest/html/users_guide/type-class-extensions.html#functional-dependencies

Rank2Types
Like RankNTypes but does not allow a higher-rank type to itself appear on the left of a function arrow.

http://www.haskell.org/ghc/docs/latest/html/users_guide/other-type-extensions.html#universal-quantification

RankNTypes
Allow a universally-quantified type to occur on the left of a function arrow.

http://www.haskell.org/ghc/docs/latest/html/users_guide/other-type-extensions.html#universal-quantification

PolymorphicComponents
Allow data constructors to have polymorphic arguments. Unlike RankNTypes, does not allow this for ordinary functions.

http://www.haskell.org/ghc/docs/latest/html/users_guide/other-type-extensions.html#universal-quantification

ExistentialQuantification
Allow existentially-quantified data constructors.

http://www.haskell.org/ghc/docs/latest/html/users_guide/data-type-extensions.html#existential-quantification

ScopedTypeVariables
Cause a type variable in a signature, which has an explicit forall quantifier, to scope over the definition of the accompanying value declaration.

http://www.haskell.org/ghc/docs/latest/html/users_guide/other-type-extensions.html#scoped-type-variables

PatternSignatures
Deprecated, use ScopedTypeVariables instead.

ImplicitParams
Enable implicit function parameters with dynamic scope.

http://www.haskell.org/ghc/docs/latest/html/users_guide/other-type-extensions.html#implicit-parameters

FlexibleContexts
Relax some restrictions on the form of the context of a type signature.

http://www.haskell.org/ghc/docs/latest/html/users_guide/other-type-extensions.html#flexible-contexts

FlexibleInstances
Relax some restrictions on the form of the context of an instance declaration.

http://www.haskell.org/ghc/docs/latest/html/users_guide/type-class-extensions.html#instance-rules

EmptyDataDecls
Allow data type declarations with no constructors.

http://www.haskell.org/ghc/docs/latest/html/users_guide/data-type-extensions.html#nullary-types

CPP
Run the C preprocessor on Haskell source code.

http://www.haskell.org/ghc/docs/latest/html/users_guide/options-phases.html#c-pre-processor

KindSignatures
Allow an explicit kind signature giving the kind of types over which a type variable ranges.

http://www.haskell.org/ghc/docs/latest/html/users_guide/other-type-extensions.html#kinding

BangPatterns
Enable a form of pattern which forces evaluation before an attempted match, and a form of strict let/where binding.

http://www.haskell.org/ghc/docs/latest/html/users_guide/bang-patterns.html

TypeSynonymInstances
Allow type synonyms in instance heads.

http://www.haskell.org/ghc/docs/latest/html/users_guide/type-class-extensions.html#flexible-instance-head

TemplateHaskell
Enable Template Haskell, a system for compile-time metaprogramming.

http://www.haskell.org/ghc/docs/latest/html/users_guide/template-haskell.html

ForeignFunctionInterface
Enable the Foreign Function Interface. In GHC, implements the standard Haskell 98 Foreign Function Interface Addendum, plus some GHC-specific extensions.

http://www.haskell.org/ghc/docs/latest/html/users_guide/ffi.html

Arrows
Enable arrow notation.

http://www.haskell.org/ghc/docs/latest/html/users_guide/arrow-notation.html

Generics
(deprecated) Enable generic type classes, with default instances defined in terms of the algebraic structure of a type.

http://www.haskell.org/ghc/docs/latest/html/users_guide/generic-classes.html

ImplicitPrelude
Enable the implicit importing of the module Prelude. When disabled, when desugaring certain built-in syntax into ordinary identifiers, use whatever is in scope rather than the Prelude -- version.

http://www.haskell.org/ghc/docs/latest/html/users_guide/syntax-extns.html#rebindable-syntax

NamedFieldPuns
Enable syntax for implicitly binding local names corresponding to the field names of a record. Puns bind specific names, unlike RecordWildCards.

http://www.haskell.org/ghc/docs/latest/html/users_guide/syntax-extns.html#record-puns

PatternGuards
Enable a form of guard which matches a pattern and binds variables.

http://www.haskell.org/ghc/docs/latest/html/users_guide/syntax-extns.html#pattern-guards

GeneralizedNewtypeDeriving
Allow a type declared with newtype to use deriving for any class with an instance for the underlying type.

http://www.haskell.org/ghc/docs/latest/html/users_guide/deriving.html#newtype-deriving

ExtensibleRecords
Enable the "Trex" extensible records system.

http://cvs.haskell.org/Hugs/pages/users_guide/hugs-only.html#TREX

RestrictedTypeSynonyms
Enable type synonyms which are transparent in some definitions and opaque elsewhere, as a way of implementing abstract datatypes.

http://cvs.haskell.org/Hugs/pages/users_guide/restricted-synonyms.html

HereDocuments
Enable an alternate syntax for string literals, with string templating.

http://cvs.haskell.org/Hugs/pages/users_guide/here-documents.html

MagicHash
Allow the character # as a postfix modifier on identifiers. Also enables literal syntax for unboxed values.

http://www.haskell.org/ghc/docs/latest/html/users_guide/syntax-extns.html#magic-hash

TypeFamilies
Allow data types and type synonyms which are indexed by types, i.e. ad-hoc polymorphism for types.

http://www.haskell.org/ghc/docs/latest/html/users_guide/type-families.html

StandaloneDeriving
Allow a standalone declaration which invokes the type class deriving mechanism.

http://www.haskell.org/ghc/docs/latest/html/users_guide/deriving.html#stand-alone-deriving

UnicodeSyntax
Allow certain Unicode characters to stand for certain ASCII character sequences, e.g. keywords and punctuation.

http://www.haskell.org/ghc/docs/latest/html/users_guide/syntax-extns.html#unicode-syntax

UnliftedFFITypes
Allow the use of unboxed types as foreign types, e.g. in foreign import and foreign export.

http://www.haskell.org/ghc/docs/latest/html/users_guide/ffi.html#id681687

InterruptibleFFI
Enable interruptible FFI.

http://www.haskell.org/ghc/docs/latest/html/users_guide/ffi.html#ffi-interruptible

CApiFFI
Allow use of CAPI FFI calling convention (foreign import capi).

http://www.haskell.org/ghc/docs/latest/html/users_guide/ffi.html#ffi-capi

LiberalTypeSynonyms
Defer validity checking of types until after expanding type synonyms, relaxing the constraints on how synonyms may be used.

http://www.haskell.org/ghc/docs/latest/html/users_guide/data-type-extensions.html#type-synonyms

TypeOperators
Allow the name of a type constructor, type class, or type variable to be an infix operator.

RecordWildCards
Enable syntax for implicitly binding local names corresponding to the field names of a record. A wildcard binds all unmentioned names, unlike NamedFieldPuns.

http://www.haskell.org/ghc/docs/latest/html/users_guide/syntax-extns.html#record-wildcards

RecordPuns
Deprecated, use NamedFieldPuns instead.

DisambiguateRecordFields
Allow a record field name to be disambiguated by the type of the record it's in.

http://www.haskell.org/ghc/docs/latest/html/users_guide/syntax-extns.html#disambiguate-fields

TraditionalRecordSyntax
Enable traditional record syntax (as supported by Haskell 98)

http://www.haskell.org/ghc/docs/latest/html/users_guide/syntax-extns.html#traditional-record-syntax

OverloadedStrings
Enable overloading of string literals using a type class, much like integer literals.

http://www.haskell.org/ghc/docs/latest/html/users_guide/type-class-extensions.html#overloaded-strings

GADTs
Enable generalized algebraic data types, in which type variables may be instantiated on a per-constructor basis. Implies GADTSyntax.

http://www.haskell.org/ghc/docs/latest/html/users_guide/data-type-extensions.html#gadt

GADTSyntax
Enable GADT syntax for declaring ordinary algebraic datatypes.

http://www.haskell.org/ghc/docs/latest/html/users_guide/data-type-extensions.html#gadt-style

MonoPatBinds
Make pattern bindings monomorphic.

http://www.haskell.org/ghc/docs/latest/html/users_guide/monomorphism.html#id630981

RelaxedPolyRec
Relax the requirements on mutually-recursive polymorphic functions.

http://www.haskell.org/ghc/docs/latest/html/users_guide/other-type-extensions.html#typing-binds

ExtendedDefaultRules
Allow default instantiation of polymorphic types in more situations.

http://www.haskell.org/ghc/docs/latest/html/users_guide/interactive-evaluation.html#extended-default-rules

UnboxedTuples
Enable unboxed tuples.

http://www.haskell.org/ghc/docs/latest/html/users_guide/primitives.html#unboxed-tuples

DeriveDataTypeable
Enable deriving for classes Typeable and Data.

http://www.haskell.org/ghc/docs/latest/html/users_guide/deriving.html#deriving-typeable

DeriveGeneric
Enable deriving for Generic and Generic1.

http://www.haskell.org/ghc/docs/latest/html/users_guide/deriving.html#deriving-typeable

DefaultSignatures
Enable support for default signatures.

http://www.haskell.org/ghc/docs/latest/html/users_guide/type-class-extensions.html#class-default-signatures

InstanceSigs
Allow type signatures to be specified in instance declarations.

http://www.haskell.org/ghc/docs/latest/html/users_guide/type-class-extensions.html#instance-sigs

ConstrainedClassMethods
Allow a class method's type to place additional constraints on a class type variable.

http://www.haskell.org/ghc/docs/latest/html/users_guide/type-class-extensions.html#class-method-types

PackageImports
Allow imports to be qualified by the package name the module is intended to be imported from, e.g.

import "network" Network.Socket

http://www.haskell.org/ghc/docs/latest/html/users_guide/syntax-extns.html#package-imports

ImpredicativeTypes
(deprecated) Allow a type variable to be instantiated at a polymorphic type.

http://www.haskell.org/ghc/docs/6.12.3/html/users_guide/other-type-extensions.html#impredicative-polymorphism

NewQualifiedOperators
(deprecated) Change the syntax for qualified infix operators.

http://www.haskell.org/ghc/docs/6.12.3/html/users_guide/syntax-extns.html#new-qualified-operators

PostfixOperators
Relax the interpretation of left operator sections to allow unary postfix operators.

http://www.haskell.org/ghc/docs/latest/html/users_guide/syntax-extns.html#postfix-operators

QuasiQuotes
Enable quasi-quotation, a mechanism for defining new concrete syntax for expressions and patterns.

http://www.haskell.org/ghc/docs/latest/html/users_guide/template-haskell.html#th-quasiquotation

TransformListComp
Enable generalized list comprehensions, supporting operations such as sorting and grouping.

http://www.haskell.org/ghc/docs/latest/html/users_guide/syntax-extns.html#generalised-list-comprehensions

MonadComprehensions
Enable monad comprehensions, which generalise the list comprehension syntax to work for any monad.

http://www.haskell.org/ghc/docs/latest/html/users_guide/syntax-extns.html#monad-comprehensions

ViewPatterns
Enable view patterns, which match a value by applying a function and matching on the result.

http://www.haskell.org/ghc/docs/latest/html/users_guide/syntax-extns.html#view-patterns

XmlSyntax
Allow concrete XML syntax to be used in expressions and patterns, as per the Haskell Server Pages extension language: http://www.haskell.org/haskellwiki/HSP. The ideas behind it are discussed in the paper "Haskell Server Pages through Dynamic Loading" by Niklas Broberg, from Haskell Workshop '05.

RegularPatterns
Allow regular pattern matching over lists, as discussed in the paper "Regular Expression Patterns" by Niklas Broberg, Andreas Farre and Josef Svenningsson, from ICFP '04.

TupleSections
Enable the use of tuple sections, e.g. (, True) desugars into x -> (x, True).

http://www.haskell.org/ghc/docs/latest/html/users_guide/syntax-extns.html#tuple-sections

GHCForeignImportPrim
Allow GHC primops, written in C--, to be imported into a Haskell file.

NPlusKPatterns
Support for patterns of the form n + k, where k is an integer literal.

http://www.haskell.org/ghc/docs/latest/html/users_guide/syntax-extns.html#n-k-patterns

DoAndIfThenElse
Improve the layout rule when if expressions are used in a do block.

MultiWayIf
Enable support for multi-way if-expressions.

http://www.haskell.org/ghc/docs/latest/html/users_guide/syntax-extns.html#multi-way-if

LambdaCase
Enable support lambda-case expressions.

http://www.haskell.org/ghc/docs/latest/html/users_guide/syntax-extns.html#lambda-case

RebindableSyntax
Makes much of the Haskell sugar be desugared into calls to the function with a particular name that is in scope.

http://www.haskell.org/ghc/docs/latest/html/users_guide/syntax-extns.html#rebindable-syntax

ExplicitForAll
Make forall a keyword in types, which can be used to give the generalisation explicitly.

http://www.haskell.org/ghc/docs/latest/html/users_guide/other-type-extensions.html#explicit-foralls

DatatypeContexts
Allow contexts to be put on datatypes, e.g. the Eq a in data Eq a => Set a = NilSet | ConsSet a (Set a).

http://www.haskell.org/ghc/docs/latest/html/users_guide/data-type-extensions.html#datatype-contexts

MonoLocalBinds
Local (let and where) bindings are monomorphic.

http://www.haskell.org/ghc/docs/latest/html/users_guide/other-type-extensions.html#mono-local-binds

DeriveFunctor
Enable deriving for the Functor class.

http://www.haskell.org/ghc/docs/latest/html/users_guide/deriving.html#deriving-typeable

DeriveTraversable
Enable deriving for the Traversable class.

http://www.haskell.org/ghc/docs/latest/html/users_guide/deriving.html#deriving-typeable

DeriveFoldable
Enable deriving for the Foldable class.

http://www.haskell.org/ghc/docs/latest/html/users_guide/deriving.html#deriving-typeable

NondecreasingIndentation
Enable non-decreasing indentation for do blocks.

http://www.haskell.org/ghc/docs/latest/html/users_guide/bugs-and-infelicities.html#infelicities-syntax

SafeImports
Allow imports to be qualified with a safe keyword that requires the imported module be trusted as according to the Safe Haskell definition of trust.

import safe Network.Socket

http://www.haskell.org/ghc/docs/latest/html/users_guide/safe-haskell.html#safe-imports

Safe
Compile a module in the Safe, Safe Haskell mode -- a restricted form of the Haskell language to ensure type safety.

http://www.haskell.org/ghc/docs/latest/html/users_guide/safe-haskell.html#safe-trust

Trustworthy
Compile a module in the Trustworthy, Safe Haskell mode -- no restrictions apply but the module is marked as trusted as long as the package the module resides in is trusted.

http://www.haskell.org/ghc/docs/latest/html/users_guide/safe-haskell.html#safe-trust

Unsafe
Compile a module in the Unsafe, Safe Haskell mode so that modules compiled using Safe, Safe Haskell mode can't import it.

http://www.haskell.org/ghc/docs/latest/html/users_guide/safe-haskell.html#safe-trust

ConstraintKinds
Allow type classimplicit parameterequality constraints to be used as types with the special kind constraint. Also generalise the (ctxt => ty) syntax so that any type of kind constraint can occur before the arrow.

http://www.haskell.org/ghc/docs/latest/html/users_guide/constraint-kind.html

PolyKinds
Enable kind polymorphism.

http://www.haskell.org/ghc/docs/latest/html/users_guide/kind-polymorphism.html

DataKinds
Enable datatype promotion.

http://www.haskell.org/ghc/docs/latest/html/users_guide/promotion.html

ParallelArrays
Enable parallel arrays syntax ([:, :]) for Data Parallel Haskell.

http://www.haskell.org/haskellwiki/GHC/Data_Parallel_Haskell

RoleAnnotations
Enable explicit role annotations, like in (type role Foo representational representational).

http://www.haskell.org/ghc/docs/latest/html/users_guide/roles.html

OverloadedLists
Enable overloading of list literals, arithmetic sequences and list patterns using the IsList type class.

http://www.haskell.org/ghc/docs/latest/html/users_guide/type-class-extensions.html#overloaded-lists

EmptyCase
Enable case expressions that have no alternatives. Also applies to lambda-case expressions if they are enabled.

http://www.haskell.org/ghc/docs/latest/html/users_guide/syntax-extns.html#empty-case

AutoDeriveTypeable
Triggers the generation of derived Typeable instances for every datatype and type class declaration.

http://www.haskell.org/ghc/docs/latest/html/users_guide/deriving.html#auto-derive-typeable

NegativeLiterals
Desugars negative literals directly (without using negate).

http://www.haskell.org/ghc/docs/latest/html/users_guide/syntax-extns.html#negative-literals

BinaryLiterals
Allow the use of binary integer literal syntax (e.g. 0b11001001 to denote 201).

http://www.haskell.org/ghc/docs/latest/html/users_guide/syntax-extns.html#binary-literals

NumDecimals
Allow the use of floating literal syntax for all instances of Num, including Int and Integer.

http://www.haskell.org/ghc/docs/latest/html/users_guide/syntax-extns.html#num-decimals

NullaryTypeClasses
Enable support for type classes with no type parameter.

http://www.haskell.org/ghc/docs/latest/html/users_guide/type-class-extensions.html#nullary-type-classes

ExplicitNamespaces
Enable explicit namespaces in module import/export lists.

http://www.haskell.org/ghc/docs/latest/html/users_guide/syntax-extns.html#explicit-namespaces

AllowAmbiguousTypes
Allow the user to write ambiguous types, and the type inference engine to infer them.

http://www.haskell.org/ghc/docs/latest/html/users_guide/other-type-extensions.html#ambiguity

JavaScriptFFI
Enable foreign import javascript.

PatternSynonyms
Allow giving names to and abstracting over patterns.

http://www.haskell.org/ghc/docs/latest/html/users_guide/other-type-extensions.html#pattern-synonyms

PartialTypeSignatures
Allow anonymous placeholders (underscore) inside type signatures. The type inference engine will generate a message describing the type inferred at the hole's location.

http://www.haskell.org/ghc/docs/latest/html/users_guide/other-type-extensions.html#partial-type-signatures

NamedWildCards
Allow named placeholders written with a leading underscore inside type signatures. Wildcards with the same name unify to the same type.

http://www.haskell.org/ghc/docs/latest/html/users_guide/other-type-extensions.html#named-wildcards

DeriveAnyClass
Enable deriving for any class.

http://www.haskell.org/ghc/docs/latest/html/users_guide/other-type-extensions.html#derive-any-class

DeriveLift
Enable deriving for the Lift class.

http://www.haskell.org/ghc/docs/latest/html/users_guide/deriving.html#deriving-lift

StaticPointers
Enable support for 'static pointers' (and the static keyword) to refer to globally stable names, even across different programs.

http://www.haskell.org/ghc/docs/latest/html/users_guide/static-pointers.html

StrictData
Switches data type declarations to be strict by default (as if they had a bang using BangPatterns), and allow opt-in field laziness using ~.

Strict
Switches all pattern bindings to be strict by default (as if they had a bang using BangPatterns), ordinary patterns are recovered using ~. Implies StrictData.

ApplicativeDo
Allows do-notation for types that are Applicative as well as Monad. When enabled, desugaring do notation tries to use (*) and fmap and join as far as possible.

DuplicateRecordFields
Allow records to use duplicated field labels for accessors.

TypeApplications
Enable explicit type applications with the syntax id @Int.

TypeInType
Dissolve the distinction between types and kinds, allowing the compiler to reason about kind equality and therefore enabling GADTs to be promoted to the type-level.

UndecidableSuperClasses
Allow recursive (and therefore undecideable) super-class relationships.

MonadFailDesugaring
A temporary extension to help library authors check if their code will compile with the new planned desugaring of fail.

TemplateHaskellQuotes
A subset of TemplateHaskell including only quasi-quoting.

OverloadedLabels
Allows use of the #label syntax.

knownExtensions :: [KnownExtension]

Deprecated: KnownExtension is an instance of Enum and Bounded, use those instead.

deprecatedExtensions :: [(Extension, Maybe Extension)]

Extensions that have been deprecated, possibly paired with another extension that replaces it.